% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_cnps.R
\name{bafLikelihood}
\alias{bafLikelihood}
\title{Calculate the likelihood of the observed B allele frequencies for a given copy number model}
\usage{
bafLikelihood(cn.model, snpdata)
}
\arguments{
\item{cn.model}{a copy number model}

\item{snpdata}{a \code{SummarizedExperiment} with assay element `GT` containing an integer coding (1, 2, and 3) for the generic genotypes AA, AB, and BB, respectively.}
}
\value{
a list. The first element is the log likeihood and the second element is the copy number model that maximized the likelihood.
}
\description{
Calculate the likelihood of the observed B allele frequencies for a given copy number model
}
\examples{

}
