% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-DensityModel.R, R/methods-MixtureModel.R, R/methods-ModelParams.R
\docType{methods}
\name{batch}
\alias{batch}
\alias{batch,DensityModel-method}
\alias{batch,MixtureModel-method}
\alias{batch,ModelParams-method}
\title{Retrieve batches from object.}
\usage{
batch(object)

\S4method{batch}{DensityModel}(object)

\S4method{batch}{MixtureModel}(object)

\S4method{batch}{ModelParams}(object)
}
\arguments{
\item{object}{see \code{showMethods(batch)}}
}
\description{
The batches are represented as a vector of integers.
}
\examples{
\dontrun{
     data(BatchModelExample)
     batch(BatchModelExample)
}
}

