% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batch_likelihood.R
\name{batchLikelihood}
\alias{batchLikelihood}
\title{Compute the hierarchical likelihood of a converged model.}
\usage{
batchLikelihood(model, niter = 1000L)
}
\arguments{
\item{model}{An object of class \code{BatchModel}, or a list of
\code{BatchModel}'s.}

\item{niter}{The number of iterations for the reduced Gibb's sampler.}
}
\description{
Compute the hierarchical likelihood of a converged model.
}

