% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-BatchModel.R
\name{BatchModel}
\alias{BatchModel}
\title{Create an object for running hierarchical MCMC simulations.}
\usage{
BatchModel(data = numeric(), k = 2L, batch, hypp, mcmc.params)
}
\arguments{
\item{data}{the data for the simulation.}

\item{k}{An integer value specifying the number of latent classes.}

\item{batch}{a vector of the different batch numbers (must be sorted)}

\item{hypp}{An object of class `Hyperparameters` used to specify the hyperparameters of the model.}

\item{mcmc.params}{An object of class 'McmcParams'}
}
\value{
An object of class `BatchModel`
}
\description{
Create an object for running hierarchical MCMC simulations.
}
\examples{
model <- BatchModel(rnorm(10), k=1, batch=rep(1:2, each=5))
}

