% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BatchModelExample}
\alias{BatchModelExample}
\title{This data is a simulated example of Batch data}
\usage{
BatchModelExample
}
\description{
This data is a simulated example of Batch data
}
\author{
Jacob Carey
BatchModelExample
}

