% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BatchModel.R
\name{BatchModelList}
\alias{BatchModelList}
\title{Constructor for list of batch models}
\usage{
BatchModelList(data = numeric(), k = numeric(), batch,
  mcmc.params = McmcParams(), ...)
}
\arguments{
\item{data}{numeric vector of average log R ratios}

\item{k}{numeric vector indicating the number of mixture components for each model}

\item{...}{additional arguments to \code{HyperparametersBatch}}
}
\value{
a list. Each element of the list is a \code{BatchModel}
}
\description{
An object of class BatchModel is constructed for each k, creating a list of
BatchModels.
}
\examples{
mlist <- BatchModelList(data=y(BatchModelExample), k=1:4, batch=batch(BatchModelExample))
}

