% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{bayesFactor}
\alias{bayesFactor}
\title{Compute the Bayes factor}
\usage{
bayesFactor(model.list, prior.odds = 1)
}
\arguments{
\item{model.list}{list of models from \code{gibbs}}

\item{prior.odds}{scalar}
}
\description{
Calculated as log(ML1) - log(ML2) + log prior odds
where ML1 is the marginal likelihood of the model with the most free parameters
}
