% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/deprecated_functions.R,
%   R/methods-MultiBatchModel.R, R/methods-SingleBatchModel.R
\docType{methods}
\name{bic}
\alias{bic}
\alias{bic,MarginalModel-method}
\alias{bic,BatchModel-method}
\alias{bic,MultiBatchModel-method}
\alias{bic,SingleBatchModel-method}
\title{Calculate BIC of a model}
\usage{
bic(object)

\S4method{bic}{MarginalModel}(object)

\S4method{bic}{BatchModel}(object)

\S4method{bic}{MultiBatchModel}(object)

\S4method{bic}{SingleBatchModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(bic)}}
}
\value{
The BIC of the model.
}
\description{
Calculate BIC of a model
}
\examples{
     bic(BatchModelExample)
}
