% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{burnin<-}
\alias{burnin<-}
\title{Reset number of burnin iterations.}
\usage{
burnin(object) <- value
}
\arguments{
\item{object}{see \code{showMethods("burnin<-")}}

\item{value}{new number of burnin iterations}
}
\description{
This function changes the number of burnin simulations to be discarded.
}

