% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{burnin}
\alias{burnin}
\title{Number of burnin iterations.}
\usage{
burnin(object)
}
\arguments{
\item{object}{see \code{showMethods(burnin)}}
}
\description{
This function retrieves the number of burnin simulations to be discarded.
}

