% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{chains}
\alias{chains}
\alias{chains,MixtureModel-method}
\title{Retrieve simulated chains from model object.}
\usage{
chains(object)

\S4method{chains}{MixtureModel}(object)
}
\arguments{
\item{object}{\code{showMethods(chains)}}
}
\description{
Retrieve simulated chains from model object.
}

