% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-DensityModel.R
\docType{methods}
\name{clusters}
\alias{clusters}
\alias{clusters,DensityModel-method}
\title{Accessor for extracting the kmeans clusters from a DensityModel
instance}
\usage{
clusters(object)

\S4method{clusters}{DensityModel}(object)
}
\arguments{
\item{object}{an instance of class 'DensityModel'}
}
\value{
k-means clustering of the component means using the modes as centers.
}
\description{
DensityModel constructor and methods are Deprecated
}

