% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{clusters}
\alias{clusters}
\title{Accessor for extracting the kmeans clusters from a DensityModel
instance}
\usage{
clusters(object)
}
\arguments{
\item{object}{an object of class 'DensityModel'}
}
\description{
Accessor for extracting the kmeans clusters from a DensityModel
instance
}
\examples{
truth <- simulateData(N=2500, p=rep(1/3, 3),
                      theta=c(-1, 0, 1),
                      sds=rep(0.1, 3))
dm <- DensityModel(truth)
clusters(dm)
}
\seealso{
\code{\link{DensityModel-class}}
}

