% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{cnv_models}
\alias{cnv_models}
\title{Fit a series of Gaussian finite mixture models at a CNV region for a collection of samples}
\usage{
cnv_models(mb, grange, snp_se, mp = McmcParams(iter = 1000, burnin = 200))
}
\arguments{
\item{mb}{a MultiBatch instance}

\item{grange}{GRanges object indicating the CNV region}

\item{snp_se}{a SummarizedExperiment with assay data containing B allele frequencies at SNPs}

\item{mp}{a McmcParams instance}
}
\value{
a MultiBatch instance
}
\description{
Fits a number of Gaussian finite mixture models at a CNV region for a collection of samples and returns a single best-fitting model.
}
\examples{
path <- system.file("extdata", package="CNPBayes")
set.seed(555)
cnp_se <- readRDS(file.path(path, "cnp_se.rds"))["CNP_022", ]
snp_se <- readRDS(file.path(path, "snp_se.rds"))
snp_se <- subsetByOverlaps(snp_se, cnp_se)
path2 <- file.path(path, "CNP_022")
mb.subsamp <- readRDS(file.path(path2, "mb_subsamp.rds"))
\dontrun{
    mb <- cnv_models(mb.subsamp,
                     rowRanges(cnp_se)["CNP_022"],
                     snp_se)
}
}
\seealso{
\code{\link{MultiBatch}}
}
