% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MultiBatchModel.R,
%   R/methods-SummarizedExperiment.R
\docType{methods}
\name{collapseBatch}
\alias{collapseBatch}
\alias{collapseBatch,MultiBatchModel-method}
\alias{collapseBatch,SummarizedExperiment-method}
\alias{collapseBatch,numeric-method}
\title{Estimate batch from any sample-level surrogate variables that capture aspects of sample processing, such as the PCR experiment (e.g., the 96 well chemistry plate), laboratory, DNA source, or DNA extraction method.}
\usage{
collapseBatch(object, provisional_batch, THR = 0.1, nchar = 8)

\S4method{collapseBatch}{MultiBatchModel}(object)

\S4method{collapseBatch}{SummarizedExperiment}(object, provisional_batch,
  THR = 0.1)

\S4method{collapseBatch}{numeric}(object, provisional_batch, THR = 0.1,
  nchar = 8)
}
\arguments{
\item{object}{see \code{showMethods(collapseBatch)}}

\item{provisional_batch}{a vector labelling from which batch each observation came from.}

\item{THR}{p-value threshold below which the null hypothesis should be rejected and batches are collapsed}

\item{nchar}{integer specifying the maximum number of characters in the batch labels}
}
\value{
The new batch value.
}
\description{
In high-throughput assays, low-level summaries of copy number at
copy number polymorphic loci (e.g., the mean log R ratio for each
sample, or a principal-component derived summary) often differ
between groups of samples due to technical sources of variation
such as reagents, technician, or laboratory.  Technical (as opposed
to biological) differences between groups of samples are referred
to as batch effects.  A useful surrogate for batch is the chemistry
plate on which the samples were hybridized. In large studies, a
Bayesian hierarchical mixture model with plate-specific means and
variances is computationally prohibitive.  However, chemistry
plates processed at similar times may be qualitatively similar in
terms of the distribution of the copy number summary statistic.
Further, we have observed that some copy number polymorphic loci
exhibit very little evidence of a batch effect, while other loci
are more prone to technical variation.  We suggest combining plates
that are qualitatively similar in terms of the Kolmogorov-Smirnov
two-sample test of the distribution and to implement this test
independently for each candidate copy number polymophism identified
in a study.  The \code{collapseBatch} function is a wrapper to the
\code{ks.test} implemented in the \code{stats} package that
compares all pairwise combinations of plates.  The \code{ks.test}
is performed recursively on the batch variables defined for a given
CNP until no batches can be combined. For smaller values of THR, plates are more likely to be judged as similar and combined.
}
\examples{
mb.ex <- MultiBatchModelExample
batches <- batch(mb.ex)
bt <- collapseBatch(y(mb.ex), batches)
batches <- as.integer(factor(bt))
hp <- hpList(k=k(mb.ex))[["MB"]]
model <- MB(dat=y(mb.ex),
            hp=hp,
            batch=batches,
            mp=mcmcParams(mb.ex))
}
