% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/functions_sc.R
\docType{methods}
\name{computeLoglik}
\alias{computeLoglik}
\alias{computeLoglik,BatchModel-method}
\alias{computeLoglik,MarginalModel-method}
\title{Compute log likelihood for a BatchModel or MarginalModel}
\usage{
computeLoglik(object)

\S4method{computeLoglik}{BatchModel}(object)

\S4method{computeLoglik}{MarginalModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(computeLoglik)}}
}
\description{
Compute log likelihood for a BatchModel or MarginalModel
}

