% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MarginalModel.R
\name{computeMarginalLik}
\alias{computeMarginalLik}
\title{Compute Marginal Likelihood}
\usage{
computeMarginalLik(y, batch, K = 1:4, T = 1000, burnin = 200, T2 = 200,
  maxperm = 3, nchains = 3, thin = 1)
}
\arguments{
\item{y}{data on which to perform MCMC simulation}

\item{batch}{a vector labelling from which batch each observation came from}

\item{K}{integer vector of possible copy numbers (latent variable states)}

\item{T}{number of MCMC iterations}

\item{burnin}{number of MCMC burnin iterations to be discarded}

\item{T2}{RS: Can you help with this paramaeter?}

\item{maxperm}{RS: Can you help with this paramaeter?}

\item{nchains}{number of parallel MCMC chains to run}

\item{thin}{number of thinning intervals}
}
\description{
This function is used to compute marginal likelihoods of posterior simulations for chains of different component sizes. Likelihoods are computed using marginal, berkhof and chibs. Summaries are calculated for each chain of each component size.
}

