% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MarginalModel.R
\name{computeMarginalLik2}
\alias{computeMarginalLik2}
\title{Estimate the marginal density of a mixture model with k components}
\usage{
computeMarginalLik2(modlist, post.iter = 200, maxperm = 3,
  method = "chib")
}
\arguments{
\item{modlist}{A list of converged models with different component sizes.}

\item{post.iter}{number of MCMC iterations after permuting the modes.  See \code{maxperm}.}

\item{maxperm}{a length-one integer vector.  For a mixture model
with K components, there are K! possible modes.  \code{maxperm}
indicates the maximum number of permutations to explore.}

\item{method}{specifies chib or laplace method}
}
\description{
This function is used to estimate the marginal density of a mixture
model with k components. A list of coverged models is passed by the user
to this function. Each model must have already converged and should be of
a different component size.
}
\seealso{
\code{\link{logBayesFactor}} for computing the bayes
factor for two models, \code{\link{orderModels}} for ordering a
list of models by decreasing marginal density, and \code{plot} for
visualizing the component densities.
}

