% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-BatchModel.R, R/methods-MarginalModel.R
\docType{methods}
\name{computeModes}
\alias{computeModes}
\alias{computeModes,BatchModel-method}
\alias{computeModes,MarginalModel-method}
\title{Calculates the modes for each of the parameters estimated by a model.}
\usage{
computeModes(object)

\S4method{computeModes}{BatchModel}(object)

\S4method{computeModes}{MarginalModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(computeModes)}}
}
\description{
Calculates the modes for each of the parameters estimated by a model.
}

