% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{consensusRegion}
\alias{consensusRegion}
\title{Identifty consensus start.}
\usage{
consensusRegion(g)
}
\arguments{
\item{g}{A GenomicRange with id metadata column.}
}
\description{
Find the minimum basepair that is spanned by at least half of all identified CNVs. RS: Will you take a look at this description?
}

