% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{SingleBatchCopyNumber-class}
\alias{SingleBatchCopyNumber-class}
\alias{MultiBatchCopyNumber-class}
\alias{MultiBatchCopyNumberPooled-class}
\title{Mixture model container where mixture components have been genotyped}
\description{
The components in a mixture model need not correspond to distinct copy number states. For example, when batch or copy number does not explain skewness or heavy-tails.
}
\details{
Suppose a mixture model with four components is selected, where the 3rd and 4th components both correspond to the diploid state.  The mapping slot will be the vector "0", "1", "2", and "2".
}
\section{Slots}{

\describe{
\item{\code{mapping}}{character string vector indicating the copy number states.  Typically '0', '1', '2', '3', or '4'.}
}}

