% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/copynumber-models.R
\docType{methods}
\name{CopyNumberModel}
\alias{CopyNumberModel}
\alias{mapping}
\alias{mapping<-}
\alias{mapping,SingleBatchCopyNumber-method}
\alias{mapping,MultiBatchCopyNumber-method}
\alias{mapping,MultiBatchCopyNumberPooled-method}
\alias{mapping,MultiBatchCopyNumber-method}
\alias{mapping<-,SingleBatchCopyNumber,numeric-method}
\alias{mapping,SingleBatchCopyNumber,numeric-method}
\alias{mapping<-,MultiBatchCopyNumber,numeric-method}
\alias{mapping,MultiBatchCopyNumber,numeric-method}
\alias{mapping<-,MultiBatchCopyNumberPooled,numeric-method}
\alias{mapping,MultiBatchCopyNumberPooled,numeric-method}
\alias{probCopyNumber}
\alias{probCopyNumber,SingleBatchCopyNumber-method}
\alias{probCopyNumber,MultiBatchCopyNumber-method}
\alias{probCopyNumber,MultiBatchCopyNumberPooled-method}
\alias{mapComponents}
\alias{SingleBatchCopyNumber}
\alias{MultiBatchCopyNumber}
\alias{MultiBatchCopyNumberPooled}
\alias{CopyNumberModel,SingleBatchModel-method}
\alias{CopyNumberModel,MultiBatchModel-method}
\alias{CopyNumberModel,MultiBatchPooled-method}
\alias{mapCopyNumber}
\title{Constructs a CopyNumberModel from SB, SBP, MB, or MBP models}
\usage{
CopyNumberModel(model)

mapping(object)

mapping(object) <- value

\S4method{mapping}{SingleBatchCopyNumber}(object)

\S4method{mapping}{MultiBatchCopyNumber}(object)

\S4method{mapping}{MultiBatchCopyNumberPooled}(object)

\S4method{mapping}{SingleBatchCopyNumber,numeric}(object) <- value

\S4method{mapping}{MultiBatchCopyNumber,numeric}(object) <- value

\S4method{mapping}{MultiBatchCopyNumberPooled,numeric}(object) <- value

probCopyNumber(model)

\S4method{probCopyNumber}{SingleBatchCopyNumber}(model)

\S4method{probCopyNumber}{MultiBatchCopyNumber}(model)

\S4method{probCopyNumber}{MultiBatchCopyNumberPooled}(model)

mapComponents(model, params = mapParams())

SingleBatchCopyNumber(model)

MultiBatchCopyNumber(model)

MultiBatchCopyNumberPooled(model)

\S4method{CopyNumberModel}{SingleBatchModel}(model)

\S4method{CopyNumberModel}{MultiBatchModel}(model)

\S4method{CopyNumberModel}{MultiBatchPooled}(model)

mapCopyNumber(model, params = mapParams())
}
\arguments{
\item{model}{a SB, SBP, MB, or MBP model}

\item{object}{a \code{SingleBatchCopyNumber} or \code{MultiBatchCopyNumber} instance}

\item{value}{a numeric vector mapping component indices to copy number state indices}

\item{params}{a list of mapping parameters}
}
\value{
a \code{MultiBatchCopyNumber} instance

a \code{MultiBatchCopyNumber} instance
}
\description{
Constructs a CopyNumberModel from SB, SBP, MB, or MBP models

Mapping of mixture component indices to copy number states

In contrast to posterior probabilities for mixture components, this function
returns posterior probabilities for distinct copy number states.
a \code{SingleBatchCopyNumber} or \code{MultiBatchCopyNumber} instance
}
\examples{
sb <- SingleBatchModelExample
cn.model <- CopyNumberModel(sb)
cn.model <- CopyNumberModel(SingleBatchModelExample)
## manually remap first two components to the same copy number state
mapping(cn.model) <- c(1, 1, 2)
\dontrun{
 ggMixture(cn.model)
}
## Batch model
bmodel <- MultiBatchModelExample
cn.model <- CopyNumberModel(bmodel)
mapping(cn.model)
\dontrun{
  ggMixture(cn.model)
}
}
