% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/copynumber-models.R
\docType{methods}
\name{copyNumber}
\alias{copyNumber}
\alias{copyNumber,SingleBatchCopyNumber-method}
\alias{copyNumber,MultiBatchCopyNumber-method}
\alias{copyNumber,MultiBatchCopyNumberPooled-method}
\title{Extract copy number estimates from a `CopyNumberModel`}
\usage{
copyNumber(object)

\S4method{copyNumber}{SingleBatchCopyNumber}(object)

\S4method{copyNumber}{MultiBatchCopyNumber}(object)

\S4method{copyNumber}{MultiBatchCopyNumberPooled}(object)
}
\arguments{
\item{object}{a \code{SingleBatchCopyNumber} or \code{MultiBatchCopyNumber} object}
}
\description{
Extract copy number estimates from a `CopyNumberModel`
}
\examples{
sb <- SingleBatchModelExample
cn.model <- CopyNumberModel(sb)
copyNumber(cn.model)

## here is an identity mapping
mapping(cn.model) <- 1:3
identical(copyNumber(cn.model), z(cn.model))
table(copyNumber(cn.model))

## here, we map the first two mixture components to one copy number state
mapping(cn.model) <- c(1, 1, 2)
table(copyNumber(cn.model))
}
\seealso{
\code{\link{mapComponents}} \code{\link{CopyNumberModel}}
}
