% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{defineCnpRegions}
\alias{defineCnpRegions}
\title{Define Copy Number Polymorphism regions}
\usage{
defineCnpRegions(grl, thr = 0.02)
}
\arguments{
\item{grl}{A GRangesList (list of GenomicRanges)}

\item{thr}{Threshold for identified CNP regions.}
}
\description{
The number of occurrences of disjoint regions are counted. A disjoint region is identified as a CNP if the number of disjoint regions contained in all of the GenomicRanges divided by the distinct GenomicRanges exceeds the argument \code{thr}. RS: will you read over this description?
}

