% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DensityModel-class}
\alias{DensityBatchModel-class}
\alias{DensityModel-class}
\title{Deprecated class for storing estimated mixture model densities}
\description{
Deprecated class for storing estimated mixture model densities
}
\section{Slots}{

\describe{
\item{\code{component}}{The component densities.}

\item{\code{overall}}{The overall (marginal across batches and components) estimate of the density.}

\item{\code{modes}}{A numeric vector providing the estimated modes in the
overall density.  The modes are defined by a crude estimate of the
first derivative of the overall density (see \code{findModes}).}

\item{\code{data}}{A numeric vector containing the data}

\item{\code{clusters}}{A vector providing the k-means clustering of the
component means using the modes as centers.  If an object of class
\code{DensityModel} is instantiated with \code{merge=FALSE}, this
slot takes values 1, ..., K, where K is the number of components.}
}}
\seealso{
\code{\link{DensityModel}}
}

