% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deprecated-classes.R
\docType{class}
\name{DensityModel-class}
\alias{DensityModel-class}
\alias{DensityBatchModel-class}
\alias{MarginalModel-class}
\alias{BatchModel-class}
\alias{DensityBatchModel-class}
\alias{BatchModel-class}
\alias{MarginalModel-class}
\title{Deprecated classes in CNPBayes}
\description{
The following classes in CNPBayes are deprecated and are provided only for compatability.

Run hierarchical MCMC for batch model.

Run marginal MCMC simulation
}
\section{Slots}{

\describe{
\item{\code{component}}{The component densities.}

\item{\code{overall}}{The overall (marginal across batches and components) estimate of the density.}

\item{\code{modes}}{A numeric vector providing the estimated modes in the
overall density.  The modes are defined by a crude estimate of the
first derivative of the overall density (see \code{findModes}).}

\item{\code{data}}{A numeric vector containing the data}

\item{\code{clusters}}{A vector providing the k-means clustering of the
component means using the modes as centers.  If an object of class
\code{DensityModel} is instantiated with \code{merge=FALSE}, this
slot takes values 1, ..., K, where K is the number of components.}

\item{\code{k}}{An integer value specifying the number of latent classes.}

\item{\code{hyperparams}}{An object of class `Hyperparameters` used to specify the hyperparameters of the model.}

\item{\code{theta}}{the means of each component and batch}

\item{\code{sigma2}}{the variances of each component and batch}

\item{\code{nu.0}}{the shape parameter for sigma2}

\item{\code{sigma2.0}}{the rate parameter for sigma2}

\item{\code{pi}}{mixture probabilities which are assumed to be the same for all batches}

\item{\code{mu}}{overall mean}

\item{\code{tau2}}{overall variance}

\item{\code{data}}{the data for the simulation.}

\item{\code{data.mean}}{the empirical means of the components}

\item{\code{data.prec}}{the empirical precisions}

\item{\code{z}}{latent variables}

\item{\code{zfreq}}{table of latent variables}

\item{\code{probz}}{n x k matrix of probabilities}

\item{\code{logprior}}{log likelihood of prior: log(p(sigma2.0)p(nu.0)p(mu))}

\item{\code{loglik}}{log likelihood: \eqn{\sum p_k \Phi(\theta_k, \sigma_k)}}

\item{\code{mcmc.chains}}{an object of class 'McmcChains' to store MCMC samples}

\item{\code{batch}}{a vector of the different batch numbers}

\item{\code{batchElements}}{a vector labeling from which batch each observation came from}

\item{\code{modes}}{the values of parameters from the iteration which maximizes log likelihood and log prior}

\item{\code{mcmc.params}}{An object of class 'McmcParams'}

\item{\code{label_switch}}{length-one logical vector indicating whether label-switching occurs (possibly an overfit model)}

\item{\code{.internal.constraint}}{Constraint on parameters. For internal use only.}

\item{\code{k}}{An integer value specifying the number of latent classes.}

\item{\code{hyperparams}}{An object of class `Hyperparameters` used to specify the hyperparameters of the model.}

\item{\code{theta}}{the means of each component and batch}

\item{\code{sigma2}}{the variances of each component and batch}

\item{\code{nu.0}}{the shape parameter for sigma2}

\item{\code{sigma2.0}}{the rate parameter for sigma2}

\item{\code{pi}}{mixture probabilities which are assumed to be the same for all batches}

\item{\code{mu}}{means from batches, averaged across batches}

\item{\code{tau2}}{variances from batches,  weighted by precisions}

\item{\code{data}}{the data for the simulation.}

\item{\code{data.mean}}{the empirical means of the components}

\item{\code{data.prec}}{the empirical precisions}

\item{\code{z}}{latent variables}

\item{\code{zfreq}}{table of latent variables}

\item{\code{probz}}{n x k matrix of probabilities}

\item{\code{logprior}}{log likelihood of prior: log(p(sigma2.0)p(nu.0)p(mu))}

\item{\code{loglik}}{log likelihood: \eqn{\sum p_k \Phi(\theta_k, \sigma_k)}}

\item{\code{mcmc.chains}}{an object of class 'McmcChains' to store MCMC samples}

\item{\code{batch}}{a vector of the different batch numbers}

\item{\code{batchElements}}{a vector labeling from which batch each observation came from}

\item{\code{modes}}{the values of parameters from the iteration which maximizes log likelihood and log prior}

\item{\code{label_switch}}{length-one logical vector indicating whether label-switching occurs (possibly an overfit model)}

\item{\code{mcmc.params}}{An object of class 'McmcParams'}

\item{\code{.internal.constraint}}{Constraint on parameters. For internal use only.}

\item{\code{k}}{An integer value specifying the number of latent classes.}

\item{\code{hyperparams}}{An object of class `Hyperparameters` used to specify the hyperparameters of the model.}

\item{\code{theta}}{the means of each component and batch}

\item{\code{sigma2}}{the variances of each component and batch}

\item{\code{nu.0}}{the shape parameter for sigma2}

\item{\code{sigma2.0}}{the rate parameter for sigma2}

\item{\code{pi}}{mixture probabilities which are assumed to be the same for all batches}

\item{\code{mu}}{means from batches, averaged across batches}

\item{\code{tau2}}{variances from batches,  weighted by precisions}

\item{\code{data}}{the data for the simulation.}

\item{\code{data.mean}}{the empirical means of the components}

\item{\code{data.prec}}{the empirical precisions}

\item{\code{z}}{latent variables}

\item{\code{zfreq}}{table of latent variables}

\item{\code{probz}}{n x k matrix of probabilities}

\item{\code{logprior}}{log likelihood of prior: log(p(sigma2.0)p(nu.0)p(mu))}

\item{\code{loglik}}{log likelihood: \eqn{\sum p_k \Phi(\theta_k, \sigma_k)}}

\item{\code{mcmc.chains}}{an object of class 'McmcChains' to store MCMC samples}

\item{\code{batch}}{a vector of the different batch numbers}

\item{\code{batchElements}}{a vector labeling from which batch each observation came from}

\item{\code{modes}}{the values of parameters from the iteration which maximizes log likelihood and log prior}

\item{\code{label_switch}}{length-one logical vector indicating whether label-switching occurs (possibly an overfit model)}

\item{\code{mcmc.params}}{An object of class 'McmcParams'}

\item{\code{.internal.constraint}}{Constraint on parameters. For internal use only.}

\item{\code{k}}{An integer value specifying the number of latent classes.}

\item{\code{hyperparams}}{An object of class `Hyperparameters` used to specify the hyperparameters of the model.}

\item{\code{theta}}{the means of each component and batch}

\item{\code{sigma2}}{the variances of each component and batch}

\item{\code{nu.0}}{the shape parameter for sigma2}

\item{\code{sigma2.0}}{the rate parameter for sigma2}

\item{\code{pi}}{mixture probabilities which are assumed to be the same for all batches}

\item{\code{mu}}{overall mean}

\item{\code{tau2}}{overall variance}

\item{\code{data}}{the data for the simulation.}

\item{\code{data.mean}}{the empirical means of the components}

\item{\code{data.prec}}{the empirical precisions}

\item{\code{z}}{latent variables}

\item{\code{zfreq}}{table of latent variables}

\item{\code{probz}}{n x k matrix of probabilities}

\item{\code{logprior}}{log likelihood of prior: log(p(sigma2.0)p(nu.0)p(mu))}

\item{\code{loglik}}{log likelihood: \eqn{\sum p_k \Phi(\theta_k, \sigma_k)}}

\item{\code{mcmc.chains}}{an object of class 'McmcChains' to store MCMC samples}

\item{\code{batch}}{a vector of the different batch numbers}

\item{\code{batchElements}}{a vector labeling from which batch each observation came from}

\item{\code{modes}}{the values of parameters from the iteration which maximizes log likelihood and log prior}

\item{\code{mcmc.params}}{An object of class 'McmcParams'}

\item{\code{label_switch}}{length-one logical vector indicating whether label-switching occurs (possibly an overfit model)}

\item{\code{.internal.constraint}}{Constraint on parameters. For internal use only.}
}}

