% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\docType{methods}
\name{tau2,MarginalModel-method}
\alias{tau2,MarginalModel-method}
\alias{bic,MarginalModel-method}
\alias{theta,MarginalModel-method}
\alias{sigma2,MarginalModel-method}
\alias{sigma2,BatchModel-method}
\alias{tau2,BatchModel-method}
\alias{theta,BatchModel-method}
\alias{marginalLikelihood,MarginalModel-method}
\alias{marginalLikelihood,MarginalModel,ANY-method}
\alias{marginalLikelihood,BatchModel-method}
\alias{marginalLikelihood,BatchModel,ANY-method}
\alias{ggMultiBatch,BatchModel-method}
\alias{DensityModel}
\alias{batch,DensityModel-method}
\alias{modes,DensityModel-method}
\alias{k,DensityModel-method}
\alias{y,DensityModel-method}
\alias{plot,DensityModel,ANY-method}
\alias{plot,DensityModel,numeric-method}
\alias{plot,MarginalModel,ANY-method}
\alias{plot,BatchModel,ANY-method}
\alias{plot,DensityBatchModel,ANY-method}
\alias{plot,BatchModel,ANY-method}
\alias{HyperparametersBatch}
\alias{HyperparametersMarginal}
\alias{BatchModelList}
\alias{BatchModel}
\alias{posteriorSimulation,MixtureModel,integer-method}
\alias{posteriorSimulation,MixtureModel,numeric-method}
\alias{DensityModel}
\alias{batch,DensityModel-method}
\alias{modes,DensityModel-method}
\alias{k,DensityModel-method}
\alias{y,DensityModel-method}
\alias{plot,DensityModel,ANY-method}
\alias{plot,DensityModel,numeric-method}
\alias{plot,MarginalModel,ANY-method}
\alias{plot,BatchModel,ANY-method}
\alias{plot,DensityBatchModel,ANY-method}
\alias{plot,BatchModel,ANY-method}
\alias{ggMultiBatch,MultiBatchModel-method}
\alias{ggMultiBatch,MultiBatchPooled-method}
\alias{ggSingleBatch,MarginalModel-method}
\alias{ggSingleBatch,SingleBatchModel-method}
\alias{ggMultiBatch,MultiBatchCopyNumber-method}
\alias{ggSingleBatch,SingleBatchCopyNumber-method}
\alias{downsample}
\alias{downSampleEachBatch}
\alias{MultiBatchModel}
\alias{posteriorSimulation,list,ANY-method}
\alias{posteriorSimulation,list-method}
\alias{plot}
\alias{clusters}
\alias{labelSwitching}
\alias{labelSwitching,MixtureModel-method}
\alias{[,BatchModel,ANY,ANY,ANY-method}
\alias{[,BatchModel-method}
\alias{[,BatchModel,ANY-method}
\alias{[,BatchModel,ANY,ANY-method}
\alias{SingleBatchModel}
\alias{multiBatchDensities}
\title{DensityModel constructor has been deprecated.}
\usage{
\S4method{tau2}{MarginalModel}(object)

\S4method{bic}{MarginalModel}(object)

\S4method{theta}{MarginalModel}(object)

\S4method{sigma2}{MarginalModel}(object)

\S4method{sigma2}{BatchModel}(object)

\S4method{tau2}{BatchModel}(object)

\S4method{theta}{BatchModel}(object)

\S4method{marginalLikelihood}{MarginalModel}(model, params = mlParams())

\S4method{marginalLikelihood}{BatchModel}(model, params = mlParams())

\S4method{ggMultiBatch}{BatchModel}(model, bins)

DensityModel(object, merge = FALSE)

\S4method{batch}{DensityModel}(object)

\S4method{modes}{DensityModel}(object)

\S4method{k}{DensityModel}(object)

\S4method{y}{DensityModel}(object)

\S4method{plot}{DensityModel,ANY}(x, y, ...)

\S4method{plot}{MarginalModel,ANY}(x, y, ...)

\S4method{plot}{BatchModel,ANY}(x, y, show.batch = TRUE, ...)

\S4method{plot}{DensityBatchModel,ANY}(x, show.batch = TRUE, ...)

HyperparametersBatch(k = 3L, mu.0 = 0, tau2.0 = 100, eta.0 = 1800,
  m2.0 = 1/60, alpha, beta = 0.1, a = 1.8, b = 6)

HyperparametersMarginal(k = 0L, mu.0 = 0, tau2.0 = 100, eta.0 = 1,
  m2.0 = 0.1, alpha, beta = 0.1, a = 1.8, b = 6)

BatchModelList(data = numeric(), k = numeric(), batch,
  mcmc.params = McmcParams(), ...)

BatchModel(data = numeric(), k = 3, batch, hypp, mcmc.params)

\S4method{posteriorSimulation}{MixtureModel,integer}(object, k)

\S4method{posteriorSimulation}{MixtureModel,numeric}(object, k)

DensityModel(object, merge = FALSE)

\S4method{batch}{DensityModel}(object)

\S4method{modes}{DensityModel}(object)

\S4method{k}{DensityModel}(object)

\S4method{y}{DensityModel}(object)

\S4method{plot}{DensityModel,ANY}(x, y, ...)

\S4method{plot}{MarginalModel,ANY}(x, y, ...)

\S4method{plot}{BatchModel,ANY}(x, y, show.batch = TRUE, ...)

\S4method{plot}{DensityBatchModel,ANY}(x, show.batch = TRUE, ...)

\S4method{ggMultiBatch}{MultiBatchModel}(model, bins)

\S4method{ggMultiBatch}{MultiBatchPooled}(model, bins)

\S4method{ggSingleBatch}{MarginalModel}(model, bins)

\S4method{ggSingleBatch}{SingleBatchModel}(model, bins)

\S4method{ggMultiBatch}{MultiBatchCopyNumber}(model, bins)

\S4method{ggSingleBatch}{SingleBatchCopyNumber}(model, bins)

downsample(batch.file, plate, y, ntiles = 250, THR = 0.1)

downSampleEachBatch(y, nt, batch)

MultiBatchModel(data = numeric(), k = 3, batch, hypp, mcmc.params)

\S4method{posteriorSimulation}{list,ANY}(object)

plot(x, y, ...)

clusters(object)

labelSwitching(object, merge = TRUE)

\S4method{labelSwitching}{MixtureModel}(object, merge = TRUE)

\S4method{[}{BatchModel,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

SingleBatchModel(data = numeric(), k = 3, hypp, mcmc.params)

multiBatchDensities(model)
}
\arguments{
\item{object}{see \code{showMethods(DensityModel)}}

\item{model}{MarginalModel}

\item{params}{list of parameters for computing marginal likelihood}

\item{bins}{length-one numeric vector specifying number of bins for plotting}

\item{merge}{Logical.  Whether to use kmeans clustering to cluster
the component means using the estimated modes from the overall
density as the centers for the \code{kmeans} function.}

\item{x}{a \code{DensityModel}-derived object, or a
\code{MixtureModel}-derived object.
numeric vector of the one-dimensional summaries for a given copy
number polymorphism. If \code{x} is a \code{MixtureModel}, \code{y}
is ignored.}

\item{y}{in memory data}

\item{...}{additional arguments to \code{HyperparametersBatch}}

\item{show.batch}{a logical. If true, batch specific densities
will be plotted.}

\item{k}{length-one integer vector specifying number of components
(typically 1 <= k <= 4)}

\item{mu.0}{length-one numeric vector of the mean for the normal
prior of the component means}

\item{tau2.0}{length-one numeric vector of the variance for the normal
prior of the component means}

\item{eta.0}{length-one numeric vector of the shape parameter for
the Inverse Gamma prior of the component variances.  The shape
parameter is parameterized as 1/2 * eta.0.}

\item{m2.0}{length-one numeric vector of the rate parameter for
the Inverse Gamma prior of the component variances.  The rate
parameter is parameterized as 1/2 * eta.0 * m2.0.}

\item{alpha}{length-k numeric vector of the shape parameters for
the dirichlet prior on the mixture probabilities}

\item{beta}{length-one numeric vector for the parameter of the
geometric prior for nu.0 (nu.0 is the shape parameter of the
Inverse Gamma sampling distribution for the component-specific
variances).  beta is a probability and must be in the interval
[0,1].}

\item{a}{length-one numeric vector of the shape parameter for the
Gamma prior used for sigma2.0 (sigma2.0 is the shape parameter of
the Inverse Gamma sampling distribution for the component-specific
variances)}

\item{b}{a length-one numeric vector of the rate parameter for the
Gamma prior used for sigma2.0 (sigma2.0 is the rate parameter of
the Inverse Gamma sampling distribution for the component-specific
variances)}

\item{data}{numeric vector of average log R ratios}

\item{batch}{a vector of the different batch numbers (must be sorted)}

\item{mcmc.params}{a \code{McmcParams} object}

\item{hypp}{An object of class `Hyperparameters` used to specify the hyperparameters of the model.}

\item{batch.file}{the name of a file contaning RDS data to be read in.}

\item{plate}{a vector containing the labels  from which batch each observation came from.}

\item{ntiles}{number of tiles in a batch}

\item{THR}{threshold above which to merge batches in Kolmogorov-Smirnov test.}

\item{nt}{the number of observations per batch}

\item{i}{integer}

\item{j}{integer}

\item{drop}{Not used.}
}
\value{
An object of class 'DensityModel'

An object of class HyperparametersBatch

An object of class HyperparametersMarginal

a list. Each element of the list is a \code{BatchModel}

An object of class `BatchModel`

An object of class 'DensityModel'

Tile labels for each observation

Tile labels for each observation

An object of class `MultiBatchModel`

A plot showing the density estimate

A single proportion for a \code{MarginalModel} or a vector of proportions, one for each batch for a \code{BatchModel}

An object of class 'BatchModel'
}
\description{
DensityModel constructor has been deprecated.

Instantiates an instance of 'DensityModel' (or 'DensityBatchModel')
from a MarginalModel or BatchModel object. See the corresponding
class for additional details and examples.

Create an object for running hierarchical MCMC simulations.

DensityModel constructor and methods are Deprecated
}
\seealso{
See \code{\link{ggSingleBatch}} and \code{\link{ggMultiBatch}} for visualization

\code{\link{BatchModel}}.  For single-batch data, use

\code{\link{DensityModel-class}} \code{\link{kmeans}}

\code{\link[dplyr]{ntile}}
}
