% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\docType{methods}
\name{dfr,MixtureModel-method}
\alias{dfr,MixtureModel-method}
\title{Accessor for degrees of freedom}
\usage{
\S4method{dfr}{MixtureModel}(object)
}
\arguments{
\item{object}{a MixtureModel-derived object}
}
\description{
Accessor for degrees of freedom
}
