% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{dfr}
\alias{dfr}
\title{Accessor for degrees of freedom}
\usage{
dfr(object)
}
\arguments{
\item{object}{a Hyperparameters or Hyperparameters-derived class}
}
\description{
Accessor for degrees of freedom
}
\examples{
hp <- Hyperparams()
dfr(hp)
}
