% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{eta.0}
\alias{eta.0}
\alias{eta.0,Hyperparameters-method}
\alias{eta.0,MixtureModel-method}
\title{Retrieve the rate parameter for the tau2 distribution.}
\usage{
eta.0(object)

\S4method{eta.0}{MixtureModel}(object)

\S4method{eta.0}{Hyperparameters}(object)
}
\arguments{
\item{object}{see \code{showMethods(eta.0)}}
}
\value{
eta.0 of a 'MixtureModel'
}
\description{
Retrieve the rate parameter for the tau2 distribution.
}
\examples{
eta.0(MarginalModelExample)
}

