% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-McmcChains.R,
%   R/methods-MultiBatchModel.R
\docType{methods}
\name{[,McmcChains,ANY,ANY,ANY-method}
\alias{[,McmcChains,ANY,ANY,ANY-method}
\alias{[,McmcChains-method}
\alias{[,McmcChains,ANY-method}
\alias{extract}
\alias{[,MultiBatchModel-method}
\alias{[,MultiBatchModel,ANY-method}
\alias{[,MultiBatchModel,ANY,ANY-method}
\alias{[,MultiBatchModel,ANY,ANY,ANY-method}
\title{extract estimated parameters at particular iteration of simulation.}
\usage{
\S4method{[}{McmcChains,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{MultiBatchModel,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{An object of class MultiBatchModel, McmcChains, or McmcParams}

\item{i}{An element of the instance to be extracted.}

\item{j}{Not used.}

\item{...}{Not used.}

\item{drop}{Not used.}
}
\value{
An object of class 'McmcChains'

An object of class 'MultiBatchModel'
}
\description{
extract estimated parameters at particular iteration of simulation.

extract data, latent variable, and batch for given observation
}
