% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{fit_restricted2}
\alias{fit_restricted2}
\title{Wrapper for running many independent chains with a short burnin and one longer run}
\usage{
fit_restricted2(mb, model = "MBP2", ...)
}
\arguments{
\item{mb}{a `MultiBatch` object}

\item{model}{name of model to evaluate (e.g., MB2, MBP2, SB2, ...).}

\item{...}{additional arguments passed to warmup}
}
\description{
Wrapper for running many independent chains with a short burnin and one longer run
}
\seealso{
\code{\link{warmup}}
}
