% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{genotype_model}
\alias{genotype_model}
\title{Map mixture component labels to integer copy numbers}
\usage{
genotype_model(model, snpdat)
}
\arguments{
\item{model}{a `MultiBatch` instance}

\item{snpdat}{a `SummarizedExperiment` containing B allele frequences}
}
\value{
a `MultiBatch` instance. The mapping is now accessible through `mapping`.
}
\description{
Mapping is performed using the B allele frquencies for SNPs spanned by the candidate deletion/duplication polymorphism. Multiple components can be mapped to a single copy number state, but one-to-many mapping is not permitted
}
