% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{ggPredictive}
\alias{ggPredictive}
\title{Compare the posterior predictive distribution to the empirical data}
\usage{
ggPredictive(model, predict, adjust = 1/3)
}
\arguments{
\item{model}{a SB, MB, SBP, or MBP model}

\item{predict}{a \code{tibble} of the posterior predictive values, batch (only for MB and MBP models), and mixture component assignments}

\item{adjust}{a length-one numeric vector passed to \code{geom_density} -- controls the smoothness of the kernal density}
}
\value{
a `gg` object
}
\description{
Compare the posterior predictive distribution to the empirical data
}
\examples{
  bmodel <- MultiBatchModelExample
  mp <- McmcParams(iter=500, burnin=150, nStarts=4)
  mcmcParams(bmodel) <- mp
  \dontrun{
     ## this is preferred to posteriorSimulation, but takes longer
     bmodel <- gibbs(model="MB", dat=y(bmodel), mp=mp, hp.list=hpList()[["MB"]],
                     batches=batch(bmodel))
  }
  bmodel <- posteriorSimulation(bmodel)
  tab <- posteriorPredictive(bmodel)
  \dontrun{
      ggPredictive(bmodel, tab)
  }
}
