% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{gibbs}
\alias{gibbs}
\title{Run a Gibbs sampler on one or multiple types of Bayesian Gaussian mixture models}
\usage{
gibbs(model = c("SB", "MB", "SBP", "MBP"), dat, mp, hp.list, batches,
  k_range = c(1, 4), max_burnin = 32000)
}
\arguments{
\item{model}{a character vector indicating which models to fit (any combination of 'SB', 'MB', 'SBP', and 'MBP')}

\item{dat}{numeric vector of the summary copy number data for each sample at a single CNP (e.g., the median log R ratio for each sample)}

\item{hp.list}{a list of hyperparameters for each of the different models.  If missing, this list will be generated automatically with default hyperparameters that work well for copy number data}

\item{batches}{an integer vector of the same length as \code{dat} indicating the batch in which the sample was processed}

\item{k_range}{a length-two numeric vector providing the minimum and maximum number of components to model.  For example, c(1, 3) will fit mixture models with 1, 2, and 3 components.}

\item{max_burnin}{the maximum number of burnin iterations. See details.}
}
\description{
Model types:\\
SB (SingleBatchModel):  hierarchical model with mixture component-specific means and variances\\
MB (MultiBatchModel): hierarchical model with mixture component- and batch-specific means and variances\\
SBP (SingleBatchPooled):  similar to SB model but with a pooled estimate of the variance across all mixture components\\
MBP (MultiBatchPooled):  similar to MB model but with a pooled estimate of the variance (across mixture components) for each batch.
}
\details{
For each model specified, a Gibbs sampler will be initiated for \code{nStarts} independently simulated starting values (we suggest \code{nStarts > 2}).  The burnin, number of iterations after burnin, and the thin parameters are specified in the \code{mp} argument.    If the effective number of independent MCMC draws for any of the parameter chains is less than 500 or if the multivariate Gelman Rubin convergence diagnostic is less than 1.2, the thin and the burnin will be doubled and we start over -- new chains are initalized independently and the Gibbs sampler is restarted. This process is repeated until the effective sample size is greater than 500 and the Gelman Rubin convergence diagnostic is less than 1.2.
}
\examples{
  set.seed(100)
  nbatch <- 3
  k <- 3
  means <- matrix(c(-2.1, -2, -1.95, -0.41, -0.4, -0.395, -0.1,
      0, 0.05), nbatch, k, byrow = FALSE)
  sds <- matrix(0.15, nbatch, k)
  sds[, 1] <- 0.3
  N <- 1000
  truth <- simulateBatchData(N = N, batch = rep(letters[1:3],
                                                length.out = N),
                             p = c(1/10, 1/5, 1 - 0.1 - 0.2),
                             theta = means,
                             sds = sds)

}
\seealso{
\code{\link[coda]{gelman.diag}} \code{\link[coda]{effectiveSize}}
}
