% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{hemdel_model}
\alias{hemdel_model}
\title{Fits a series of models consistent for a deletion CNV where only hemizygous deletions are present (no homozygous deletions)}
\usage{
hemdel_model(mb, mp, ...)
}
\arguments{
\item{mb}{a MultiBatch instance}

\item{mp}{a McmcParams instance}

\item{...}{additional arguments passed to warmup}
}
\description{
Fits a series of models consistent for a deletion CNV where only hemizygous deletions are present (no homozygous deletions)
}
\seealso{
[warmup()]
}
