% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{homdel_model}
\alias{homdel_model}
\title{Wrapper for evaluating mixture models consistent with copy number deletion polymorphisms}
\usage{
homdel_model(mb, mp, augment = TRUE, ...)
}
\arguments{
\item{mb}{a `MultiBatch` instance}

\item{mp}{a `McmcParams` instance}

\item{augment}{whether to use data augmentation for rare observations that may be present in a subset of batches}

\item{...}{additional arguments passed to \code{warmup}}
}
\description{
Fits single-batch and multi-batch models consistent with a deletion polymorphism.
}
\seealso{
\code{\link{warmup}}
}
