% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{homdeldup_model}
\alias{homdeldup_model}
\title{Wrapper for evaluating mixture models consistent with copy number deletion and gain polymorphisms}
\usage{
homdeldup_model(mb, mp, augment = TRUE, ...)
}
\arguments{
\item{mb}{a `MultiBatch` instance}

\item{mp}{a `McmcParams` instance}

\item{augment}{whether to use data augmentation for rare observations that may be present in a subset of batches}
}
\description{
Wrapper for evaluating mixture models consistent with copy number deletion and gain polymorphisms
}
\seealso{
[mcmcParams()]
}
