% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-Hyperparameters.R
\name{Hyperparameters}
\alias{Hyperparameters}
\title{Create an object of class 'Hyperparameters'}
\usage{
Hyperparameters(type = "batch", k = 2L, ...)
}
\arguments{
\item{type}{specifies 'marginal' or 'batch'}

\item{k}{number of components}

\item{...}{optional parameters, i.e. mu.0, tau2.0, eta.0, etc.}
}
\description{
Create an object of class 'Hyperparameters'
}

