% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{HyperparametersMarginal-class}
\alias{HyperparametersMarginal-class}
\title{An object to specify the hyperparameters of a marginal model.}
\description{
This class inherits from the Hyperparameters class. This class is for hyperparameters which are marginal over the batches.
}
\section{Slots}{

\describe{
\item{\code{k}}{Number of components}

\item{\code{mu.0}}{a priori mean?}

\item{\code{tau2.0}}{a priori precision?}

\item{\code{eta.0}}{?}

\item{\code{m2.0}}{?}

\item{\code{alpha}}{gamma parameter?}

\item{\code{beta}}{gamma parameter?}

\item{\code{a}}{?}

\item{\code{b}}{?}
}}

