% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-Hyperparameters.R
\name{HyperparametersMarginal}
\alias{HyperparametersMarginal}
\title{Create an object of class 'HyperparametersMarginal' for the
marginal mixture model}
\usage{
HyperparametersMarginal(k = 0L, mu.0 = 0, tau2.0 = 100, eta.0 = 1,
  m2.0 = 0.1, alpha, beta = 0.1, a = 1.8, b = 6)
}
\arguments{
\item{k}{length-one integer vector specifying number of components
(typically 1 <= k <= 4)}

\item{mu.0}{length-one numeric vector of the mean for the normal
prior of the component means}

\item{tau2.0}{length-one numeric vector of the variance for the normal
prior of the component means}

\item{eta.0}{length-one numeric vector of the shape parameter for
the Inverse Gamma prior of the component variances.  The shape
parameter is parameterized as 1/2 * eta.0.}

\item{m2.0}{length-one numeric vector of the rate parameter for
the Inverse Gamma prior of the component variances.  The rate
parameter is parameterized as 1/2 * eta.0 * m2.0.}

\item{alpha}{length-k numeric vector of the shape parameters for
the dirichlet prior on the mixture probabilities}

\item{beta}{length-one numeric vector for the parameter of the
geometric prior for nu.0 (nu.0 is the shape parameter of the
Inverse Gamma sampling distribution for the component-specific
variances).  beta is a probability and must be in the interval
[0,1].}

\item{a}{length-one numeric vector of the shape parameter for the
Gamma prior used for sigma2.0 (sigma2.0 is the shape parameter of
the Inverse Gamma sampling distribution for the component-specific
variances)}

\item{b}{a length-one numeric vector of the rate parameter for the
Gamma prior used for sigma2.0 (sigma2.0 is the rate parameter of
the Inverse Gamma sampling distribution for the component-specific
variances)}
}
\value{
An object of class HyperparametersMarginal
}
\description{
Create an object of class 'HyperparametersMarginal' for the
marginal mixture model
}
\examples{
HyperparametersMarginal(k=3)
}

