% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Hyperparameters.R
\name{HyperparametersTrios}
\alias{HyperparametersTrios}
\title{Create an object of class 'Hyperparameters' with additional parameters for Trios}
\usage{
HyperparametersTrios(states = 1:4, k = length(states), mu.0 = 0,
  tau2.0 = 0.4, eta.0 = 32, m2.0 = 0.5, alpha, beta = 0.1, a = 1.8,
  b = 6, dfr = 100)
}
\arguments{
\item{states}{specifies the specific copy number states}
}
\value{
An object of class HyperparameterTrios
}
\description{
Create an object of class 'Hyperparameters' with additional parameters for Trios
}
\examples{
    hyp.trio <- HyperparameterTrios(states=1:4)
    
}
