% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{hyperParams}
\alias{hyperParams}
\alias{hyperParams,MixtureModel-method}
\title{Retried specified hyper parameters from model.}
\usage{
hyperParams(object)

\S4method{hyperParams}{MixtureModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(hyperParams)}}
}
\description{
Retried specified hyper parameters from model.
}

