% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{id}
\alias{id}
\alias{id,MultiBatch-method}
\alias{id,MultiBatchList-method}
\title{Accessor for sample id}
\usage{
id(object)
}
\arguments{
\item{object}{A MultiBatch instance}
}
\description{
Extract vector of sample identifiers
}
