% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{isSimulated}
\alias{isSimulated}
\alias{isSimulated,MixtureModel-method}
\alias{isSimulated,MultiBatch-method}
\alias{isSimulated,MultiBatchList-method}
\title{Accessor for which observations were simulated}
\usage{
isSimulated(object)
}
\arguments{
\item{object}{A MultiBatch instance}
}
\description{
Additional details about data augmentation are provided in the vignette.
}
