% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcParams.R, R/methods-MixtureModel.R, R/methods-ModelList.R
\docType{methods}
\name{iter<-}
\alias{iter}
\alias{iter,McmcParams-method}
\alias{iter,MixtureModel-method}
\alias{iter,ModelList-method}
\alias{iter<-}
\alias{iter<-,McmcParams-method}
\alias{iter<-,MixtureModel-method}
\alias{iter<-,ModelList-method}
\title{Reset number of iterations.}
\usage{
iter(object, force = FALSE) <- value

iter(object)

\S4method{iter}{McmcParams}(object)

\S4method{iter}{McmcParams}(object, force = FALSE) <- value

\S4method{iter}{MixtureModel}(object)

\S4method{iter}{MixtureModel}(object, force = FALSE) <- value

\S4method{iter}{ModelList}(object)

\S4method{iter}{ModelList}(object) <- value
}
\arguments{
\item{object}{see \code{showMethods(iter)}}

\item{force}{Allow changing of the size of the elements?}

\item{value}{new number of iterations}
}
\description{
This function changes the number of simulations.

This function retrieves the number of iterations of an MCMC simulation.
}

