% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{iter<-}
\alias{iter<-}
\title{Reset number of iterations.}
\usage{
iter(object, force = FALSE) <- value
}
\arguments{
\item{object}{see \code{showMethods("iter<-")}}

\item{force}{Allow changing of the size of the elements?}

\item{value}{new number of iterations}
}
\description{
This function changes the number of simulations.
}

