% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{iter}
\alias{iter}
\title{Number of MCMC iterations.}
\usage{
iter(object)
}
\arguments{
\item{object}{see \code{showMethods(iter)}}
}
\description{
This function retrieves the number of iterations of an MCMC simulation.
}

