% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Hyperparameters.R
\name{k}
\alias{k}
\alias{k,McmcChains-method}
\alias{k,MixtureModel-method}
\alias{k,MultiBatch-method}
\alias{k,MultiBatchList-method}
\alias{k<-}
\alias{k<-,MultiBatchList,ANY-method}
\alias{k<-,Hyperparameters-method}
\alias{k,Hyperparameters-method}
\title{Number of components.}
\usage{
k(object)

k(object) <- value

\S4method{k}{Hyperparameters}(object) <- value

\S4method{k}{Hyperparameters}(object)
}
\arguments{
\item{object}{A Hyperparameters instance}

\item{value}{An integer specifying the number of mixture components for the mixture model}
}
\description{
Number of components.

Accessor for number of mixture components
}
