% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{k}
\alias{k}
\title{Number of components.}
\usage{
k(object)
}
\arguments{
\item{object}{see \code{showMethods(k)}}
}
\description{
This function retrieves the number of a priori components.
}

