% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{label_switch}
\alias{label_switch}
\title{Accessor for determing whether label switching occurred during MCMC}
\usage{
label_switch(object)
}
\arguments{
\item{a}{MixtureModel-derived class}
}
\description{
Accessor for determing whether label switching occurred during MCMC
}
\examples{
label_switch(MarginalModelExample)

}
