% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{labelSwitching}
\alias{labelSwitching}
\title{Calculate proportion of relabeling instances}
\usage{
labelSwitching(object, merge = TRUE)
}
\arguments{
\item{object}{An object of class \code{MarginalModel} or \code{BatchModel}}

\item{merge}{A logical indicating whether the components should be merged before checking for label switching}
}
\value{
A single proportion for a \code{MarginalModel} or a vector of proportions, one for each batch for a \code{BatchModel}
}
\description{
When fitting an object of class \code{MixtureModel}, label switching can occur i.e. the mean
of component one can be less than the mean of component two at one iteration of the
MCMC sampler and at the next instance, the order is switched. Label switching should be
kept at a minimum. This function returns the proportion of MCMC sample iterations where
label switching has occurred.
}
\examples{
labelSwitching(MarginalModelExample)
}

