% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-ModelList.R
\docType{methods}
\name{length,ModelList-method}
\alias{length,ModelList-method}
\title{Retrieve the number of Marginal or BatchModels contained.}
\usage{
\S4method{length}{ModelList}(x)
}
\arguments{
\item{x}{a ModelList of MarginalModel or BatchModel.}
}
\description{
Retrieve the number of Marginal or BatchModels contained.
}

