% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R
\docType{methods}
\name{log_lik}
\alias{log_lik}
\alias{log_lik,McmcChains-method}
\title{Retrieve log likelihood.}
\usage{
log_lik(object)

\S4method{log_lik}{McmcChains}(object)
}
\arguments{
\item{object}{see showMethods(log_lik)}
}
\value{
The log likelihood
}
\description{
Retrieve log likelihood.
}
\examples{
log_lik(SingleBatchModelExample)
}
