% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MarginalModel.R,
%   R/methods-SingleBatchModel.R
\name{logBayesFactor}
\alias{logBayesFactor}
\alias{logBayesFactor}
\title{Compute the log bayes factor between models.}
\usage{
logBayesFactor(x)

logBayesFactor(x)
}
\arguments{
\item{x}{the result of a call to \code{computeMarginalLik}.}

\item{x}{the result of a call to \code{computeMarginalLik}.}
}
\value{
Log Bayes factor comparing the two models with highest likelihood.

Log Bayes factor comparing the two models with highest likelihood.
}
\description{
Models of varying component sizes are compared. The log bayes factor is 
calculated comparing each set of two models by marginal likelihood, 
as computed by \code{marginalLikelihood}.

Models of varying component sizes are compared. The log bayes factor is 
calculated comparing each set of two models by marginal likelihood, 
as computed by \code{marginalLikelihood}.
}
