% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R, R/methods-MixtureModel.R
\docType{methods}
\name{logLik}
\alias{logLik}
\alias{logLik,McmcChains-method}
\alias{logLik,MixtureModel-method}
\title{Retrieve log likelihood.}
\usage{
logLik(object, ...)

\S4method{logLik}{McmcChains}(object)

\S4method{logLik}{MixtureModel}(object)
}
\arguments{
\item{object}{see showMethods(logLik)}
}
\description{
Retrieve log likelihood.
}

