% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{logpotential}
\alias{logpotential}
\title{Retrieve the log potential from a MixtureModel}
\usage{
logpotential(object)
}
\arguments{
\item{object}{a MixtureModel.}
}
\description{
The log potential of a model is the log(likelihood * likelihood of prior) + log(probability(theta)) + log(probability(sigma.2))
}

