% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R, R/methods-MixtureModel.R
\docType{methods}
\name{logPrior}
\alias{logPrior}
\alias{logPrior,McmcChains-method}
\alias{logPrior,MixtureModel-method}
\title{Calculate log likelihood of prior for model}
\usage{
logPrior(object)

\S4method{logPrior}{McmcChains}(object)

\S4method{logPrior}{MixtureModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(logPrior)}}
}
\value{
log likelihood of the prior.
}
\description{
Calculate log likelihood of prior for model
}
\examples{
     logPrior(MarginalModelExample)
}

