% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{m2.0}
\alias{m2.0}
\alias{m2.0,Hyperparameters-method}
\alias{m2.0,MixtureModel-method}
\title{Retrieve the shape parameter for the tau2 distribution.}
\usage{
m2.0(object)

\S4method{m2.0}{MixtureModel}(object)

\S4method{m2.0}{Hyperparameters}(object)
}
\arguments{
\item{object}{see \code{showMethods(m2.0)}}
}
\value{
m2.0 for a model
}
\description{
Retrieve the shape parameter for the tau2 distribution.
}
\examples{
     m2.0(MarginalModelExample)
}

