% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{map_z}
\alias{map_z}
\title{Calculate the maximum a posteriori estimate of latent variable assignment.}
\usage{
map_z(object)
}
\arguments{
\item{object}{an object of class MixtureModel.}
}
\value{
map estimate of latent variable assignment for each observation
}
\description{
Calculate the maximum a posteriori estimate of latent variable assignment.
}
\examples{
     map_z(MarginalModelExample)
}
