% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{mapCnProbability}
\alias{mapCnProbability}
\title{Probabiliistic copy number assigments.}
\usage{
mapCnProbability(model)
}
\arguments{
\item{model}{An object of class MixtureModel.}
}
\description{
Calculate probabilistic copy number assignments using Bayes Rule applied at the MAP estimates of the cluster mean, variance, and class proportion parameters
}

