% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber-models.R
\name{mapCopyNumber}
\alias{mapCopyNumber}
\title{Name biological equivalent of components}
\usage{
mapCopyNumber(model, params = mapParams())
}
\arguments{
\item{x}{A vector of component LRR means}

\item{approx}{A vector indicating the approximate locations of labels}

\item{homozygous_check}{A boolean indicating whether to check for homozygous deletion first}
}
\value{
A vector of labels e.g. \code{c("Homozygous Deletion", "Diploid")}
}
\description{
Name biological equivalent of components
}
\examples{
cn.model <- SingleBatchModel(MarginalModelExample)
mapping(cn.model) <- mapComponents(cn.model)
mapCopyNumber(MarginalModelExample)
}

