% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber-models.R
\name{mapParams}
\alias{mapParams}
\title{Parameters for merging two components}
\usage{
mapParams(threshold = 0.1, proportion.subjects = 0.5)
}
\arguments{
\item{threshold}{numeric value in [0, 0.5]. For a given observation (sample),
mixture component probabilities > threshold and less than 1-threshold are
combined.}

\item{proportion.subjects}{numeric value in [0, 1]. Two components are
combined if the fraction of subjects with component probabilities in the
range [threshold, 1-threshold] exceeds this value.}
}
\description{
Parameters for merging two components
}

