% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{mapping}
\alias{mapping}
\alias{mapping<-}
\title{Map mixture components to copy number states}
\usage{
mapping(object)

mapping(object) <- value
}
\arguments{
\item{object}{a SB, SBP, MB, or MBP model}

\item{value}{a k-length numeric vector with values in {1, 2, ..., k}, where k is the number of mixture components}
}
\description{
Map mixture components to copy number states
}
\examples{
cn.model <- CopyNumberModel(SingleBatchModelExample)
## manually remap first two components to the same copy number state
\dontrun{
 mapping(cn.model) <- c(1, 1, 2)
 ggMixture(cn.model)
}
}
\seealso{
\code{\link{CopyNumberModel}}
}
