% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{marginal_lik}
\alias{marginal_lik}
\alias{marginal_lik,MixtureModel-method}
\alias{marginal_lik,MultiBatch-method}
\alias{marginal_lik,MultiBatchList-method}
\alias{marginal_lik<-}
\alias{marginal_lik<-,MixtureModel,numeric-method}
\alias{marginal_lik<-,MultiBatch,numeric-method}
\title{Accessor for the log marginal likelihood of a SB, SBP, MB, or MBP model}
\usage{
marginal_lik(object)

marginal_lik(object) <- value
}
\arguments{
\item{object}{a SB, SBP, MB, or MBP model}

\item{value}{the marginal likelihood}
}
\description{
The marginal likelihood is computed by Chib's estimator (JASA, Volume 90 (435), 1995).
}
\examples{
sb <- SingleBatchModelExample
marginal_lik(sb)
}
\seealso{
See \code{\link{marginalLikelihood}} for computing the marginal likelihood of a mixture model.
}
