% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{marginalLik}
\alias{marginalLik}
\title{Extract marginal likelihoods from a list of models}
\usage{
marginalLik(models)
}
\arguments{
\item{models}{list of models}
}
\description{
Extract marginal likelihoods from a list of models
}
