% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/marginal_likelihood.R
\docType{methods}
\name{marginalLikelihood}
\alias{marginalLikelihood}
\alias{marginalLikelihood,BatchModel,ANY-method}
\alias{marginalLikelihood,BatchModel-method}
\alias{marginalLikelihood,MarginalModel,ANY-method}
\alias{marginalLikelihood,MarginalModel-method}
\alias{marginalLikelihood,SingleBatchPooledVar,ANY-method}
\alias{marginalLikelihood,SingleBatchPooledVar-method}
\alias{marginalLikelihood,list,ANY-method}
\alias{marginalLikelihood,list-method}
\title{Compute the marginal likelihood of a converged model.}
\usage{
marginalLikelihood(model, params = list(niter = 1000L, root = (1/10),
  reject.threshold = 1e-50, prop.threshold = 0.5))

\S4method{marginalLikelihood}{MarginalModel}(model, params = list(niter =
  1000L, root = (1/10), reject.threshold = 1e-50, prop.threshold = 0.5))

\S4method{marginalLikelihood}{SingleBatchPooledVar}(model, params = list(niter
  = 1000L, root = (1/10), reject.threshold = 1e-50, prop.threshold = 0.5))

\S4method{marginalLikelihood}{BatchModel}(model, params = list(niter = 1000L,
  root = (1/10), reject.threshold = 1e-50, prop.threshold = 0.5))

\S4method{marginalLikelihood}{list}(model, params = list(niter = 1000L, root =
  (1/10), reject.threshold = 1e-50, prop.threshold = 0.5))
}
\arguments{
\item{model}{An object of class \code{MarginalModel}, or a list of
\code{MarginalModel}'s. Can also be an object of \code{BatchModel} or
a list of such models.}

\item{niter}{The number of iterations for the reduced Gibb's sampler.}
}
\value{
A vector of the marginal likelihood of the model(s)
}
\description{
Compute the marginal likelihood of a converged model.
}
\examples{
     marginalLikelihood(MarginalModelExample)
}

