% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/marginal_likelihood.R
\name{marginalLikelihood}
\alias{marginalLikelihood}
\title{Compute the marginal likelihood of a converged model.}
\usage{
marginalLikelihood(model, niter = 1000L)
}
\arguments{
\item{model}{An object of class \code{MarginalModel}, or a list of
\code{MarginalModel}'s.}

\item{niter}{The number of iterations for the reduced Gibb's sampler.}
}
\description{
Compute the marginal likelihood of a converged model.
}

