% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MarginalModel-class}
\alias{MarginalModel-class}
\title{The 'MarginalModel' class}
\description{
The 'MarginalModel' class
}
\section{Slots}{

\describe{
\item{\code{k}}{An integer value specifying the number of latent classes.}

\item{\code{hyperparams}}{An object of class `Hyperparameters` used to specify the hyperparameters of....}

\item{\code{theta}}{A}

\item{\code{sigma2}}{A}

\item{\code{nu.0}}{A}

\item{\code{sigma2.0}}{A}

\item{\code{pi}}{A}

\item{\code{mu}}{A}

\item{\code{tau2}}{A}

\item{\code{data}}{A}

\item{\code{data.mean}}{A}

\item{\code{data.prec}}{A}

\item{\code{z}}{A}

\item{\code{zfreq}}{A}

\item{\code{probz}}{A}

\item{\code{logprior}}{A}

\item{\code{loglik}}{A}

\item{\code{mcmc.chains}}{A}

\item{\code{batch}}{A}

\item{\code{batchElements}}{A}

\item{\code{hwe}}{A}

\item{\code{modes}}{A}

\item{\code{theta_order}}{A}

\item{\code{m.y}}{A}

\item{\code{mcmc.params}}{An object of class 'McmcParams'}
}}

