% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MarginalModelList-class}
\alias{MarginalModelList-class}
\title{An object to store a list of MarginalModel}
\description{
This class is primarily used internally for calculating the number of components of a mixture.
}
\section{Slots}{

\describe{
\item{\code{model_list}}{A list of MarginalModel}

\item{\code{names}}{The names of each model}

\item{\code{data}}{A vector containing the data}
}}

