% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MarginalModel.R,
%   R/methods-SingleBatchModel.R
\name{MarginalModelList}
\alias{MarginalModelList}
\alias{MarginalModelList}
\title{Constructor for list of single-batch models}
\usage{
MarginalModelList(data = numeric(), k = numeric(),
  mcmc.params = McmcParams(), ...)

MarginalModelList(data = numeric(), k = numeric(),
  mcmc.params = McmcParams(), ...)
}
\arguments{
\item{data}{numeric vector of average log R ratios}

\item{k}{numeric vector indicating the number of mixture components for each model}

\item{mcmc.params}{an object of class \code{McmcParams}}

\item{...}{additional arguments passed to \code{Hyperparameters}}

\item{data}{numeric vector of average log R ratios}

\item{k}{numeric vector indicating the number of mixture components for each model}

\item{mcmc.params}{an object of class \code{McmcParams}}

\item{...}{additional arguments passed to \code{Hyperparameters}}
}
\value{
a list. Each element of the list is a \code{BatchModel}

a list. Each element of the list is a \code{BatchModel}
}
\description{
An object of class MarginalModel is constructed for each k, creating a list of
MarginalModels.

An object of class MarginalModel is constructed for each k, creating a list of
MarginalModels.
}
\examples{
mlist <- MarginalModelList(data=y(MarginalModelExample), k=1:4)
mcmcParams(mlist) <- McmcParams(iter=1, burnin=1, nStarts=0)
mlist2 <- posteriorSimulation(mlist)
mlist <- MarginalModelList(data=y(MarginalModelExample), k=1:4)
mcmcParams(mlist) <- McmcParams(iter=1, burnin=1, nStarts=0)
mlist2 <- posteriorSimulation(mlist)
}
\seealso{
\code{\link{MarginalModel}} \code{\link{BatchModelList}}

\code{\link{MarginalModel}} \code{\link{BatchModelList}}
}
