% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R
\docType{methods}
\name{McmcChains}
\alias{McmcChains}
\alias{McmcChains,BatchModel-method}
\alias{McmcChains,MixtureModel-method}
\alias{McmcChains,missing-method}
\title{Initialize empty chain for model.}
\usage{
McmcChains(object)

\S4method{McmcChains}{missing}(object)

\S4method{McmcChains}{MixtureModel}(object)

\S4method{McmcChains}{BatchModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(McmcChains)}}
}
\description{
Initialize empty chain for model.
}

