% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{McmcParams-class}
\alias{McmcParams-class}
\alias{iter,burnin,nStarts,McmcParams-method}
\title{An object to specify MCMC options for a later simulation}
\description{
An object to specify MCMC options for a later simulation
}
\section{Slots}{

\describe{
\item{\code{thin}}{A one length numeric to specify thinning. A value of n indicates that every nth sample should be saved. Thinning helps to reduce autocorrelation.}

\item{\code{iter}}{A one length numeric to specify how many MCMC iterations should be sampled.}

\item{\code{burnin}}{A one length numeric to specify burnin. The first $n$ samples will be discarded.}

\item{\code{nstarts}}{A one length numeric to specify the number of chains in a simulation.}

\item{\code{param_updates}}{Indicates whether each parameter should be updated (1) or fixed (0).}

\item{\code{min_GR}}{minimum value of multivariate Gelman Rubin statistic for diagnosing convergence. Default is 1.2.}

\item{\code{min_effsize}}{the minimum mean effective size of the chains. Default is 1/3 * iter.}

\item{\code{max_burnin}}{The maximum number of burnin iterations before we give up and return the existing model.}

\item{\code{min_chains}}{minimum number of independence MCMC chains used for assessing convergence. Default is 3.}
}}

\examples{
McmcParams()
McmcParams(iter=1000)
mp <- McmcParams()
iter(mp)
}
