% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-McmcParams.R
\name{McmcParams}
\alias{McmcParams}
\title{Create an object of class 'McmcParams' to specify iterations, burnin, etc.}
\usage{
McmcParams(iter = 1000L, burnin = 0L, thin, nStarts = 1,
  param_updates = .param_updates())
}
\arguments{
\item{iter}{number of iterations}

\item{burnin}{number of burnin iterations}

\item{thin}{thinning interval}

\item{nStarts}{number of chains to run}

\item{param_updates}{labeled vector specifying whether each parameter is to be updated (1) or not (0).}
}
\value{
An object of class 'McmcParams'
}
\description{
Create an object of class 'McmcParams' to specify iterations, burnin, etc.
}
\examples{
mp <- McmcParams(iter=100, burnin=10)
}

