% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-McmcParams.R
\name{McmcParamsTrios}
\alias{McmcParamsTrios}
\title{Create an object of class 'McmcParamsTrios' to specify iterations, burnin, etc.}
\usage{
McmcParamsTrios(iter = 1000L, burnin = 0L, thin = 1L, nStarts = 1L,
  param_updates = .param_updates2(), min_GR = 1.2,
  min_effsize = round(1/3 * iter, 0), max_burnin = 32000,
  min_chains = 1)
}
\arguments{
\item{iter}{number of iterations}

\item{burnin}{number of burnin iterations}

\item{thin}{thinning interval}

\item{nStarts}{number of chains to run}

\item{param_updates}{labeled vector specifying whether each parameter is to be updated (1) or not (0).}
}
\value{
An object of class 'McmcParams'
}
\description{
Create an object of class 'McmcParamsTrios' to specify iterations, burnin, etc.
}
\examples{
     mp <- McmcParams(iter=100, burnin=10)
}
