% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{mcmcWithHomDel}
\alias{mcmcWithHomDel}
\title{Wrapper for fitting likely homozygous deletion polymorphisms}
\usage{
mcmcWithHomDel(mb, sb, restricted_model, model = "MB3")
}
\arguments{
\item{mb}{`MultiBatch` object with multiple batches (MB)}

\item{sb}{`MultiBatch` object with a single batch (SB)}

\item{restricted_model}{`MultiBatch` object that is restricted to a subset of the observations.  For example, often extreme observations in the tails are excluded.}

\item{model}{name of to evaluate.  See `modelName()`}
}
\description{
Wrapper for fitting likely homozygous deletion polymorphisms
}
\seealso{
[modelName()]
}
