% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{median_summary}
\alias{median_summary}
\title{Compute median summaries of log ratios at a CNV region}
\usage{
median_summary(se, provisional_batch, assay_index = 1, THR)
}
\arguments{
\item{se}{a SummarizedExperiment with assays containing SNP or bin-level summaries of copy number, typically log ratios.}

\item{provisional_batch}{A provisional definition of batch. Examples include PCR data, study center, DNA source, etc. See details.}

\item{assay_index}{If multiple assays are included in the SummarizedExperiment, the user should indicate which assay should be median summarized by a numeric index.}

\item{THR}{numeric value indicate that log ratios below this value are potentially samples with hemizygous or homozygous deletions.}
}
\description{
Compute median summaries of log ratios at a CNV region and define an indicator for likely deletions based on a user specified threshold for the median log ratio.  The identification of likely deletions can be helpful to flag possibly rare deletions.
}
\details{
See vignette for additional details.
}
\examples{
extdir <- system.file("extdata", package="CNPBayes")
se <- readRDS(file.path(extdir, "snp_se.rds"))
cnv_region <- GRanges("chr2", IRanges(90010895, 90248037),
                      seqinfo=seqinfo(se))
se2 <- subsetByOverlaps(se, cnv_region)
provisional_batch <- se2$Sample.Plate
full.data <- median_summary(se2,
                            provisional_batch=provisional_batch,
                            assay_index=2,
                            THR=-1)
}
