% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{mixture_layout}
\alias{mixture_layout}
\title{Layout of grid-based graphics for visualization of posterior predictive distributions and mapping of mixture components to copy number states.}
\usage{
mixture_layout(figure_list, augmented = TRUE, newpage = TRUE)
}
\arguments{
\item{figure_list}{list of grid graphics created by \code{mixture_plot}}

\item{augmented}{whether to display simulated datapoints that were used to fit rarer deletions. See details.}

\item{newpage}{whether to display the composite figure on a newpage or use existing graphics device}
}
\description{
Layout of grid-based graphics for visualization of posterior predictive distributions and mapping of mixture components to copy number states.
}
\details{
For studies of germline CNVs, extreme observations in the left-tail typically correspond to homozygous deletions and, when rare, may be present in a subset of the estimated batches.  The consequences of a rare deletion present in a subset of batches  are two-fold: (1) due to the hierarchical nature of the model, a mixture-component with a very large variance will be needed to accommodate the extreme observations and  (2) the mixture component indices may correspond to different copy number states between batches, complicating subsequent efforts to map mixture component indices to integer copy numbers.  Rather than exclude these observations, we augment the observed data with simulated homozygous deletions.  The simulated observations ensure the mixture component indices capture the same latent copy number in each batch. We rationalize this approach as being comparable to an empirically derived prior that large negative values at such germline CNV regions are not outliers of the hemizygous and diploid states but bona fide homozygous deletions.  Since our model does not assume a one-to-one mapping between mixture components and copy number nor that any of the alterations identified will be in Hardy Weinberg equilibrium (HWE), the assessment of HWE for germline CNVs can be a useful post-hoc quality control. While evidence against HWE does not necessarily indicate problems with the CNV calling, support for HWE would be unlikely if there were major sources of technical variation not yet accounted for.
}
\seealso{
\code{\link{ggMixture}} \code{\link{mixture_plot}}
}
