% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{mixture_plot}
\alias{mixture_plot}
\title{List of figures useful for evaluating model fit.}
\usage{
mixture_plot(model, snpdat, xlimit = c(-4, 1), bins = 100)
}
\arguments{
\item{model}{a MultiBatch instance for a CNV region}

\item{snpdat}{a SummarizedExperiment containing B allele frequencies at SNPs in the CNV region}

\item{xlimit}{x-axis limits for log ratio density and posterior predictive distribution}

\item{bins}{number of bins used by \code{geom_histogram} to display the primary data}
}
\value{
a list of figures
}
\description{
Creates list of figures used by CNPBayes to evaluate model fit and mapping of mixture component assignments to copy number states
}
\details{
The resulting list of figures can be plotted as a composite using the function \code{mixture_layout}.
}
\seealso{
\code{\link{mixture_layout}}
}
