% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ModelList-class}
\alias{ModelList-class}
\title{An object to store a list of MarginalModel or BatchModel}
\description{
This class is primarily used internally for calculating the number of components of a mixture. MarginalModelList and BatchModelList inherit from this class depending on the model.
}
\section{Slots}{

\describe{
\item{\code{model_list}}{A list of MarginalModel or BatchModel}

\item{\code{names}}{The names of each model}

\item{\code{data}}{A vector containing the data}
}}

