% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{modelName,MixtureModel-method}
\alias{modelName,MixtureModel-method}
\title{Abbreviated model name}
\usage{
\S4method{modelName}{MixtureModel}(object)
}
\arguments{
\item{object}{a SingleBatchModel, MultiBatchModel, etc.}
}
\description{
Abbreviated model name
}
\examples{
modelName(SingleBatchModelExample)
}
