% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{modelName}
\alias{modelName}
\alias{modelName,MultiBatch-method}
\alias{modelName,MultiBatchList-method}
\title{Accessor for name of MultiBatch model}
\usage{
modelName(object)
}
\arguments{
\item{object}{a `MultiBatch` object}
}
\description{
MultiBatch objects are named according to whether the mixture components are stratified by batch (MB) or single batch (SB), the variance across all mixture components within a batch is pooled (P), and the number of mixture components.  For example, MB3 is a multi-batch 3-component mixture model without pooled variance, while SBP4 is a 4-component single-batch mixture model using a pooled estimate of the variance.
}
