% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Hyperparameters.R
\docType{methods}
\name{ModelParamList}
\alias{ModelParamList}
\alias{ModelParamList,HyperParametersBatch-method}
\alias{ModelParamList,HyperParametersMarginal-method}
\alias{ModelParamList,HyperparametersBatch-method}
\alias{ModelParamList,HyperparametersMarginal-method}
\title{Creates a list of ModelParams for differing number of components}
\usage{
ModelParamList(hypp, K, data, mcmcp, batch)

\S4method{ModelParamList}{HyperparametersMarginal}(hypp, K, data, mcmcp, batch)

\S4method{ModelParamList}{HyperparametersBatch}(hypp, K, data, mcmcp, batch)
}
\arguments{
\item{hypp}{An object of class 'Hyperparameters'}

\item{K}{An integer vector containing the different component sizes for which you wish to create ModelParams.}

\item{data}{A vector containing the data.}

\item{mcmcp}{An object of class 'ModelParams'. Can have varying numbers of iterations, etc for each component size.}

\item{batch}{A vector indicating the batches from which each observation came from.}
}
\description{
Simplifies creating multiple objects for model parameters for multiple component sizes. Used when number of components is not known. Can be used with Marginal and Batch models.
}

