% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ModelParams-class}
\alias{ModelParams-class}
\title{An object to specify the parameters for MCMC simulation.}
\description{
This class has slots for parameters and data needed for simulation.
}
\section{Slots}{

\describe{
\item{\code{type}}{Indicates whether the model should be marginal over the batches, or hierarchical over the batches.}

\item{\code{k}}{Number of components.}

\item{\code{data}}{A numeric vector containing the data}

\item{\code{batch}}{A vector indicating from which batch each observation came.}

\item{\code{mcmc.params}}{An object of class McmcParams indicating number of iterations, etc.}
}}
\examples{
ModelParams()
ModelParams(k=3)
}

