% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-ModelParams.R
\name{ModelParams}
\alias{ModelParams}
\title{Create an object of class 'ModelParams' to specify data, parameters, etc.}
\usage{
ModelParams(type = c("marginal", "batch"), y = numeric(), k = 2L, batch,
  mcmc.params = McmcParams(iter = 1000, burnin = 0))
}
\arguments{
\item{type}{marginal or batch}

\item{y}{the data}

\item{k}{the number of components}

\item{batch}{a vector specifying batch labels}

\item{mcmc.params}{an object of class 'McmcParams'}
}
\description{
Create an object of class 'ModelParams' to specify data, parameters, etc.
}
\examples{
mp <- ModelParams("marginal", c(rnorm(10, 0),
                                     rnorm(10, 1),
                                     rnorm(10, 2)), k=3)
}

