% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-DensityModel.R, R/methods-MixtureModel.R
\docType{methods}
\name{modes}
\alias{modes}
\alias{modes,DensityModel-method}
\alias{modes,MixtureModel-method}
\alias{modes<-}
\alias{modes<-,MixtureModel-method}
\title{Retrieve the modes from a model.}
\usage{
modes(object)

modes(object) <- value

\S4method{modes}{DensityModel}(object)

\S4method{modes}{MixtureModel}(object)

\S4method{modes}{MixtureModel}(object) <- value
}
\arguments{
\item{object}{a \code{MixtureModel}-derived class}

\item{value}{a \code{list} of the modes.  See \code{mode(object)}
to obtain the correct format of the list.}
}
\value{
A list of the modes of each parameter
}
\description{
The iteration which maximizes log likelihood and log prior is found. The estimates for each parameter at this iteration are retrieved.

For a mixture model with K components, there are K! possible modes.
One can permute the ordering of the modes and assign the permuted
order to a MixtureModel derived class by this method.
}
\examples{
modes(MarginalModelExample)
}

