% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R
\docType{methods}
\name{muchd}
\alias{muchd}
\alias{muchd,McmcChains-method}
\alias{muchd,McmcChains-method}
\title{Calculates a frequency table of mu for offspring by observation.}
\usage{
muchd(object)

\S4method{muchd}{McmcChains}(object)

\S4method{muchd}{McmcChains}(object)
}
\arguments{
\item{object}{see \code{showMethods(muchd)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of mu for offspring by observation.
}
\examples{
     muchd(TrioBatchModelExample)
}
