% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R, R/MultiBatchList.R
\name{length,CnList-method}
\alias{length,CnList-method}
\alias{length,MultiBatchList-method}
\alias{names,MultiBatchList-method}
\title{MultiBatch accessors}
\usage{
\S4method{length}{CnList}(x)

\S4method{names}{MultiBatchList}(x)
}
\arguments{
\item{x}{object of class `CnList`}
}
\description{
Find length (number of samples), model names of MultiBatchList object etc.
}
