% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MultiBatch-class}
\alias{MultiBatch-class}
\title{An object for running MCMC simulations.}
\description{
BatchModel and MarginalModel both inherit from this class.
}
\section{Slots}{

\describe{
\item{\code{data}}{a tibble with one-dimensional summaries (oned), id, and batch}

\item{\code{parameters}}{list of parameters}

\item{\code{chains}}{object of class McmcChains}

\item{\code{current_values}}{current value of each chain}

\item{\code{summaries}}{list of empirical data and model summaries}

\item{\code{flags}}{list of model flags}
}}

