% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{MultiBatch}
\alias{MultiBatch}
\title{Constructor for MultiBatch objects
MultiBatch is the container used by CNPBayes for the organization of the primary data, current values of the mixture model parameters, hyperparameters, and MCMC chains.}
\usage{
MultiBatch(
  model = "MB3",
  data = modelData(),
  specs = model_spec(model, data),
  iter = 1000L,
  burnin = 200L,
  thin = 1L,
  nStarts = 4L,
  max_burnin = burnin,
  hp = Hyperparameters(k = specs$k),
  mp = McmcParams(iter = iter, thin = thin, burnin = burnin, nStarts = nStarts,
    max_burnin = max_burnin),
  parameters = modelParameters(mp = mp, hp = hp),
  chains = mcmc_chains(specs, parameters),
  current_values = modelValues2(specs, data, hp),
  summaries = modelSummaries(specs),
  flags = modelFlags()
)
}
\arguments{
\item{model}{character string abbreviation for the type of mixture model.  See details.}

\item{data}{one-dimensional summaries of log-ratios at a CNV-region for a collection of samples.  See details for required columns.}

\item{specs}{additional model specifications}

\item{iter}{total number of saved MCMC iterations (after thinning and does not include burnin)}

\item{burnin}{number of burnin MCMC simulations}

\item{thin}{number indicating how often MCMC updates are saved in the McmcChains slot of a MultiBatch instance}

\item{nStarts}{number of independent chains}

\item{max_burnin}{ignored}

\item{hp}{a Hyperparameters instance}

\item{mp}{a McmcParams instance}

\item{parameters}{Parameters of the finite Bayesian mixture model}

\item{chains}{a McmcChains instance}

\item{current_values}{values of mixture model parameters from the last iteration of the MCMC. These values can be used to initialize another chain if more MCMC simulations are needed.}

\item{summaries}{list of model summaries}

\item{flags}{list of flags that could indicate possible problems with convergence}
}
\value{
a MultiBatch instance
}
\description{
Constructor for MultiBatch objects
MultiBatch is the container used by CNPBayes for the organization of the primary data, current values of the mixture model parameters, hyperparameters, and MCMC chains.
}
\details{
CNPBayes fits finite mixture models with batch-specific means and variances, where multi-batch models are denoted by MB. Special cases of the MB model include a pooled variance model (MBP) with a single variance estimate per batch.  In addition, we evaluate models with a single batch (SB) and a single batch model with pooled variances (SBP). The abbreviation MB3  indicates a multi-batch model with 3 mixture components, while SBP2 corresponds to a single batch model with pooled variance with 2 mixture components.  This short-hand for the type of model and number of mixture components can be passed as the `model` argument to the MultiBatch constructor.  As described in the vignette, our general strategy that works well in most cases is to instantiate a default MultiBatch instance without to create a general container without regard to the number of mixture components and whether to pool at this stage.  Next, we fit some of the wrappers provided with CNPBayes that will explore several possible models. We refer the reader to the vignette for a detailed example of an example workflow.  While an effort has been made to automate these processes, exploratory data analyses and associated visualizations are extremely helpful.  Several such visualizations are provided in the vignette using the `ggMixture` function.
}
\examples{
 extdir <- system.file("extdata", package="CNPBayes")
 fname <- file.path(extdir, "CNP_001",
                    "batched_data.rds")
 batched.data <- readRDS(fname)
 mb <- MultiBatch(data=batched.data)
 mb
 ## Batch information is ignored if a SB model is created, but can be recovered using `revertToMultiBatch`.
 sb <- MultiBatch(model="SBP2", data=batched.data)
 mb2 <- revertToMultiBatch(sb)
}
\seealso{
\code{\link{ggMixture}} \code{\link{revertToMultiBatch}} \code{\link{median_summary}} \code{\link{kolmogorov_batches}}
}
