% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatchList.R
\name{MultiBatchList}
\alias{MultiBatchList}
\title{Constructor for a list of MultiBatch objects}
\usage{
MultiBatchList(
  models,
  data = modelData(),
  K,
  specs = modelSpecs(models, K, data),
  num_models = nrow(specs),
  burnin = 1000L,
  iter = 1000L,
  thin = 1L,
  nStarts = 4L,
  max_burnin = 32000L,
  hp = Hyperparameters(),
  mp = McmcParams(iter = iter, thin = thin, nStarts = nStarts, burnin = burnin,
    max_burnin = max_burnin),
  parameters = modelParameters(hp = hp, mp = mp),
  chains = listChains2(specs, parameters),
  current_values,
  summaries = listSummaries(specs),
  flags = listFlags(specs)
)
}
\arguments{
\item{models}{vector of abbreviated model names}

\item{data}{one-dimensional summaries of log-ratios at a CNV-region for a collection of samples.  See details for required columns.}

\item{K}{vector for numer of mixture components in the list object}

\item{specs}{additional model specifications}

\item{num_models}{ignored}

\item{burnin}{number of burnin MCMC simulations}

\item{iter}{total number of saved MCMC iterations (after thinning and does not include burnin)}

\item{thin}{number indicating how often MCMC updates are saved in the McmcChains slot of a MultiBatch instance}

\item{nStarts}{number of independent chains}

\item{max_burnin}{ignored}

\item{hp}{a Hyperparameters instance}

\item{mp}{a McmcParams instance}

\item{parameters}{Parameters of the finite Bayesian mixture model}

\item{chains}{a McmcChains instance}

\item{current_values}{values of mixture model parameters from the last iteration of the MCMC. These values can be used to initialize another chain if more MCMC simulations are needed.}

\item{summaries}{list of model summaries}

\item{flags}{list of flags that could indicate possible problems with convergence}
}
\value{
a MultiBatchList
}
\description{
MultiBatch is the container used by CNPBayes for the organization of the primary data, current values of the mixture model parameters, hyperparameters, and MCMC chains. A list of MultiBatch objects is often created internally to evaluate and compare different models.
}
\seealso{
\code{\link{MultiBatch}}
}
