% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MultiBatchModel.R
\name{MultiBatchModel2}
\alias{MultiBatchModel2}
\title{Constructor for MultiBatchModel}
\usage{
MultiBatchModel2(dat = numeric(), hp = HyperparametersMultiBatch(),
  mp = McmcParams(iter = 1000, thin = 10, burnin = 1000, nStarts = 4),
  batches = integer())
}
\arguments{
\item{data}{the data for the simulation.}

\item{k}{An integer value specifying the number of latent classes.}

\item{batch}{a vector of the different batch numbers (must be sorted)}

\item{hypp}{An object of class `Hyperparameters` used to specify the hyperparameters of the model.}

\item{mcmc.params}{An object of class 'McmcParams'}
}
\value{
An object of class `MultiBatchModel`
}
\description{
Initializes a MultiBatchModel, a container for storing data, parameters, and MCMC output for mixture models with batch- and component-specific means and variances.
}
\examples{
     model <- MultiBatchModel2(rnorm(10), batch=rep(1:2, each=5))
}
