% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MultiBatchModelExample}
\alias{MultiBatchModelExample}
\title{This data is an instance of MultiBatchModel}
\usage{
MultiBatchModelExample
}
\value{
An example of a `MultiBatchModel`
MultiBatchModelExample
}
\description{
This data is an instance of MultiBatchModel
}
\author{
Jacob Carey
}
