% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MultiBatchModel.R
\name{MultiBatchModelList}
\alias{MultiBatchModelList}
\title{Constructor for list of batch models}
\usage{
MultiBatchModelList(data = numeric(), k = numeric(), batch,
  mcmc.params = McmcParams(), ...)
}
\arguments{
\item{data}{numeric vector of average log R ratios}

\item{k}{numeric vector indicating the number of mixture components for each model}

\item{batch}{vector of batch labels}

\item{mcmc.params}{a \code{McmcParams} object}

\item{...}{additional arguments to \code{HyperparametersBatch}}
}
\value{
a list. Each element of the list is a \code{BatchModel}
}
\description{
An object of class MultiBatchModel is constructed for each k, creating a list of
BatchModels.
}
\examples{
mlist <- BatchModelList(data=y(MultiBatchModelExample), k=1:4, batch=batch(MultiBatchModelExample))
mcmcParams(mlist) <- McmcParams(iter=1, burnin=1, nStarts=0)
mlist2 <- posteriorSimulation(mlist)
}
\seealso{
\code{\link{BatchModel}}.  For single-batch data, use \code{\link{MarginalModelList}}.
}
