% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MultiBatchPooledExample}
\alias{MultiBatchPooledExample}
\title{This data is an instance of MultiBatchPooled}
\usage{
MultiBatchPooledExample
}
\value{
An example of a `MultiBatchPooled`
MultiBatchPooledExample
}
\description{
This data is an instance of MultiBatchPooled
}
\author{
Jacob Carey
}
