% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{muMean}
\alias{muMean}
\title{Retrieve overall mean averaged across MCMC simulations.}
\usage{
muMean(object)
}
\arguments{
\item{object}{an object of class MarginalModel or BatchModel}
}
\description{
Retrieve overall mean averaged across MCMC simulations.
}

