% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{mupar}
\alias{mupar}
\title{Calculates a frequency table of mu for parents by observation.}
\usage{
mupar(object)
}
\arguments{
\item{object}{see \code{showMethods(mupar)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of mu for parents by observation.
}
\examples{
     mupar(TrioBatchModelExample)
}
