% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-McmcChains.R, R/methods-ModelList.R
\docType{methods}
\name{names,McmcChains-method}
\alias{names,McmcChains-method}
\alias{names,ModelList-method}
\title{Retrieve the names of the parameters estimated in the MCMC chain.}
\usage{
\S4method{names}{McmcChains}(x)

\S4method{names}{ModelList}(x)
}
\arguments{
\item{x}{an object of class 'McmcChains'}
}
\description{
Retrieve the names of the parameters estimated in the MCMC chain.

Retrieve the names of Marginal or BatchModels contained.
}

