% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-McmcParams.R
\name{nStarts,McmcParams-method}
\alias{nStarts,McmcParams-method}
\alias{nStarts<-,McmcParams,ANY-method}
\alias{nStarts<-,McmcParams-method}
\title{Number of random starts for MCMC}
\usage{
\S4method{nStarts}{McmcParams}(object)

\S4method{nStarts}{McmcParams,ANY}(object) <- value
}
\description{
Number of random starts for MCMC
}
