% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{nStarts<-}
\alias{nStarts<-}
\title{Reset number of MCMC chains in simulation.}
\usage{
nStarts(object) <- value
}
\arguments{
\item{object}{see showMethods(nStarts)}

\item{value}{new number of chains}
}
\description{
This function changes the number of chains used for an MCMC simulation.
}

