% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{nStarts}
\alias{nStarts}
\title{Number of MCMC chains.}
\usage{
nStarts(object)
}
\arguments{
\item{object}{see showMethods(nStarts)}
}
\description{
This function retrieves the number of chains used for an MCMC simulation.
}

