% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-McmcChains.R
\name{nu.0,McmcChains-method}
\alias{nu.0,McmcChains-method}
\title{Accessor for nu.0 slot of McmcChains class}
\usage{
\S4method{nu.0}{McmcChains}(object)
}
\description{
Accessor for nu.0 slot of McmcChains class
}
