% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R
\docType{methods}
\name{nuchd}
\alias{nuchd}
\alias{nuchd,McmcChains-method}
\title{Calculates a frequency table of nu for offspring by observation.}
\usage{
nuchd(object)

\S4method{nuchd}{McmcChains}(object)
}
\arguments{
\item{object}{see \code{showMethods(nuchd)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of nu for offspring by observation.
}
\examples{
     nuchd(TrioBatchModelExample)
}
