% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{numberObs}
\alias{numberObs}
\alias{numberObs,MixtureModel-method}
\title{Number of observations in mixture model object}
\usage{
numberObs(model)
}
\arguments{
\item{model}{A MultiBatch instance}
}
\description{
Number of observations in mixture model object
}
