% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{nupar}
\alias{nupar}
\title{Calculates a frequency table of nu for parents by observation.}
\usage{
nupar(object)
}
\arguments{
\item{object}{see \code{showMethods(nupar)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of nu for parents by observation.
}
\examples{
     nupar(TrioBatchModelExample)
}
