% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{oned}
\alias{oned}
\alias{oned,MixtureModel-method}
\alias{oned,MultiBatch-method}
\alias{oned,MultiBatchList-method}
\title{Retrieve one-dimensional summaries of the log ratios for each subject.}
\usage{
oned(object)
}
\arguments{
\item{object}{see \code{showMethods(oned)}}
}
\value{
A vector the length of the data
}
\description{
Retrieve one-dimensional summaries of the log ratios for each subject.
}
