% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MarginalModel.R
\name{orderModels}
\alias{orderModels}
\title{Reorder models of varying component sizes.}
\usage{
orderModels(x, maxdev = 5)
}
\arguments{
\item{x}{the result of a call to \code{computeMarginalLik}.}

\item{maxdev}{the maximum of the sum of ranges of marginal likelihoods that is considered acceptable.}
}
\description{
Models are ordered according to marginal likelihood. The marginal likelihood is computed for each chain of each component size model separately. The mean is taken by model, and ordering by this mean marginal is performed. For each model, the difference of marginal likelihoods is calculated for each chain and the range is taken. If the sum of these ranges across models is greater than \code{maxdev}, a NULL is returned and a warning message printed.
}

