% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\docType{methods}
\name{p,MixtureModel-method}
\alias{p,MixtureModel-method}
\title{Retrieve mixture proportions.}
\usage{
\S4method{p}{MixtureModel}(object)
}
\arguments{
\item{object}{an object of class MarginalModel or BatchModel}
}
\value{
A vector of length the number of components
}
\description{
Retrieve mixture proportions.
}
\examples{
     p(SingleBatchModelExample)
}
