% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{p}
\alias{p}
\title{Retrieve mixture proportions.}
\usage{
p(object)
}
\arguments{
\item{object}{an object of class MarginalModel or BatchModel}
}
\description{
Retrieve mixture proportions.
}

