% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{paramUpdates<-}
\alias{paramUpdates<-}
\title{Change MCMC parameter status.}
\usage{
paramUpdates(x) <- value
}
\arguments{
\item{value}{new MCMC parameter status.}

\item{object}{see showMethods("paramUpdates<-")}
}
\description{
This function changes the status of MCMC parameters for a model.
}

