% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{paramUpdates}
\alias{paramUpdates}
\title{MCMC parameter status.}
\usage{
paramUpdates(x)
}
\arguments{
\item{object}{see showMethods(paramUpdates)}
}
\description{
This function retrieves the status of MCMC parameters for a model.
}

