% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{pic}
\alias{pic}
\title{Retrieve mixture proportions at each iteration of the MCMC.}
\usage{
pic(object)
}
\arguments{
\item{object}{an object of class MarginalModel or BatchModel}
}
\value{
A matrix of size MCMC iterations x Number of components
}
\description{
Retrieve mixture proportions at each iteration of the MCMC.
}
\examples{
     pic(MarginalModelExample)
}

