% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-BatchModel.R, R/methods-MixtureModel.R
\docType{methods}
\name{pMean}
\alias{pMean}
\alias{pMean,BatchModel-method}
\alias{pMean,MixtureModel-method}
\title{Calculate mean mixture proportion of components.}
\usage{
pMean(object)

\S4method{pMean}{BatchModel}(object)

\S4method{pMean}{MixtureModel}(object)
}
\arguments{
\item{object}{see showMethods(pMean)}
}
\description{
Calculate mean mixture proportion of components.
}

