% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{posteriorPredictive}
\alias{posteriorPredictive}
\title{Simulate data from the posterior predictive distribution}
\usage{
posteriorPredictive(model)
}
\arguments{
\item{model}{a MarginalModel or MultiBatchModel}
}
\description{
Simulating from the posterior predictive distribution can be helpful for assessing the adequacy of the mixture model.
}
\examples{

 model <- MarginalModelExample
 mp <- McmcParams(iter=200, burnin=50)
 mcmcParams(model) <- mp
 model <- posteriorSimulation(model)
 pd <- posteriorPredictive(model)
 if(FALSE) qqplot(pd, y(model))

\dontrun{
    bmodel <- BatchModelExample
    mp <- McmcParams(iter=500, burnin=150, nStarts=20)
    mcmcParams(bmodel) <- mp
    bmodel <- posteriorSimulation(bmodel)
    batchy <- posteriorPredictive(bmodel)
}

}
