% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{posteriorSimulation}
\alias{posteriorSimulation}
\alias{posteriorSimulation,MixtureModel-method}
\title{Run the MCMC simulation.}
\usage{
posteriorSimulation(object)

\S4method{posteriorSimulation}{MixtureModel}(object)
}
\arguments{
\item{object}{see showMethods(posteriorSimulation)}
}
\description{
nStarts chains are run. b burnin iterations are run and then discarded. Next, s iterations are run in each train. The mode of the MCMC simulation is also calculated.
}

