% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-TrioBatchModel.R
\name{pp}
\alias{pp}
\title{Retrieve mixture proportions of parents.}
\usage{
pp(object)
}
\arguments{
\item{object}{an object of class TrioBatchModel}
}
\value{
A vector of length the number of components
}
\description{
Retrieve mixture proportions of parents.
}
\examples{
  \dontrun{
     pp(TrioBatchModelExample)
   } 
}
