% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-functions.R
\name{predictiveDataTable}
\alias{predictiveDataTable}
\title{Create a tibble of posterior predictive distributions for a list of models}
\usage{
predictiveDataTable(models)
}
\arguments{
\item{models}{list of models}
}
\description{
Create a tibble of posterior predictive distributions for a list of models
}
\examples{
  sb <- SingleBatchModelExample
  mcmcParams(sb) <- McmcParams(iter=500, burnin=50)
  sb <- posteriorSimulation(sb)
  models <- list(MultiBatchModelExample, sb)
  tab <- predictiveDataTable(models)
  \dontrun{
    library(ggplot2)
    ggplot(tab, aes(y, fill=predictive)) +
      geom_density(alpha=0.4, adjust=1/2) +
      facet_grid(model~batch) +
      guides(fill=guide_legend(title="")) +
      theme(panel.background=element_rect(fill="white"))
  }
}
