% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber-models.R
\name{probCopyNumber}
\alias{probCopyNumber}
\title{Posterior probabilities for copy number states}
\usage{
probCopyNumber(model)
}
\arguments{
\item{model}{a \code{SingleBatchCopyNumber} or \code{MultiBatchCopyNumber} instance}
}
\description{
In contrast to posterior probabilities for mixture components, this function
returns posterior probabilities for distinct copy number states.
}

