% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{probz}
\alias{probz}
\alias{probz,MixtureModel-method}
\alias{probz,MultiBatch-method}
\title{Retrieve the probability of latent variable membership by observation.}
\usage{
probz(object)
}
\arguments{
\item{object}{see \code{showMethods(probz)}}
}
\value{
A matrix of size number of observations x number of components
}
\description{
Retrieve the probability of latent variable membership by observation.
}
\examples{
     head(probz(SingleBatchModelExample))
}
