% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R,
%   R/methods-TrioBatchModel.R
\docType{methods}
\name{probzchd}
\alias{probzchd}
\alias{probzchd,McmcChains-method}
\alias{probzchd,TrioBatchModel-method}
\title{Retrieve the probability of latent variable membership by observation for offspring.}
\usage{
probzchd(object)

\S4method{probzchd}{McmcChains}(object)

\S4method{probzchd}{TrioBatchModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(probzchd)}}
}
\value{
A matrix of size number of observations x number of components
}
\description{
Retrieve the probability of latent variable membership by observation for offspring.
}
\examples{
     probzchd(TrioBatchModelExample)
}
