% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-TrioBatchModel.R
\docType{methods}
\name{probzpar}
\alias{probzpar}
\alias{probzpar,TrioBatchModel-method}
\title{Retrieve the probability of latent variable membership by observation for parents.}
\usage{
probzpar(object)

\S4method{probzpar}{TrioBatchModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(probzpar)}}
}
\value{
A matrix of size number of observations x number of components
}
\description{
Retrieve the probability of latent variable membership by observation for parents.
}
\examples{
\dontrun{
     probzpar(TrioBatchModelExample)
}
}
