% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MarginalModel.R
\docType{methods}
\name{relabel}
\alias{relabel}
\alias{relabel,BatchModel-method}
\alias{relabel,MarginalModel-method}
\title{Switch labels of latent variables}
\usage{
relabel(object, zindex)

\S4method{relabel}{MarginalModel}(object, zindex)

\S4method{relabel}{BatchModel}(object, zindex)
}
\arguments{
\item{object}{see \code{showMethods(relabel)}}

\item{zindex}{an integer vector containing the labels in the desired order}
}
\description{
Switch labels of latent variables
}

