% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{sample2}
\alias{sample2}
\title{Downsampling will not work well if the range of the downsampled data is not similar to the range of the original data}
\usage{
sample2(mb, N)
}
\description{
Downsampling will not work well if the range of the downsampled data is not similar to the range of the original data
}
