% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SummarizedExperiment.R
\name{saveBatch}
\alias{saveBatch}
\title{Save se data}
\usage{
saveBatch(se, batch.file, THR = 0.1)
}
\arguments{
\item{se}{a SummarizedExperiment object}

\item{batch.file}{the file name to which to save the data}

\item{THR}{threshold below which the null hypothesis should be rejected and batches are collapsed.}
}
\value{
A vector of collapsed batch labels
}
\description{
Batches drawn from the same distribution as identified by Kolmogorov-Smirnov test are combined.
}

