% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SingleBatchModel.R
\name{SB}
\alias{SB}
\title{Constructors for SB and SBP models}
\usage{
SB(dat = numeric(), hp = Hyperparameters(), mp = McmcParams(iter =
  1000, burnin = 1000, thin = 10, nStarts = 4))
}
\arguments{
\item{dat}{numeric vector of average log R ratios}

\item{hp}{an object of class \code{Hyperparameters}}

\item{mp}{an object of class \code{McmcParams}}
}
\value{
An instance of \code{MultiBatchModel}
}
\description{
Create objects of class SingleBatchModel or SingleBatchPooled. This function creates a MultiBatchModel with a single batch.
}
\examples{
SB()
}
\seealso{
\code{\link{MultiBatchModel2}}
}
