% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{sigma}
\alias{sigma}
\title{Retrieve standard deviations of each component/batch mean.}
\usage{
sigma(object)
}
\arguments{
\item{object}{an object of class MarginalModel or BatchModel}
}
\description{
Retrieve standard deviations of each component/batch mean.
}

