% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R,
%   R/methods-MultiBatchPooled.R
\docType{methods}
\name{sigma2}
\alias{sigma2}
\alias{sigma2,McmcChains-method}
\alias{sigma2,missing-method}
\alias{sigma2,MultiBatchPooled-method}
\alias{sigma2<-,MultiBatchPooled,ANY-method}
\alias{sigma2<-,MultiBatchCopyNumberPooled-method}
\alias{sigma2,MultiBatchCopyNumberPooled-method}
\alias{sigma,MultiBatchCopyNumberPooled-method}
\alias{sigma2<-,MultiBatchCopyNumberPooled,ANY-method}
\alias{sigma<-,MultiBatchCopyNumberPooled-method}
\alias{sigma_,MultiBatchCopyNumberPooled-method}
\title{Retrieve the variances of each component and batch distribution}
\usage{
sigma2(object)

\S4method{sigma2}{McmcChains}(object)

\S4method{sigma2}{MultiBatchPooled}(object)

\S4method{sigma2}{MultiBatchPooled,ANY}(object) <- value

\S4method{sigma2}{MultiBatchCopyNumberPooled}(object)

\S4method{sigma}{MultiBatchCopyNumberPooled}(object)

\S4method{sigma2}{MultiBatchCopyNumberPooled,ANY}(object) <- value

\S4method{sigma}{MultiBatchCopyNumberPooled}(object) <- value

\S4method{sigma_}{MultiBatchCopyNumberPooled}(object)
}
\arguments{
\item{object}{see \code{showMethods(sigma2)}}
}
\value{
A vector of length number of components or a matrix of size 
number of batches x number of components
}
\description{
For a MarginalModel, this function returns a vector of variances. For a BatchModel, returns a matrix of size number of batches by number of components.
}
\examples{
     sigma2(SingleBatchModelExample)
}
