% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{sigma2.0}
\alias{sigma2.0}
\alias{sigma2.0,McmcChains-method}
\title{Accessor for sigma2.0 slot of McmcChains class}
\usage{
sigma2.0(object)
}
\arguments{
\item{object}{a Hyperparameters instance}
}
\description{
Accessor for sigma2.0 slot of McmcChains class
}
