% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R
\docType{methods}
\name{sigma2.0chd}
\alias{sigma2.0chd}
\alias{sigma2.0chd,McmcChains-method}
\title{Calculates a frequency table of sigma2.0 for offspring by observation.}
\usage{
sigma2.0chd(object)

\S4method{sigma2.0chd}{McmcChains}(object)
}
\arguments{
\item{object}{see \code{showMethods(sigma2.0chd)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of sigma2.0 for offspring by observation.
}
\examples{
     sigma2.0chd(TrioBatchModelExample)
}
