% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R
\docType{methods}
\name{sigma2chd}
\alias{sigma2chd}
\alias{sigma2chd,McmcChains-method}
\title{Calculates a frequency table of sigma2 for offspring by observation.}
\usage{
sigma2chd(object)

\S4method{sigma2chd}{McmcChains}(object)
}
\arguments{
\item{object}{see \code{showMethods(sigma2chd)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of sigma2 for offspring by observation.
}
\examples{
     sigma2chd(TrioBatchModelExample)
}
