% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{sigma2par}
\alias{sigma2par}
\title{Calculates a frequency table of sigma2 for parents by observation.}
\usage{
sigma2par(object)
}
\arguments{
\item{object}{see \code{showMethods(sigma2par)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of sigma2 for parents by observation.
}
\examples{
     sigma2par(TrioBatchModelExample)
}
