% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{sigmac}
\alias{sigmac}
\title{Retrieve standard deviation of each component/batch mean at each iteration of the MCMC.}
\usage{
sigmac(object)
}
\arguments{
\item{object}{an object of class MarginalModel or BatchModel}
}
\description{
Retrieve standard deviation of each component/batch mean at each iteration of the MCMC.
}
\examples{
sigmac(MarginalModelExample)
}

