% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-BatchModel.R, R/methods-MixtureModel.R
\docType{methods}
\name{sigmaMean}
\alias{sigmaMean}
\alias{sigmaMean,BatchModel-method}
\alias{sigmaMean,MixtureModel-method}
\title{Calculate mean variance of each component/batch.}
\usage{
sigmaMean(object)

\S4method{sigmaMean}{BatchModel}(object)

\S4method{sigmaMean}{MixtureModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(sigmaMean)}}
}
\description{
Mean is taken across simulated MCMC chain.
}

