% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulateData}
\alias{simulateData}
\title{Create simulated data for testing.}
\usage{
simulateData(N, p, theta, sds, df = 10)
}
\arguments{
\item{N}{number of observations}

\item{p}{a vector indicating probability of membership to each component}

\item{theta}{a vector of means, one per component}

\item{sds}{a vector of standard deviations, one per component}

\item{df}{length-1 numeric vector for the t-distribution degrees of freedom}
}
\value{
An object of class 'SingleBatchModel'
}
\description{
Create simulated data for testing.
}
