% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber-models.R
\name{SingleBatchCopyNumber}
\alias{SingleBatchCopyNumber}
\title{Constructor for copy number models (single-batch)}
\usage{
SingleBatchCopyNumber(model)
}
\arguments{
\item{model}{a \code{MarginalModel}}
}
\description{
Used for mapping mixture components to distict copy number states
}
\examples{
SingleBatchCopyNumber(MarginalModelExample)
}

