% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatchList.R
\docType{methods}
\name{singleBatchGuided,MultiBatchList,MultiBatch-method}
\alias{singleBatchGuided,MultiBatchList,MultiBatch-method}
\title{Use values from SingleBatch model to simulate reasonable starting values for the SingleBatch-pooled and MultiBatch models of the same number of components}
\usage{
\S4method{singleBatchGuided}{MultiBatchList,MultiBatch}(x, guide)
}
\description{
Use values from SingleBatch model to simulate reasonable starting values for the SingleBatch-pooled and MultiBatch models of the same number of components
}
