% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SingleBatchModel.R
\name{SingleBatchModel}
\alias{SingleBatchModel}
\title{Create an object for running single batch MCMC simulations.}
\usage{
SingleBatchModel(dat = numeric(), hp = Hyperparameters(),
  mp = McmcParams(iter = 1000, burnin = 1000, thin = 10, nStarts = 4))
}
\arguments{
\item{data}{the data for the simulation.}

\item{k}{An integer value specifying the number of latent classes.}

\item{hypp}{An object of class `Hyperparameters` used to specify the hyperparameters of the model.}

\item{mcmc.params}{An object of class 'McmcParams'}
}
\value{
An object of class 'SingleBatchModel'
}
\description{
Create an object for running single batch MCMC simulations.
}
\examples{
     model <- SingleBatchModel(data=rnorm(10), k=1)
}
