% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SingleBatchModel.R,
%   R/methods-SingleBatchPooled.R
\name{SingleBatchModel2}
\alias{SingleBatchModel2}
\alias{SingleBatchPooled}
\alias{SBP}
\title{Constructors for SB and SBP models}
\usage{
SingleBatchModel2(dat = numeric(), hp = Hyperparameters(),
  mp = McmcParams(iter = 1000, burnin = 1000, thin = 10, nStarts = 4))

SingleBatchPooled(dat = numeric(), hp = Hyperparameters(),
  mp = McmcParams(iter = 1000, burnin = 1000, thin = 10, nStarts = 4))

SBP(dat = numeric(), hp = Hyperparameters(), mp = McmcParams(iter =
  1000, burnin = 1000, thin = 10, nStarts = 4))
}
\arguments{
\item{dat}{numeric vector of average log R ratios}

\item{hp}{an object of class \code{Hyperparameters}}

\item{mp}{an object of class \code{McmcParams}}

\item{dat}{numeric vector}

\item{hp}{object of class Hyperparameters created by \code{Hyperparameters} function}

\item{mp}{a \code{McmcParams} object}

\item{dat}{numeric vector}

\item{hp}{object of class Hyperparameters created by \code{Hyperparameters} function}

\item{mp}{a \code{McmcParams} object}
}
\value{
An instance of \code{MultiBatchModel}
}
\description{
This creates a MultiBatchModel object with a single batch.
}
\examples{
SingleBatchModel2()
SingleBatchModel2(dat=rnorm(100), hpList(k=2)[["SB"]])
SingleBatchPooled()
SingleBatchPooled()
## or, equivalently
SBP()
SingleBatchPooled()
## or, equivalently
SBP()
}
\seealso{
\code{\link{MultiBatchModel2}}
}
