% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SingleBatchModelExample}
\alias{SingleBatchModelExample}
\title{This data is an instance of SingleBatchModel}
\usage{
SingleBatchModelExample
}
\value{
An example of a `SingleBatchModel`
SingleBatchModelExample
}
\description{
This data is an instance of SingleBatchModel
}
\author{
Jacob Carey
}
