% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_initialization.R
\docType{methods}
\name{startingValues,MarginalModel-method}
\alias{startingValues,MarginalModel-method}
\title{##}
\usage{
\S4method{startingValues}{MarginalModel}(object)
}
\description{
##
}

