% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{summarize_region}
\alias{summarize_region}
\title{Wrapper for summarizing CNV regions}
\usage{
summarize_region(
  se,
  provisional_batch,
  THR = -1,
  assay_index = 1,
  KS_cutoff = 0.001,
  S = 1000,
  min_size = 250
)
}
\arguments{
\item{se}{A SummarizedExperiment containing log ratios at SNPs or genomic bins for a collection of samples at a single CNV region}

\item{provisional_batch}{A provisional batch label such as date of PCR, study center, or DNA source.}

\item{THR}{log ratios below this value are potentially hemizygous or homozygous deletions}

\item{assay_index}{index of the assay element in the SummarizedExperiment object that contains the log ratios that are to be summarized.}

\item{KS_cutoff}{Cutoff for Kolmogorov-Smirnov (KS) p-value.  For two batches that have a KS p-value above this threshold, the batches are combined to form a single batch.}

\item{S}{desired number of samples to include in the down-sampled dataset}

\item{min_size}{integer indicating the number of samples to randomly select for each batch. The actual number of samples included may be larger as samples flagged as likely deleted are not down-sampled.}
}
\description{
Computes a median summary of log ratios for each sample at CNV regions, identifies batches from the provisional batch labels, and down samples the resulting summarized data for subsequent evaluation by finite mixture models.
}
\details{
Helpful to provide a provisional batch label that is fairly granular, allowing `kolmogorov_batches` to provide a further coarsening of the batch labels.
}
\seealso{
\code{\link{kolmogorov_batches}} \code{\link{median_summary}} \code{\link{down_sample2}} \code{\link{MultiBatch}}
}
