% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{tau}
\alias{tau}
\title{Retrieve overall standard deviation.}
\usage{
tau(object)
}
\arguments{
\item{object}{an object of class MarginalModel or BatchModel}
}
\value{
A vector of standard deviations
}
\description{
Retrieve overall standard deviation.
}
\examples{
     tau(MarginalModelExample)
}

