% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-McmcChains.R
\name{tau2,McmcChains-method}
\alias{tau2,McmcChains-method}
\title{Accessor for tau2 slot of McmcChains class}
\usage{
\S4method{tau2}{McmcChains}(object)
}
\description{
Accessor for tau2 slot of McmcChains class
}
