% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R
\docType{methods}
\name{tau2chd}
\alias{tau2chd}
\alias{tau2chd,McmcChains-method}
\title{Calculates a frequency table of tau2 for offspring by observation.}
\usage{
tau2chd(object)

\S4method{tau2chd}{McmcChains}(object)
}
\arguments{
\item{object}{see \code{showMethods(tau2chd)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of tau2 for offspring by observation.
}
\examples{
     tau2chd(TrioBatchModelExample)
}
