% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{tau2par}
\alias{tau2par}
\title{Calculates a frequency table of tau2 for parents by observation.}
\usage{
tau2par(object)
}
\arguments{
\item{object}{see \code{showMethods(tau2par)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of tau2 for parents by observation.
}
\examples{
     tau2par(TrioBatchModelExample)
}
