% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{tauc}
\alias{tauc}
\title{Retrieve overall standard deviation at each iteration of the MCMC.}
\usage{
tauc(object)
}
\arguments{
\item{object}{an object of class MarginalModel or BatchModel}
}
\description{
Retrieve overall standard deviation at each iteration of the MCMC.
}
\examples{
tauc(MarginalModelExample)
}

