% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{tauMean}
\alias{tauMean}
\title{Retrieve overall standard deviation averaged across MCMC simulations.}
\usage{
tauMean(object)
}
\arguments{
\item{object}{an object of class MarginalModel or BatchModel}
}
\description{
Retrieve overall standard deviation averaged across MCMC simulations.
}
\examples{
tauMean(MarginalModelExample)
}

