% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{theta}
\alias{theta}
\title{Retrieve theta.}
\usage{
theta(object)
}
\arguments{
\item{object}{see \code{showMethods(theta)}}
}
\description{
This function retrieves theta from an object. Theta is a vector or matrix representing the means of each component distribution.
}

