% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{thetac}
\alias{thetac}
\alias{thetac,MixtureModel-method}
\title{Retrieve simulated thetas from a MixtureModel}
\usage{
thetac(object)

\S4method{thetac}{MixtureModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(thetac)}}
}
\description{
Retrieve simulated thetas from a MixtureModel
}

