% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R
\docType{methods}
\name{thetachd}
\alias{thetachd}
\alias{thetachd,McmcChains-method}
\title{Calculates a frequency table of theta for offspring by observation.}
\usage{
thetachd(object)

\S4method{thetachd}{McmcChains}(object)
}
\arguments{
\item{object}{see \code{showMethods(thetachd)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of theta for offspring by observation.
}
\examples{
     thetachd(TrioBatchModelExample)
}
