% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{thetapar}
\alias{thetapar}
\title{Calculates a frequency table of thetas for parents by observation.}
\usage{
thetapar(object)
}
\arguments{
\item{object}{see \code{showMethods(thetapar)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of thetas for parents by observation.
}
\examples{
     thetapar(TrioBatchModelExample)
}
