% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcParams.R, R/methods-MixtureModel.R
\docType{methods}
\name{thin}
\alias{thin}
\alias{thin,McmcParams-method}
\alias{thin,MixtureModel-method}
\alias{thin<-}
\title{Number of thinning intervals.}
\usage{
thin(object)

thin(object) <- value

\S4method{thin}{McmcParams}(object)

\S4method{thin}{MixtureModel}(object)
}
\arguments{
\item{object}{see showMethods(thin)}

\item{object}{see showMethods("thin<-")}
}
\value{
An integer of the number of thinning intervals

An object of the same class
}
\description{
This function retrieves the number of thinning intervals used for an MCMC simulation.

This function updates the number of thinning intervals used for an MCMC simulation.
}
\examples{
     thin(MarginalModelExample)
     thin(MarginalModelExample) <- 10L
}

