% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{thin}
\alias{thin}
\title{Number of thinning intervals.}
\usage{
thin(object)
}
\arguments{
\item{object}{see showMethods(thin)}
}
\description{
This function retrieves the number of thinning intervals used for an MCMC simulation.
}

