% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tileMedians}
\alias{tileMedians}
\title{Create tile labels for each observation}
\usage{
tileMedians(y, nt, batch)
}
\arguments{
\item{y}{vector containing data}

\item{nt}{the number of observations per batch}

\item{batch}{a vector containing the labels from which batch each observation came from.}
}
\value{
A tibble with a tile assigned to each log R ratio
}
\description{
For large datasets (several thousand subjects), the computational burden for fitting Bayesian mixture models can be high.  Downsampling can reduce the computational burden with little effect on inference.  The function tileMedians is useful for putting the median log R ratios for each subject in a bucket. The observations in each bucket are averaged.  This is done independently for each batch and the range of median log R ratios within each bucket is guaranteed to be less than 0.05.  Note this function requires specification of a batch variable. If the study was small enough such that all the samples were processed in a single batch, then downsampling would not be needed.  By summarizing the observations in each bucket by batch, the SingleBatchModels (SB or SBP) and MultiBatchModels (MB or MBP) will be fit to the same data and are still comparable by marginal likelihoods or Bayes Factors.
}
\examples{
  mb <- MultiBatchModelExample
  tiled.medians <- tileMedians(y(mb), 200, batch(mb))
  expected <- tiled.medians \%>\% group_by(tile) \%>\%
    summarize(avgLRR=mean(logratio),
              batch=unique(batch))
  tile.summaries <- tileSummaries(tiled.medians)
  expect_identical(tile.summaries, expected)
  mp <- McmcParams(iter=50, burnin=100)
  mb <- MultiBatchModel2(dat=tile.summaries$avgLRR,
                         batches=tile.summaries$batch, mp=mp)
  mb <- posteriorSimulation(mb)
  ggMixture(mb)
  mb2 <- upSample(mb, tiled.medians)
  ggMixture(mb2)
}
\seealso{
\code{\link[dplyr]{ntile}}
}
