% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/visualization.R
\docType{methods}
\name{tracePlot}
\alias{tracePlot}
\alias{tracePlot,BatchModel-method}
\title{Create a trace plot of a parameter estimated by MCMC.}
\usage{
tracePlot(object, name, ...)

\S4method{tracePlot}{BatchModel}(object, name, ...)
}
\arguments{
\item{object}{see \code{showMethods(tracePlot)}}

\item{name}{the name of the parameter for which to plot values. Can be 'theta', 'sigma', 'p', 'mu', or 'tau'.}

\item{...}{Other argument to pass to plot.}
}
\value{
A traceplot of a parameter value
}
\description{
Create a trace plot of a parameter estimated by MCMC.
}
\examples{
tracePlot(BatchModelExample, "theta")
tracePlot(BatchModelExample, "sigma")
}

