% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-TrioBatchModel.R
\name{TrioBatchModel}
\alias{TrioBatchModel}
\title{Constructor for TrioBatchModel}
\usage{
TrioBatchModel(triodata = tibble(), hp = HyperparametersTrios(),
  mp = McmcParams(iter = 1000, thin = 10, burnin = 1000, nStarts = 4),
  mprob = mprob)
}
\arguments{
\item{triodata}{the data for the simulation.}

\item{hp}{An object of class `Hyperparameters` used to specify the hyperparameters of the model.}

\item{mp}{An object of class 'McmcParams'}

\item{batches}{an integer-vector of the different batches}
}
\value{
An object of class `TrioBatchModel`
}
\description{
Initializes a TrioBatchModel, a container for storing data, parameters, and MCMC output for mixture models with batch- and component-specific means and variances.
}
