% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{triodata_lrr}
\alias{triodata_lrr}
\title{Retrieves intensity data from trios}
\usage{
triodata_lrr(object)
}
\arguments{
\item{object}{see \code{showMethods(triodata_lrr)}}
}
\value{
An integer vector of length the number of components
}
\description{
Retrieves intensity data from trios
}
\examples{
\dontrun{
     triodata_lrr(TrioBatchModelExample)
}
}
