% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{upsample2}
\alias{upsample2}
\title{Provides probabilistic assignments for mixture components to the entire study after down-sampling}
\usage{
upsample2(model, dat)
}
\arguments{
\item{model}{a MultiBatch instance}

\item{dat}{a \code{tibble} of the full dataset typically created by \code{median_summary}}
}
\value{
a \code{tibble} for the full dataset with mixture component assignments
}
\description{
For large studies, tens of thousands of observations are not needed to approximate the mixture component densities and down-sampling is used to speed up computation.  The inverse, up-sampling, is used after fitting mixture models to provide mixture component probabilities for all participants in the study.
}
\details{
See the vignette or README for an example workflow.
}
\seealso{
\code{\link{down_sample2}}
}
