% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{warmup}
\alias{warmup}
\title{Run short burnin from multiple independent starts and select a single model for further MCMC}
\usage{
warmup(
  tib,
  model1,
  model2 = NULL,
  model2.penalty = 50,
  Nrep = 10,
  .burnin = 100
)
}
\arguments{
\item{tib}{a tibble containing data, sample identifiers, and batch information}

\item{model1}{Name of model (e.g., "MB3", "MBP3", "SB3", ...)}

\item{model2}{Second model to evaluate (optional)}

\item{model2.penalty}{Typically we compare a simple model (SB3) to a more complicated model (MB3).  This allows the user to specify a penalty for the more complicated model.}

\item{Nrep}{Integer specifying number of independent starts}

\item{.burnin}{Integer specifying number of burnin simulations}
}
\description{
Run short burnin from multiple independent starts and select a single model for further MCMC
}
