% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R,
%   R/deprecated_functions.R
\docType{methods}
\name{y}
\alias{y}
\alias{y,MixtureModel-method}
\alias{y,DensityModel-method}
\title{Retrieve data.}
\usage{
y(object)

\S4method{y}{MixtureModel}(object)

\S4method{y}{DensityModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(y)}}
}
\value{
A vector containing the data
}
\description{
Retrieve data.
}
\examples{
     y(MarginalModelExample)
}
