% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{z}
\alias{z}
\title{Retrieve latent variable assignments.}
\usage{
z(object)
}
\arguments{
\item{object}{see \code{showMethods(z)}}
}
\value{
A vector the length of the data
}
\description{
Retrieves the simulated latent variable assignments of each observation at each MCMC simulation.
}
\examples{
     head(z(SingleBatchModelExample))
}
