% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-McmcChains.R
\name{zFreq,McmcChains-method}
\alias{zFreq,McmcChains-method}
\alias{zfreq,McmcChains-method}
\title{Accessor for frequency of mixture component label in McmcChains class}
\usage{
\S4method{zFreq}{McmcChains}(object)
}
\description{
Accessor for frequency of mixture component label in McmcChains class
}
