% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R
\docType{methods}
\name{zFreqChd}
\alias{zFreqChd}
\alias{zFreqChd,McmcChains-method}
\alias{zfreqchd,McmcChains-method}
\title{Calculates a frequency table of latent variable assigments for offspring by observation.}
\usage{
zFreqChd(object)

\S4method{zFreqChd}{McmcChains}(object)
}
\arguments{
\item{object}{see \code{showMethods(zfreqchd)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of latent variable assigments for offspring by observation.
}
\examples{
     zfreqchd(TrioBatchModelExample)
}
