% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R
\docType{methods}
\name{zFreqPar}
\alias{zFreqPar}
\alias{zFreqPar,McmcChains-method}
\alias{zfreqpar,McmcChains-method}
\title{Calculates a frequency table of latent variable assigments for parents by observation.}
\usage{
zFreqPar(object)

\S4method{zFreqPar}{McmcChains}(object)
}
\arguments{
\item{object}{see \code{showMethods(zfreqpar)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of latent variable assigments for parents by observation.
}
\examples{
\dontrun{
     zfreqpar(TrioBatchModelExample)
}
}
