% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{batch}
\alias{batch}
\alias{batch,MixtureModel-method}
\alias{batch,MultiBatch-method}
\alias{batch,MultiBatchList-method}
\title{Retrieve batches from object.}
\usage{
batch(object)
}
\arguments{
\item{object}{see \code{showMethods(batch)}}
}
\value{
The batch of each data element.
}
\description{
The batches are represented as a vector of integers.
}
\examples{
     batch(MultiBatchModelExample)
}
