% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{batch}
\alias{batch}
\title{Extract integer vector of batch labels from MixtureModel-derived class}
\usage{
batch(object)
}
\value{
An vector of length of the data
}
\description{
Extract integer vector of batch labels from MixtureModel-derived class
}
\examples{
model <- BatchModel()
batch(model)
}

