% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BatchModelExample}
\alias{BatchModelExample}
\title{This data is a simulated example of Batch data}
\format{\preformatted{Formal class 'BatchModel' [package "CNPBayes"] with 24 slots
  ..@ k                   : int 3
  ..@ hyperparams         :Formal class 'HyperparametersBatch' [package "CNPBayes"] with 9 slots
  .. .. ..@ k     : int 3
  .. .. ..@ mu.0  : num 0
  .. .. ..@ tau2.0: num 100
  .. .. ..@ eta.0 : num 1800
  .. .. ..@ m2.0  : num 0.0167
  .. .. ..@ alpha : num [1:3] 1 1 1
  .. .. ..@ beta  : num 0.1
  .. .. ..@ a     : num 1.8
  .. .. ..@ b     : num 6
  ..@ theta               : num [1:3, 1:3] -0.2169 0.0107 0.1919 0.8081 0.9827 ...
  ..@ sigma2              : num [1:3, 1:3] 0.01349 0.01027 0.00887 0.01103 0.01133 ...
  ..@ nu.0                : num 1
  ..@ sigma2.0            : num 0.0129
  ..@ pi                  : num [1:3] 0.343 0.43 0.227
  ..@ mu                  : num [1:3] 0.0618 0.912 -0.9163
  ..@ tau2                : num [1:3] 0.0174 0.0169 0.0177
  ..@ data                : num [1:1500] -1.23 0.755 0.923 -0.168 -1.034 ...
  ..@ data.mean           : num [1:3, 1:3] -0.20719 0.00279 0.1921 0.79998 0.98798 ...
  ..@ data.prec           : num [1:3, 1:3] 92 104 104.5 95.2 90.5 ...
  ..@ z                   : int [1:1500] 3 2 2 1 3 1 1 2 1 3 ...
  ..@ zfreq               : int [1:3] 532 656 312
  ..@ probz               : int [1:1500, 1:3] 299 0 0 0 299 0 0 0 0 299 ...
  ..@ logprior            : num -11.9
  ..@ loglik              : num -299
  ..@ mcmc.chains         :Formal class 'McmcChains' [package "CNPBayes"] with 11 slots
  .. .. ..@ theta   : num [1:300, 1:9] -0.21 -0.202 -0.211 -0.201 -0.198 ...
  .. .. ..@ sigma2  : num [1:300, 1:9] 0.012 0.0111 0.0103 0.0121 0.0107 ...
  .. .. ..@ pi      : num [1:300, 1:3] 0.313 0.333 0.353 0.366 0.348 ...
  .. .. ..@ mu      : num [1:300, 1:3] -0.00275 -0.05788 -0.03821 -0.14485 -0.04377 ...
  .. .. ..@ tau2    : num [1:300, 1:3] 0.0154 0.0164 0.0166 0.0164 0.0169 ...
  .. .. ..@ nu.0    : num [1:300] 4 3 7 8 15 1 1 1 6 6 ...
  .. .. ..@ sigma2.0: num [1:300] 0.0227 0.0411 0.0742 0.0803 0.1756 ...
  .. .. ..@ logprior: num [1:300] -12.3 -11.8 -11.9 -12 -12.7 ...
  .. .. ..@ loglik  : num [1:300] -300 -296 -291 -292 -290 ...
  .. .. ..@ zfreq   : num [1:300, 1:3] 532 532 532 532 532 532 532 532 532 532 ...
  .. .. ..@ z       : int [1:300, 1:1500] 3 3 3 3 3 3 3 3 3 3 ...
  ..@ batch               : int [1:1500] 1 1 1 1 1 1 1 1 1 1 ...
  ..@ batchElements       : Named int [1:3] 500 500 500
  .. ..- attr(*, "names")= chr [1:3] "1" "2" "3"
  ..@ hwe                 : num(0) 
  ..@ modes               :List of 10
  .. ..$ theta   : num [1:3, 1:3] -0.20335 0.00587 0.18989 0.7954 0.99248 ...
  .. ..$ sigma2  : num [1:3, 1:3] 0.01093 0.00951 0.00915 0.00982 0.01016 ...
  .. ..$ mixprob : num [1:3] 0.353 0.437 0.211
  .. ..$ mu      : num [1:3] 0.00109 0.94015 -0.99498
  .. ..$ tau2    : num [1:3] 0.0177 0.0169 0.0165
  .. ..$ nu0     : num 17
  .. ..$ sigma2.0: num 0.182
  .. ..$ zfreq   : num [1:3] 532 656 312
  .. ..$ loglik  : num -286
  .. ..$ logprior: num -12.9
  ..@ mcmc.params         :Formal class 'McmcParams' [package "CNPBayes"] with 5 slots
  .. .. ..@ thin         : int 1
  .. .. ..@ iter         : int 300
  .. .. ..@ burnin       : int 100
  .. .. ..@ nstarts      : int 5
  .. .. ..@ param_updates: Named int [1:8] 1 1 1 1 1 1 1 1
  .. .. .. ..- attr(*, "names")= chr [1:8] "theta" "sigma2" "p" "mu" ...
  ..@ .internal.constraint: num 5e-04
}}
\usage{
BatchModelExample
}
\description{
This data is a simulated example of Batch data
}
\author{
Jacob Carey
}
\keyword{datasets}

