% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BatchModelExample}
\alias{BatchModelExample}
\title{This data is a simulated example of Batch data}
\usage{
BatchModelExample
}
\value{
An example of a `BatchModel`
BatchModelExample
}
\description{
This data is a simulated example of Batch data
}
\author{
Jacob Carey
}

