% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{bic}
\alias{bic}
\alias{bic,MultiBatch-method}
\alias{bic,MultiBatchModel-method}
\alias{bic,MultiBatchP-method}
\alias{bic,MultiBatchPooled-method}
\title{Calculate BIC of a model}
\usage{
bic(object)
}
\arguments{
\item{object}{see \code{showMethods(bic)}}
}
\value{
The BIC of the model.
}
\description{
Calculate BIC of a model
}
\examples{
mb <- MultiBatchModelExample
mcmcParams(mb) <- McmcParams(iter=100, burnin=50)
mb <- posteriorSimulation(mb)
bic(mb)
}
