% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-BatchModel.R, R/methods-MarginalModel.R
\docType{methods}
\name{bic}
\alias{bic}
\alias{bic,BatchModel-method}
\alias{bic,MarginalModel-method}
\title{Calculate BIC of a model}
\usage{
bic(object)

\S4method{bic}{BatchModel}(object)

\S4method{bic}{MarginalModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(bic)}}
}
\value{
The BIC of the model.
}
\description{
Calculate BIC of a model
}
\examples{
bic(BatchModelExample)
}

