% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcParams.R
\docType{methods}
\name{burnin<-}
\alias{burnin<-}
\alias{burnin,McmcParams-method}
\alias{burnin<-,McmcParams,ANY-method}
\alias{burnin<-,McmcParams-method}
\title{Reset number of burnin iterations.}
\usage{
burnin(object) <- value

\S4method{burnin}{McmcParams}(object)

\S4method{burnin}{McmcParams,ANY}(object) <- value
}
\arguments{
\item{value}{new number of burnin iterations}
}
\description{
This function changes the number of burnin simulations to be discarded.
}
