% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcParams.R, R/methods-MixtureModel.R, R/methods-ModelList.R
\docType{methods}
\name{burnin}
\alias{burnin}
\alias{burnin,McmcParams-method}
\alias{burnin,MixtureModel-method}
\alias{burnin,ModelList-method}
\title{Number of burnin iterations.}
\usage{
burnin(object)

\S4method{burnin}{McmcParams}(object)

\S4method{burnin}{MixtureModel}(object)

\S4method{burnin}{ModelList}(object)
}
\arguments{
\item{object}{see \code{showMethods(burnin)}}
}
\description{
This function retrieves the number of burnin simulations to be discarded.
}

