% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{burnin}
\alias{burnin}
\title{Number of burnin iterations.}
\usage{
burnin(object)
}
\arguments{
\item{object}{see \code{showMethods(burnin)}}
}
\value{
The number of burnin simulations.
}
\description{
This function retrieves the number of burnin simulations to be discarded.
}
\examples{
burnin(SingleBatchModelExample)
mp <- mcmcParams(SingleBatchModelExample)
burnin(mp)
}
