% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{chains}
\alias{chains}
\alias{chains,MixtureModel-method}
\alias{chains,MultiBatch-method}
\alias{chains,MultiBatchList-method}
\title{Retrieve simulated chains from model object.}
\usage{
chains(object)
}
\arguments{
\item{object}{\code{showMethods(chains)}}
}
\value{
The simulated chains.
}
\description{
The method \code{chains} applied to a \code{MixtureModel}-derived
class will return an object of class \code{McmcChains} that
contains the chains for all simulated parameters. Typically, chains
is called in conjunction with an accessor for one of these
parameters.
}
\examples{
    theta.chain <- theta(chains(MultiBatchModelExample))
    dim(theta.chain)
    plot.ts(theta.chain, plot.type="single",
            col=seq_len(k(SingleBatchModelExample)))
}
