% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{chains<-}
\alias{chains}
\alias{chains,MixtureModel-method}
\alias{chains<-}
\alias{chains<-,MixtureModel-method}
\title{Replace model's simulated MCMC chain.}
\usage{
chains(object) <- value

chains(object)

\S4method{chains}{MixtureModel}(object)

\S4method{chains}{MixtureModel}(object) <- value
}
\arguments{
\item{object}{\code{showMethods(chains)}}

\item{value}{object with which to replace the McmcChains}
}
\description{
Replace model's simulated MCMC chain.

Retrieve simulated chains from model object.
}

