% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{chains<-}
\alias{chains}
\alias{chains,MixtureModel-method}
\alias{chains<-}
\alias{chains<-,MixtureModel-method}
\title{Replace model's simulated MCMC chain.}
\usage{
chains(object) <- value

chains(object)

\S4method{chains}{MixtureModel}(object)

\S4method{chains}{MixtureModel}(object) <- value
}
\arguments{
\item{object}{\code{showMethods(chains)}}

\item{value}{object with which to replace the McmcChains}
}
\description{
Replace model's simulated MCMC chain.

The method \code{chains} applied to a \code{MixtureModel}-derived
class will return an object of class \code{McmcChains} that
contains the chains for all simulated parameters. Typically, chains
is called in conjunction with an accessor for one of these
parameters.
}
\examples{
\dontrun{
   data(MarginalModelExample)
   theta.chain <- theta(chains(MarginalModelExample))
   dim(theta.chain)
   plot.ts(theta.chain, plot.type="single",
           col=seq_len(k(MarginalModelExample)))
 }
}

