% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{chromosome}
\alias{chromosome}
\title{Extract character vector of sequence names}
\usage{
chromosome(object, ...)
}
\arguments{
\item{object}{a \code{GRanges} instance}

\item{...}{currently ignored}
}
\value{
A character vector
}
\description{
Short cut for \code{as.character(seqnames(g))} where g is a
\code{GRanges} object.
}
\examples{
\dontrun{
   g <- GRanges("chr1", IRanges(10, 15))
   chromosome(g)
}
}

