% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-DensityModel.R
\docType{methods}
\name{clusters}
\alias{clusters}
\alias{clusters,DensityModel-method}
\title{Accessor for extracting the kmeans clusters from a DensityModel
instance}
\usage{
clusters(object)

\S4method{clusters}{DensityModel}(object)
}
\arguments{
\item{object}{an instance of class 'DensityModel'}
}
\value{
k-means clustering of the component means using the modes as centers.
}
\description{
Accessor for extracting the kmeans clusters from a DensityModel
instance
}
\examples{
truth <- simulateData(N=2500, p=rep(1/3, 3),
                      theta=c(-1, 0, 1),
                      sds=rep(0.1, 3))
dm <- DensityModel(truth)
clusters(dm)
}
\seealso{
\code{\link{DensityModel-class}}
}

