% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-BatchModel.R, R/methods-SummarizedExperiment.R
\docType{methods}
\name{collapseBatch}
\alias{collapseBatch}
\alias{collapseBatch,BatchModel-method}
\alias{collapseBatch,SummarizedExperiment-method}
\alias{collapseBatch,numeric-method}
\title{Collapse batch data to fewer number of batches.}
\usage{
collapseBatch(object, plate, THR = 0.1)

\S4method{collapseBatch}{BatchModel}(object)

\S4method{collapseBatch}{SummarizedExperiment}(object, plate, THR = 0.1)

\S4method{collapseBatch}{numeric}(object, plate, THR = 0.1)
}
\arguments{
\item{object}{see \code{showMethods(collapseBatch)}}

\item{plate}{a vector labelling from which batch each observation came from.}

\item{THR}{threshold below which the null hypothesis should be rejected and batches are collapsed.}
}
\description{
For observations from different batches that are likely drawn from the same distribution, collapse the batches. This is done for computational efficiency. The Kolmogorov-Smirnov test is used.
}

