% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/copynumber-models.R
\docType{methods}
\name{CopyNumberModel}
\alias{CopyNumberModel}
\alias{SingleBatchCopyNumber}
\alias{MultiBatchCopyNumber}
\alias{MultiBatchCopyNumberPooled}
\alias{CopyNumberModel,SingleBatchModel-method}
\alias{CopyNumberModel,MultiBatchModel-method}
\alias{CopyNumberModel,MultiBatchPooled-method}
\alias{mapCopyNumber}
\title{Constructs a CopyNumberModel from SB, SBP, MB, or MBP models}
\usage{
CopyNumberModel(model, params = mapParams())

SingleBatchCopyNumber(model)

MultiBatchCopyNumber(model)

MultiBatchCopyNumberPooled(model)

\S4method{CopyNumberModel}{SingleBatchModel}(model, params = mapParams())

\S4method{CopyNumberModel}{MultiBatchModel}(model, params = mapParams())

\S4method{CopyNumberModel}{MultiBatchPooled}(model, params = mapParams())

mapCopyNumber(model, params = mapParams())
}
\arguments{
\item{model}{a SB, SBP, MB, or MBP model}

\item{params}{a list of parameters used for mapping mixture components to copy number states.}
}
\value{
a \code{MultiBatchCopyNumber} instance

a \code{MultiBatchCopyNumber} instance
}
\description{
The mixture components do not necessarily reflect distinct copy number
states, possibly due to skewed (non-Gaussian) log R ratios. While easy to fit skewed data with a finite mixture of Gaussians, additional steps are needed to assess whether the components correspond to distinct copy number states.  This accessor \code{copyNumber} returns the copy number states -- i.e., the result after mapping mixture components to copy number states.
}
\examples{
sb <- SingleBatchModelExample
cn.model <- CopyNumberModel(sb, mapParams())
}
\seealso{
\code{\link{copyNumber}}
}
