% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber-models.R
\docType{methods}
\name{mapping}
\alias{mapping}
\alias{mapping<-}
\alias{probCopyNumber,MultiBatchCopyNumber-method}
\alias{probCopyNumber,SingleBatchCopyNumber-method}
\title{Mapping of mixture component indices to copy number states}
\usage{
mapping(object)

mapping(object) <- value

\S4method{probCopyNumber}{SingleBatchCopyNumber}(model)

\S4method{probCopyNumber}{MultiBatchCopyNumber}(model)
}
\arguments{
\item{object}{a \code{SingleBatchCopyNumber} or \code{MultiBatchCopyNumber} instance}
}
\description{
Mapping of mixture component indices to copy number states
}
\examples{
cn.model <- SingleBatchCopyNumber(MarginalModelExample)
map <- mapComponents(cn.model)
mapping(cn.model) <- map
}

