% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Hyperparameters.R
\docType{methods}
\name{dfr}
\alias{dfr}
\alias{dfr<-}
\alias{dfr,Hyperparameters-method}
\alias{dfr<-,Hyperparameters,numeric-method}
\title{Accessor for degrees of freedom}
\usage{
dfr(object)

dfr(object) <- value

\S4method{dfr}{Hyperparameters}(object)

\S4method{dfr}{Hyperparameters,numeric}(object) <- value
}
\arguments{
\item{object}{a Hyperparameters or Hyperparameters-derived class}
}
\description{
Accessor for degrees of freedom
}
\examples{
  hp <- Hyperparameters()
  dfr(hp)
  dfr(hp) <- 10
  dfr(hp)
}
