% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downSample}
\alias{downSample}
\title{Down sample the observations in a mixture}
\usage{
downSample(y, batches, size = 1000, min.batchsize = 75)
}
\arguments{
\item{y}{vector containing data}

\item{batch}{a vector containing the labels from which batch each observation came from.}
}
\value{
A tibble
}
\description{
For large datasets (several thousand subjects), the computational burden for fitting Bayesian mixture models can be high.  Downsampling can reduce the computational burden with little effect on inference.  This function draws a random sample with replacement.  Batches with few observations are combined with larger batches that have a similar median log R ratio.
}
\examples{
  mb <- MultiBatchModelExample
  ds <- downSample(y(mb), 200, batch(mb))
  mp <- McmcParams(iter=50, burnin=100)
  mb <- MultiBatchModel2(dat=tile.summaries$avgLRR,
                         batches=ds$batch, mp=mp)
  mb <- posteriorSimulation(mb)
  ggMixture(mb)
  mb2 <- upSample(mb, tiled.medians)
  ggMixture(mb2)
}
