% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{downsample}
\alias{downsample}
\title{Create tile labels for each observation}
\usage{
downsample(batch.file, plate, y, ntiles = 250, THR = 0.1)
}
\arguments{
\item{batch.file}{the name of a file contaning RDS data to be read in.}

\item{plate}{a vector containing the labels  from which batch each observation came from.}

\item{y}{in memory data}

\item{ntiles}{number of tiles in a batch}

\item{THR}{threshold above which to merge batches in Kolmogorov-Smirnov test.}
}
\value{
Tile labels for each observation
}
\description{
A wrapper for function downSampleEachBatch. Batches are automatically merged as needed.
}

