% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/functions.R
\name{upSample}
\alias{upSample}
\alias{downSampleEachBatch}
\title{Restore model to original dimension}
\usage{
upSample(model, tiles)

downSampleEachBatch(y, nt, batch)
}
\arguments{
\item{y}{vector containing data}

\item{nt}{the number of observations per batch}

\item{batch}{a vector containing the labels from which batch each observation came from.}
}
\value{
a SB, MB, SBP, or MBP model

Tile labels for each observation
}
\description{
As the name suggests, the function downSampleEach batch can be used to down sample the observations in each batch. The function upSample restores the model object to the original dimension.

Create tile labels for each observation
}
\examples{

y <- runif(100)
batch <- sample(letters[1:3], 100, replace=TRUE)
ds <- downSampleEachBatch(y, 10, batch)

model <- MultiBatchModelExample
ds <- downSampleEachBatch(y(model), 100, batch(model))
model.ds <- MultiBatchModel(ds$y, batch=ds$batch, k=3)
model.ds <- posteriorSimulation(model.ds)
## map the posterior probabilities of the downsampled data back to the
## original observations
probs <- probz(model.ds)
rownames(probs) <- names(y(model.ds))
probs.ds <- probs[ds$label, ]

## compare downsampled results to that of fitting the fulll data
model <- posteriorSimulation(model)
probs.full <- probz(model)
probs.full <- round(probs.full, 1)
reduced <- round(probs.ds, 1)
not.equal <- rowSums(full != reduced) > 0
}
\seealso{
\code{\link[dplyr]{ntile}}
}
