% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{downSampleEachBatch}
\alias{downSampleEachBatch}
\title{Create tile labels for each observation}
\usage{
downSampleEachBatch(y, nt, batch)
}
\arguments{
\item{y}{vector containing data}

\item{nt}{the number of tiles in a batch}

\item{batch}{a vector containing the labels from which batch each observation came from.}
}
\value{
Tile labels for each observation
}
\description{
Create tile labels for each observation
}

