% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-McmcChains.R, R/methods-McmcParams.R,
%   R/methods-MultiBatchModel.R
\docType{methods}
\name{[,McmcChains,ANY,ANY,ANY-method}
\alias{[,McmcChains,ANY,ANY,ANY-method}
\alias{[,McmcChains-method}
\alias{[,McmcChains,ANY-method}
\alias{[,McmcParams,ANY,ANY,ANY-method}
\alias{[,McmcParams-method}
\alias{[,McmcParams,ANY-method}
\alias{extract}
\alias{[,BatchModel-method}
\alias{[,BatchModel,ANY-method}
\alias{[,BatchModel,ANY,ANY-method}
\alias{[,BatchModel,ANY,ANY,ANY-method}
\title{extract estimated parameters at particular iteration of simulation.}
\usage{
\S4method{[}{McmcChains,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{McmcParams,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{BatchModel,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{An object of class BatchModel, McmcChains, or McmcParams}

\item{i}{An element of the instance to be extracted.}

\item{j}{Not used.}

\item{...}{Not used.}

\item{drop}{Not used.}
}
\value{
An object of class 'McmcChains'

An object of class 'McmcParams'

An object of class 'BatchModel'
}
\description{
extract estimated parameters at particular iteration of simulation.

Allows a user to pass a vector for burnin, thin, and iter.

extract data, latent variable, and batch for given observation
}
