% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-McmcChains.R
\docType{methods}
\name{[,McmcChains,ANY,ANY,ANY-method}
\alias{[,McmcChains,ANY,ANY,ANY-method}
\alias{[,McmcChains-method}
\alias{[,McmcChains,ANY-method}
\alias{[,McmcChainsTrios,ANY,ANY,ANY-method}
\alias{[,McmcChainsTrios-method}
\alias{[,McmcChainsTrios,ANY-method}
\title{extract estimated parameters at particular iteration of simulation.}
\usage{
\S4method{[}{McmcChains,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{McmcChainsTrios,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\value{
An object of class 'McmcChains'

An object of class 'McmcChainsTrios'
}
\description{
extract estimated parameters at particular iteration of simulation.

extract estimated parameters at particular iteration of simulation.
}
