% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-BatchModel.R, R/methods-McmcChains.R, R/methods-McmcParams.R, R/methods-ModelList.R
\docType{methods}
\name{extract}
\alias{[,BatchModel,ANY,ANY,ANY-method}
\alias{[,BatchModel,ANY,ANY-method}
\alias{[,BatchModel,ANY-method}
\alias{[,BatchModel-method}
\alias{[,McmcChains,ANY-method}
\alias{[,McmcChains-method}
\alias{[,McmcParams,ANY-method}
\alias{[,McmcParams-method}
\alias{[,ModelList,ANY-method}
\alias{[,ModelList-method}
\alias{[[,ModelList,ANY,ANY-method}
\alias{[[,ModelList-method}
\alias{[[<-,ModelList,ANY,ANY,ANY-method}
\alias{[[<-,ModelList-method}
\alias{extract}
\title{extract data, latent variable, and batch for given observation}
\usage{
\S4method{[}{BatchModel,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{McmcChains,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{McmcParams,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[}{ModelList,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{ModelList,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{[[}{ModelList,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE) <- value
}
\arguments{
\item{x}{An object of class BatchModel, McmcChains, McmcParams, or ModelList.}

\item{i}{An element of the instance to be extracted.}

\item{j}{Not used.}

\item{...}{Not used.}

\item{drop}{Not used.}

\item{value}{A MarginalModel or BatchModel to insert.}
}
\description{
extract data, latent variable, and batch for given observation

extract estimated parameters at particular iteration of simulation.

Allows a user to pass a vector for burnin, thin, and iter.

Extract a particular Marginal or BatchModel but still as class ModelList.

Extract a particular Marginal or BatchModel.

Replace a MarginalModel or BatchModel in a ModelList.
}

