% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plot-functions.R,
%   R/deprecated_functions.R
\docType{methods}
\name{ggChains}
\alias{ggChains}
\alias{ggMixture}
\alias{ggSingleBatch}
\alias{ggMultiBatch}
\alias{ggSingleBatchChains}
\alias{multiBatchDensities}
\alias{ggMixture,MultiBatchCopyNumber-method}
\alias{ggMixture,MultiBatchCopyNumberPooled-method}
\alias{ggMixture,SingleBatchModel-method}
\alias{ggMixture,MultiBatchModel-method}
\alias{ggMixture,MultiBatchPooled-method}
\alias{ggMixture,SingleBatchCopyNumber-method}
\alias{ggChains,MultiBatchModel-method}
\alias{ggMultiBatch,BatchModel-method}
\alias{ggMultiBatch,MultiBatchModel-method}
\alias{ggMultiBatch,MultiBatchPooled-method}
\alias{ggSingleBatch,MarginalModel-method}
\alias{ggSingleBatch,SingleBatchModel-method}
\alias{ggMultiBatch,MultiBatchCopyNumber-method}
\alias{ggSingleBatch,SingleBatchCopyNumber-method}
\title{Trace plots of MCMC chains and mixture model densities}
\usage{
ggChains(model)

ggMixture(model, bins)

ggSingleBatch(model, bins)

ggMultiBatch(model, bins)

ggSingleBatchChains(model)

multiBatchDensities(object)

\S4method{ggMixture}{MultiBatchCopyNumber}(model, bins)

\S4method{ggMixture}{MultiBatchCopyNumberPooled}(model, bins)

\S4method{ggMixture}{SingleBatchModel}(model, bins)

\S4method{ggMixture}{MultiBatchModel}(model, bins)

\S4method{ggMixture}{MultiBatchPooled}(model, bins)

\S4method{ggMixture}{SingleBatchCopyNumber}(model, bins)

\S4method{ggChains}{MultiBatchModel}(model)

\S4method{ggMultiBatch}{BatchModel}(model, bins)

\S4method{ggMultiBatch}{MultiBatchModel}(model, bins)

\S4method{ggMultiBatch}{MultiBatchPooled}(model, bins)

\S4method{ggSingleBatch}{MarginalModel}(model, bins)

\S4method{ggSingleBatch}{SingleBatchModel}(model, bins)

\S4method{ggMultiBatch}{MultiBatchCopyNumber}(model, bins)

\S4method{ggSingleBatch}{SingleBatchCopyNumber}(model, bins)
}
\arguments{
\item{model}{A SB, MB, SBP, or MBP model}

\item{bins}{a length-one numeric vector indicating the number of bins -- passed to \code{geom_hist}}

\item{model}{a \code{BatchModel} or  \code{MarginalModel} object}

\item{bins}{length-one integer vector indicating the number of bins for the histograms (passed to \code{geom_histogram})}
}
\value{
A \code{gg} object

A \code{gg} object

a \code{ggplot} object

a list of \code{ggplot} objects. Chains are grouped by the length of
  the parameter vector. For example, in the single-batch model, the means
  (theta) and variances (sigma2) are component-specific (length k, where k is
  number of components) and are plotted together in a single \code{ggplot}
  object.
}
\description{
The \code{ggChains} method provides a convenient wrapper for plotting the chains of all parameters in the various mixture model implementations.  In addition to the estimated number of independent MCMC draws (effective sample size) and Gelman-Rubin convergence diagnostics implemented in \code{gibbs}, visualization of the chains is helpful for assessing convergence.

ggplot wrapper for plotting the data at a single CNP and the model-based densities

ggplot-style functions for diagnosing convergence
}
\details{
The \code{ggMixture} method overlays the posterior approximation of the Gaussian mixture on the empirical data.
}
\examples{
  sb <- SingleBatchModelExample
  iter(sb, force=TRUE) <- 1000
  burnin(sb) <- 100
  sb <- posteriorSimulation(sb)
  fig.chains <- ggChains(sb)
  ## component-specific chains
  fig.chains[["comp"]]
  ## single-parameter chains and log-likelihood
  fig.chains[["single"]]

  ## plot the mixture
  fig.mix <- ggMixture(sb)
ggMultiBatch(MultiBatchModelExample)
ggSingleBatch(SingleBatchModelExample)
plist.sb <- ggSingleBatchChains(MarginalModelExample)
## chains for parameter vectors of length k
plist.sb[["comp"]]
## chains for parameters vectors of length 1
plist.sb[["single"]]
plist.mb <- ggMultiBatchChains(BatchModelExample)
## chains for parameters that are batch- and component-specific
plist.mb[["batch"]]
## chains for parameters vectors of length k
plist.mb[["comp"]]
## chains for parameter vectors of length 1
plist.mb[["single"]]
df <- multiBatchDensities(MultiBatchModelExample)
head(df)
}
\seealso{
\code{\link{gibbs}}
}
