% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Hyperparameters-class}
\alias{Hyperparameters-class}
\alias{k,Hyperparameters-method}
\title{An object to specify the hyperparameters of a model.}
\description{
An object to specify the hyperparameters of a model.
}
\section{Slots}{

\describe{
\item{\code{k}}{Number of components}

\item{\code{mu.0}}{Prior mean for mu.}

\item{\code{tau2.0}}{prior variance on mu}

\item{\code{eta.0}}{rate paramater for tau2}

\item{\code{m2.0}}{shape parameter for tau2}

\item{\code{alpha}}{mixture probabilities}

\item{\code{beta}}{parameter for nu.0 distribution}

\item{\code{a}}{shape for sigma2.0}

\item{\code{b}}{rate for sigma2.0}
}}

