% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Hyperparameters.R
\name{Hyperparameters}
\alias{Hyperparameters}
\title{Create an object of class 'Hyperparameters'}
\usage{
Hyperparameters(type = "batch", k = 2L, ...)
}
\arguments{
\item{type}{specifies 'marginal' or 'batch'}

\item{k}{number of components}

\item{...}{additional arguments are names of the slots of the Hyperparameters class, including `mu.0`, `tau2.0`, `eta.0`, `m2.0`, `alpha`, `beta`, `a`, `b`, and `dfr`).}
}
\value{
An object of class HyperparametersMarginal or HyperparametersBatch
}
\description{
Create an object of class 'Hyperparameters'
}
\details{
Additional hyperparameters can be passed to the HyperparametersMarginal and HyperparametersBatch models.
}
