% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-Hyperparameters.R
\name{Hyperparameters}
\alias{Hyperparameters}
\title{Create an object of class 'Hyperparameters'}
\usage{
Hyperparameters(type = "batch", k = 2L, ...)
}
\arguments{
\item{type}{specifies 'marginal' or 'batch'}

\item{k}{number of components}

\item{...}{optional parameters.  See details}
}
\value{
An object of class HyperparametersMarginal or HyperparametersBatch
}
\description{
Create an object of class 'Hyperparameters'
}
\details{
Additional hyperparameters can be passed to the
HyperparametersMarginal and HyperparametersBatch models.
}
\examples{
hypp <- Hyperparameters("marginal", k=2)
}

