% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{hyperParams}
\alias{hyperParams}
\alias{hyperParams,MixtureModel-method}
\alias{hyperParams,MultiBatch-method}
\alias{hyperParams,MultiBatchList-method}
\alias{hyperParams<-}
\alias{hyperParams<-,MixtureModel,Hyperparameters-method}
\alias{hyperParams<-,MultiBatch,Hyperparameters-method}
\alias{hyperParams<-,MultiBatchList,Hyperparameters-method}
\title{Accessor for Hyperparameters object for a MixtureModel-derived object}
\usage{
hyperParams(object)

hyperParams(object) <- value
}
\arguments{
\item{object}{see \code{showMethods(hyperParams)}}

\item{value}{an object of class 'Hyperparameters'}
}
\value{
The Hyperparameters of a MixtureModel
}
\description{
Accessor for Hyperparameters object for a MixtureModel-derived object

Replace the hyperparameters for a \code{MixtureModel}-derived object
}
\examples{
    hyperParams(SingleBatchModelExample)
hypp <- Hyperparameters(type="marginal",
                        k=k(SingleBatchModelExample),
                        alpha=c(9, 9, 10))
hyperParams(SingleBatchModelExample) <- hypp

}
