% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{hyperParams}
\alias{hyperParams}
\alias{hyperParams,MixtureModel-method}
\alias{hyperParams<-}
\alias{hyperParams<-,MixtureModel,Hyperparameters-method}
\alias{hyperParams<-,MixtureModel-method}
\title{Accessor for Hyperparameters object for a MixtureModel-derived object}
\usage{
hyperParams(object)

hyperParams(object) <- value

\S4method{hyperParams}{MixtureModel}(object)

\S4method{hyperParams}{MixtureModel,Hyperparameters}(object) <- value
}
\arguments{
\item{object}{see \code{showMethods(hyperParams)}}

\item{value}{an object of class 'Hyperparameters'}
}
\description{
Accessor for Hyperparameters object for a MixtureModel-derived object

Replace the hyperparameters for a MixtureModel-derived object
}
\examples{
\dontrun{
    data(MarginalModelExample)
    hyperParams(MixtureModelExample)
}
\dontrun{
   data(MarginalModelExample)
   hypp <- Hyperparameters(type="marginal",
                           k=k(MarginalModelExample))
   ## TODO: change one or two of the hyperparameters in the object
   ## 'hypp'
   ##
   hyperParams(MarginalModelExample) <- hypp
}
}

