% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcParams.R, R/methods-MixtureModel.R, R/methods-ModelList.R
\docType{methods}
\name{iter}
\alias{iter}
\alias{iter,McmcParams-method}
\alias{iter,MixtureModel-method}
\alias{iter,ModelList-method}
\title{Number of MCMC iterations.}
\usage{
iter(object)

\S4method{iter}{McmcParams}(object)

\S4method{iter}{MixtureModel}(object)

\S4method{iter}{ModelList}(object)
}
\arguments{
\item{object}{see \code{showMethods(iter)}}
}
\description{
This function retrieves the number of iterations of an MCMC simulation.
}

