% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcParams.R,
%   R/methods-MixtureModel.R
\docType{methods}
\name{iter<-}
\alias{iter<-}
\alias{iter}
\alias{iter,McmcParams-method}
\alias{iter<-,McmcParams,ANY-method}
\alias{iter<-,McmcParams-method}
\alias{iter,MixtureModel-method}
\alias{iter<-,MixtureModel,ANY-method}
\alias{iter<-,MixtureModel-method}
\title{Reset number of iterations.}
\usage{
iter(object) <- value

iter(object)

\S4method{iter}{McmcParams}(object)

\S4method{iter}{McmcParams,ANY}(object) <- value

\S4method{iter}{MixtureModel}(object)

\S4method{iter}{MixtureModel,ANY}(object) <- value
}
\arguments{
\item{object}{see \code{showMethods(iter)}}

\item{value}{new number of iterations}

\item{force}{Allow changing of the size of the elements?}
}
\value{
The number of MCMC iterations
}
\description{
This function changes the number of simulations.

This function retrieves the number of iterations of an MCMC simulation.
}
\examples{
     iter(SingleBatchModelExample)
}
