% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-Hyperparameters.R
\name{k<-,Hyperparameters-method}
\alias{k<-,Hyperparameters-method}
\alias{k<-,HyperparametersTrios-method}
\title{Accessor for number of mixture components
'}
\usage{
\S4method{k}{Hyperparameters}(object) <- value

\S4method{k}{HyperparametersTrios}(object) <- value
}
\description{
Accessor for number of mixture components
'
}
