% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R, R/methods-ModelList.R, R/methods-ModelParams.R
\docType{methods}
\name{k}
\alias{k}
\alias{k,MixtureModel-method}
\alias{k,ModelList-method}
\alias{k,ModelParams-method}
\title{Number of components.}
\usage{
k(object)

\S4method{k}{MixtureModel}(object)

\S4method{k}{ModelList}(object)

\S4method{k}{ModelParams}(object)
}
\arguments{
\item{object}{see \code{showMethods(k)}}
}
\description{
This function retrieves the number of a priori components.
}

