% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-BatchModel.R, R/methods-DensityModel.R, R/methods-MarginalModel.R, R/methods-MixtureModel.R, R/methods-ModelList.R, R/methods-ModelParams.R
\docType{methods}
\name{k}
\alias{k}
\alias{k,DensityModel-method}
\alias{k,MixtureModel-method}
\alias{k,ModelList-method}
\alias{k,ModelParams-method}
\alias{k<-}
\alias{k<-,BatchModel-method}
\alias{k<-,MarginalModel-method}
\title{Number of components.}
\usage{
k(object)

k(object) <- value

\S4method{k}{MarginalModel}(object) <- value

\S4method{k}{DensityModel}(object)

\S4method{k}{MarginalModel}(object) <- value

\S4method{k}{MixtureModel}(object)

\S4method{k}{ModelList}(object)

\S4method{k}{ModelParams}(object)
}
\arguments{
\item{object}{see \code{showMethods(k)}}

\item{value}{An integer for the new number of components.}
}
\description{
This function retrieves the number of a priori components.

Creates a new MarginalModel or BatchModel with the same data,
hyperparameters, MCMC options, and batch (for BatchModel).
}
\examples{
data(MarginalModelExample)
k(MarginalModelExample) <- 5
}

