% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-Hyperparameters.R,
%   R/methods-MixtureModel.R, R/methods-TrioBatchModel.R
\docType{methods}
\name{k}
\alias{k}
\alias{k<-}
\alias{k<-,Hyperparameters-method}
\alias{k<-,HyperparametersTrios-method}
\alias{k,MixtureModel-method}
\alias{k<-,MixtureModel-method}
\alias{k,TrioBatchModel-method}
\alias{k<-,TrioBatchModel-method}
\title{Number of components.}
\usage{
k(object)

k(object) <- value

\S4method{k}{Hyperparameters}(object) <- value

\S4method{k}{HyperparametersTrios}(object) <- value

\S4method{k}{MixtureModel}(object)

\S4method{k}{MixtureModel}(object) <- value

\S4method{k}{TrioBatchModel}(object)

\S4method{k}{TrioBatchModel}(object) <- value
}
\arguments{
\item{object}{see \code{showMethods(k)}}

\item{value}{An integer for the new number of components.}
}
\value{
The number of components
}
\description{
This function retrieves the number of a priori components.

Updates the number of components and erases chains from a previous
posteriorSimulation (if one was performed). Draws from prior to guess
new starting values.
}
\examples{
k(SingleBatchModelExample) <- 2
}
