% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-DensityModel.R, R/methods-Hyperparameters.R, R/methods-MixtureModel.R, R/methods-ModelParams.R
\docType{methods}
\name{k}
\alias{k}
\alias{k,DensityModel-method}
\alias{k,MixtureModel-method}
\alias{k,ModelParams-method}
\alias{k<-}
\alias{k<-,Hyperparameters-method}
\alias{k<-,Hyperparemeters-method}
\alias{k<-,MixtureModel-method}
\title{Number of components.}
\usage{
k(object)

k(object) <- value

\S4method{k}{DensityModel}(object)

\S4method{k}{Hyperparameters}(object) <- value

\S4method{k}{MixtureModel}(object)

\S4method{k}{MixtureModel}(object) <- value

\S4method{k}{ModelParams}(object)
}
\arguments{
\item{object}{see \code{showMethods(k)}}

\item{value}{An integer for the new number of components.}
}
\description{
This function retrieves the number of a priori components.

Updates the number of components and erases chains from a previous
posteriorSimulation (if one was performed). Draws from prior to guess
new starting values.
}
\examples{
k(MarginalModelExample) <- 2
}

