% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiBatch.R
\name{kolmogorov_batches}
\alias{kolmogorov_batches}
\title{Group provisional batch labels by similarity of eCDFs}
\usage{
kolmogorov_batches(dat, KS_cutoff)
}
\arguments{
\item{dat}{typically a `tibble` gotten by `assays(MultiBatch)`}

\item{KS_cutoff}{scalar indicating cutoff for Kolmogorov-Smirnov p-value.  Two provisional batches with p-value above this cutoff are combined.}
}
\description{
This function groups provisional batches (e.g., chemistry plate, date, or DNA source) by similarity of the empirical cummulative distribution function (eCDF).  Similarity of the eCDFs is based on the p-value from Kolmogorov-Smirnov  test statistic.  All pairwise combinations of batches are compared, and repeated recursively until no two batches can be combined.
}
\seealso{
[ks.test()]
}
