% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{label_switch}
\alias{label_switch}
\alias{label_switch,MixtureModel-method}
\title{Accessor for determing whether label switching occurred during MCMC}
\usage{
label_switch(object)

\S4method{label_switch}{MixtureModel}(object)
}
\arguments{
\item{object}{MixtureModel-derived class}
}
\description{
Accessor for determing whether label switching occurred during MCMC
}
\examples{
label_switch(SingleBatchModelExample)
}
