% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R, R/methods-MixtureModel.R
\docType{methods}
\name{log_lik}
\alias{log_lik}
\alias{log_lik,McmcChains-method}
\alias{log_lik,MixtureModel-method}
\title{Retrieve log likelihood.}
\usage{
log_lik(object)

\S4method{log_lik}{McmcChains}(object)

\S4method{log_lik}{MixtureModel}(object)
}
\arguments{
\item{object}{see showMethods(log_lik)}
}
\value{
The log likelihood
}
\description{
Retrieve log likelihood.
}
\examples{
## retrieve log likelihood at each MCMC iteration
log_lik(chains(MarginalModelExample))
## retrieve log likelihood at last MCMC iteration
log_lik(MarginalModelExample)
}

