% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MarginalModel.R
\name{logBayesFactor}
\alias{logBayesFactor}
\title{Compute the log bayes factor between models.}
\usage{
logBayesFactor(x)
}
\arguments{
\item{x}{the result of a call to \code{computeMarginalLik}.}
}
\value{
Log Bayes factor comparing the two models with highest likelihood.
}
\description{
Models of varying component sizes are compared. The log bayes factor is
calculated comparing the two models with the highest marginal likelihood,
as computed by \code{marginalLikelihood}.
}

