% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{map}
\alias{map}
\title{Calculate the maximum a posteriori estimate of latent variable assignment.}
\usage{
map(object)
}
\arguments{
\item{object}{an object of class MixtureModel.}
}
\description{
Calculate the maximum a posteriori estimate of latent variable assignment.
}
\examples{
map(MarginalModelExample)
}

