% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{mapCnProbability}
\alias{mapCnProbability}
\title{Probabiliistic copy number assigments.}
\usage{
mapCnProbability(model)
}
\arguments{
\item{model}{An object of class MixtureModel.}
}
\value{
A matrix of size N x K where N is number of observations and K is the number of components.
}
\description{
Calculate probabilistic copy number assignments using Bayes Rule applied at the MAP estimates of the cluster mean, variance, and class proportion parameters
}

