% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber-models.R
\name{mapComponents}
\alias{mapComponents}
\title{Map mixture components to distinct copy number states}
\usage{
mapComponents(model, params = mapParams())
}
\arguments{
\item{model}{a SB, MB, SBP, or MBP model}

\item{params}{a list of mapping parameters}
}
\description{
Map mixture components to distinct copy number states
}
\examples{
## Batch model
bmodel <- MultiBatchModelExample
cn.model <- CopyNumberModel(bmodel)
mapping(cn.model)
\dontrun{
  ggMixture(cn.model)
}
}
\seealso{
\code{\link{CopyNumberModel}} \code{\link{mapParams}}
}
