% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copynumber-models.R
\name{mapComponents}
\alias{mapComponents}
\title{Map mixture components to distinct copy number states}
\usage{
mapComponents(model, params = mapParams())
}
\arguments{
\item{model}{\code{SingleBatchCopyNumber} or \code{{MultiBatchCopyNumber}} model}

\item{params}{a list of mapping parameters}
}
\description{
Map mixture components to distinct copy number states
}
\examples{
mm <- MarginalModelExample
cn.model <- SingleBatchCopyNumber(mm)
mapping(cn.model) <- mapComponents(cn.model)
mapping(cn.model)
\dontrun{
 ggSingleBatch(mm)
}
## Batch model
bmodel <- BatchModelExample
bmodel <- MultiBatchCopyNumber(bmodel)
mapping(bmodel) <- mapComponents(bmodel)
mapping(bmodel)
}

