% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/copynumber-models.R
\docType{methods}
\name{mapping}
\alias{mapping}
\alias{mapping<-}
\alias{mapping,MultiBatchCopyNumber-method}
\alias{mapping,MultiBatchCopyNumberPooled-method}
\alias{mapping,MultiBatchCopyNumber-method}
\alias{mapping<-,MultiBatchCopyNumber,character-method}
\alias{mapping,MultiBatchCopyNumber,numeric-method}
\alias{mapping<-,MultiBatchCopyNumberPooled,character-method}
\alias{mapping,MultiBatchCopyNumberPooled,numeric-method}
\title{Map mixture components to copy number states}
\usage{
mapping(object)

mapping(object) <- value

\S4method{mapping}{MultiBatchCopyNumber}(object)

\S4method{mapping}{MultiBatchCopyNumberPooled}(object)

\S4method{mapping}{MultiBatchCopyNumber,character}(object) <- value

\S4method{mapping}{MultiBatchCopyNumberPooled,character}(object) <- value
}
\arguments{
\item{object}{a SB, SBP, MB, or MBP model}

\item{value}{a k-length numeric vector with values in {1, 2, ..., k}, where k is the number of mixture components}
}
\description{
Map mixture components to copy number states
}
\examples{
cn.model <- CopyNumberModel(SingleBatchModelExample)
## manually remap first two components to the same copy number state
mapping(cn.model) <- c(1, 1, 2)
\dontrun{
 ggMixture(cn.model)
}
}
\seealso{
\code{\link{CopyNumberModel}}
}
