% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{marginal_lik}
\alias{marginal_lik}
\alias{marginal_lik,MixtureModel-method}
\title{Accessor for marginal likelihood}
\usage{
marginal_lik(object)

\S4method{marginal_lik}{MixtureModel}(object)
}
\arguments{
\item{object}{a SB, SBP, MB, or MBP model}
}
\description{
Accessor for marginal likelihood
}
\examples{
sb <- SingleBatchModelExample
marginal_lik(sb)
}
\seealso{
\code{\link{marginalLikelihood}}
}
