% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{marginal_lik}
\alias{marginal_lik}
\alias{marginal_lik<-}
\alias{marginal_lik,MixtureModel-method}
\alias{marginal_lik<-,MixtureModel,numeric-method}
\title{Accessor for the log marginal likelihood of a SB, SBP, MB, or MBP model}
\usage{
marginal_lik(object)

marginal_lik(object) <- value

\S4method{marginal_lik}{MixtureModel}(object)

\S4method{marginal_lik}{MixtureModel,numeric}(object) <- value
}
\arguments{
\item{object}{a SB, SBP, MB, or MBP model}
}
\description{
The marginal likelihood is computed by Chib's estimator (JASA, Volume 90 (435), 1995).
}
\examples{
sb <- SingleBatchModelExample
marginal_lik(sb)
}
\seealso{
See \code{\link{marginalLikelihood}} for computing the marginal likelihood of a mixture model.
}
