% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/marginal_likelihood.R
\docType{methods}
\name{marginalLikelihood}
\alias{marginalLikelihood}
\alias{marginalLikelihood,SingleBatchModel-method}
\alias{marginalLikelihood,SingleBatchModel,ANY-method}
\alias{marginalLikelihood,SingleBatchPooled-method}
\alias{marginalLikelihood,SingleBatchPooled,ANY-method}
\alias{marginalLikelihood,MultiBatchModel-method}
\alias{marginalLikelihood,MultiBatchModel,ANY-method}
\alias{marginalLikelihood,MultiBatchPooled-method}
\alias{marginalLikelihood,MultiBatchModel,ANY-method}
\alias{marginalLikelihood,list-method}
\alias{marginalLikelihood,list,ANY-method}
\title{Compute the marginal likelihood of a converged model.}
\usage{
marginalLikelihood(model, params = mlParams())

\S4method{marginalLikelihood}{SingleBatchModel}(model, params = mlParams())

\S4method{marginalLikelihood}{SingleBatchPooled}(model, params = mlParams())

\S4method{marginalLikelihood}{MultiBatchModel}(model, params = mlParams())

\S4method{marginalLikelihood}{MultiBatchPooled}(model, params = mlParams())

\S4method{marginalLikelihood}{list}(model, params = mlParams(warnings =
  FALSE))
}
\arguments{
\item{model}{An object of class \code{MarginalModel}, or a list of
\code{MarginalModel}'s. Can also be an object of \code{BatchModel} or
a list of such models.}

\item{params}{A list containing parameters for marginalLikelihood computation. See \code{mlParams} for details.}
}
\value{
A vector of the marginal likelihood of the model(s)
}
\description{
Compute the marginal likelihood of a converged model.
}
\examples{
## In practice, run a much longer burnin and increase the number of
## iterations to save after burnin
   mm <- SingleBatchModelExample
   mcmcParams(mm) <- McmcParams(iter=50, burnin=0, nStarts=0)
   mm <- posteriorSimulation(mm)
   marginalLikelihood(mm)
}
\seealso{
\code{\link{mlParams}}

Note: currently thinning of the reduced MCMC chains is not allowed.
}
