% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/marginal_likelihood.R
\docType{methods}
\name{marginalLikelihood}
\alias{marginalLikelihood}
\alias{marginalLikelihood,BatchModel,ANY-method}
\alias{marginalLikelihood,BatchModel-method}
\alias{marginalLikelihood,MarginalModel,ANY-method}
\alias{marginalLikelihood,MarginalModel-method}
\alias{marginalLikelihood,SingleBatchPooledVar,ANY-method}
\alias{marginalLikelihood,SingleBatchPooledVar-method}
\alias{marginalLikelihood,list,ANY-method}
\alias{marginalLikelihood,list-method}
\title{Compute the marginal likelihood of a converged model.}
\usage{
marginalLikelihood(model, params = list(niter = 1000L, root = (1/10),
  reject.threshold = 1e-50, prop.threshold = 0.5))

\S4method{marginalLikelihood}{MarginalModel}(model, params = list(niter =
  1000L, root = (1/10), reject.threshold = 1e-50, prop.threshold = 0.5))

\S4method{marginalLikelihood}{SingleBatchPooledVar}(model, params = list(niter
  = 1000L, root = (1/10), reject.threshold = 1e-50, prop.threshold = 0.5))

\S4method{marginalLikelihood}{BatchModel}(model, params = list(niter = 1000L,
  root = (1/10), reject.threshold = 1e-50, prop.threshold = 0.5))

\S4method{marginalLikelihood}{list}(model, params = list(niter = 1000L, root =
  (1/10), reject.threshold = 1e-50, prop.threshold = 0.5))
}
\arguments{
\item{model}{An object of class \code{MarginalModel}, or a list of
\code{MarginalModel}'s. Can also be an object of \code{BatchModel} or
a list of such models.}

\item{params}{A list containing: \cr
niter - the number of iterations for the reduced Gibb's sampler \cr
root - a tempering parameter. Before the log mean of the reduced 
Gibb's outputs are taken, the root of each iteration is taken \cr
reject.threshold - small values for reduced Gibb's output for 
\code{theta} can indicate overfitting. Values below reject.threshold
will be flagged \cr
prop.threshold - If a proportion prop.threshold or higher of the 
reduced Gibb's out for \code{theta} are smaller than 
reject.threshold, the marginalLikelihood will not be calculated 
and a warning will be displayed}
}
\value{
A vector of the marginal likelihood of the model(s)
}
\description{
Compute the marginal likelihood of a converged model.
}
\examples{
     marginalLikelihood(MarginalModelExample,
                        params=list(niter=5L,
                                    root=(1/10),
                                    reject.threshold=1e-50,
                                    prop.threshold=0.5))
}

