% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/marginal_likelihood.R
\docType{methods}
\name{marginalLikelihood}
\alias{marginalLikelihood}
\alias{marginalLikelihood,BatchModel,ANY-method}
\alias{marginalLikelihood,BatchModel,integer-method}
\alias{marginalLikelihood,BatchModel-method}
\alias{marginalLikelihood,MarginalModel,ANY-method}
\alias{marginalLikelihood,MarginalModel,integer-method}
\alias{marginalLikelihood,MarginalModel-method}
\alias{marginalLikelihood,SingleBatchPooledVar,integer-method}
\alias{marginalLikelihood,list,ANY-method}
\alias{marginalLikelihood,list,integer-method}
\alias{marginalLikelihood,list-method}
\title{Compute the marginal likelihood of a converged model.}
\usage{
marginalLikelihood(model, niter)

\S4method{marginalLikelihood}{MarginalModel,integer}(model, niter)

\S4method{marginalLikelihood}{SingleBatchPooledVar,integer}(model, niter)

\S4method{marginalLikelihood}{MarginalModel,ANY}(model)

\S4method{marginalLikelihood}{BatchModel,integer}(model, niter)

\S4method{marginalLikelihood}{BatchModel,ANY}(model)

\S4method{marginalLikelihood}{list,integer}(model, niter)

\S4method{marginalLikelihood}{list,ANY}(model, niter)
}
\arguments{
\item{model}{An object of class \code{MarginalModel}, or a list of
\code{MarginalModel}'s. Can also be an object of \code{BatchModel} or
a list of such models.}

\item{niter}{The number of iterations for the reduced Gibb's sampler.}
}
\value{
A vector of the marginal likelihood of the model(s)
}
\description{
Compute the marginal likelihood of a converged model.
}
\examples{
marginalLikelihood(MarginalModelExample)
}

