% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MarginalModelExample}
\alias{MarginalModelExample}
\title{This data is a simulated example of Marginal data}
\usage{
MarginalModelExample
}
\value{
An example of a `MarginalModel`
MarginalModelExample
}
\description{
This data is a simulated example of Marginal data
}
\author{
Jacob Carey
}

