% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MarginalModel.R
\name{MarginalModelList}
\alias{MarginalModelList}
\title{Constructor for list of single-batch models}
\usage{
MarginalModelList(data = numeric(), k = numeric(), ...)
}
\arguments{
\item{data}{numeric vector of average log R ratios}

\item{k}{numeric vector indicating the number of mixture components for each model}

\item{...}{additional arguments passed to \code{{MarginalModel}}}
}
\value{
a list. Each element of the list is a \code{BatchModel}
}
\description{
An object of class MarginalModel is constructed for each k, creating a list of
MarginalModels.
}
\examples{
mlist <- MarginalModelList(data=y(MarginalModelExample), k=1:4)
}
\seealso{
\code{\link{MarginalModel}}
}

