% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{mcmcParams}
\alias{mcmcParams}
\alias{mcmcParams,MixtureModel-method}
\alias{mcmcParams,MultiBatch-method}
\alias{mcmcParams,MultiBatchList-method}
\alias{mcmcParams,list-method}
\alias{mcmcParams<-}
\alias{mcmcParams<-,MixtureModel,ANY-method}
\alias{mcmcParams<-,MultiBatch,McmcParams-method}
\alias{mcmcParams<-,MultiBatchList,McmcParams-method}
\alias{mcmcParams<-,MultiBatchP,McmcParams-method}
\alias{mcmcParams<-,list,ANY-method}
\title{Retrieve MCMC parameters from model.}
\usage{
mcmcParams(object)

mcmcParams(object) <- value
}
\arguments{
\item{object}{see \code{showMethods(mcmcParams)}}

\item{value}{an object of class 'McmcParams' containing the new number of iterations, etc.}
}
\value{
An object of class 'McmcParams'
}
\description{
View number of iterations, burnin, etc.
}
\examples{
     mcmcParams(SingleBatchModelExample)
}
