% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R,
%   R/methods-TrioBatchModel.R
\docType{methods}
\name{mcmcParams}
\alias{mcmcParams}
\alias{mcmcParams<-}
\alias{mcmcParams,MixtureModel-method}
\alias{mcmcParams<-,MixtureModel,ANY-method}
\alias{mcmcParams<-,list,ANY-method}
\alias{mcmcParams,list-method}
\alias{mcmcParams<-,TrioBatchModel,ANY-method}
\title{Retrieve MCMC parameters from model.}
\usage{
mcmcParams(object)

mcmcParams(object) <- value

\S4method{mcmcParams}{MixtureModel}(object)

\S4method{mcmcParams}{MixtureModel,ANY}(object) <- value

\S4method{mcmcParams}{list,ANY}(object) <- value

\S4method{mcmcParams}{list}(object)

\S4method{mcmcParams}{TrioBatchModel,ANY}(object) <- value
}
\arguments{
\item{object}{see \code{showMethods(mcmcParams)}}

\item{value}{an object of class 'McmcParams' containing the new number of iterations, etc.}

\item{force}{logical value. If false (default) the update will not proceed.}
}
\value{
An object of class 'McmcParams'
}
\description{
View number of iterations, burnin, etc.

Replace number of iterations, burnin, etc. Any update of the MCMC parameters will trigger an update of the chains. However, if iter (the number of MCMC iterations) is set to a nonpositive value, the chains will not be updated and kept as is.
}
\examples{
     mcmcParams(SingleBatchModelExample)
}
