% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{modelName}
\alias{modelName}
\title{Abbreviated model name}
\usage{
modelName(model)
}
\arguments{
\item{model}{a SingleBatchModel, MultiBatchModel, etc.}
}
\description{
Abbreviated model name
}
\examples{
modelName(SingleBatchModelExample)
}
