% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-BatchModel.R,
%   R/methods-MarginalModel.R, R/methods-McmcChains.R,
%   R/methods-MultiBatchModel.R, R/methods-SingleBatchModel.R
\docType{methods}
\name{mu}
\alias{mu}
\alias{mu,BatchModel-method}
\alias{mu,MarginalModel-method}
\alias{mu,McmcChains-method}
\alias{mu,BatchModel-method}
\alias{mu,MarginalModel-method}
\title{Retrieve overall mean}
\usage{
mu(object)

\S4method{mu}{BatchModel}(object)

\S4method{mu}{MarginalModel}(object)

\S4method{mu}{McmcChains}(object)

\S4method{mu}{BatchModel}(object)

\S4method{mu}{MarginalModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(mu)}}
}
\value{
A vector containing 'mu'
}
\description{
Retrieve overall mean
}
\examples{
     mu(MarginalModelExample)
}
