% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-McmcChains.R
\name{mu,McmcChains-method}
\alias{mu,McmcChains-method}
\title{Accessor for mu slot of McmcChains class}
\usage{
\S4method{mu}{McmcChains}(object)
}
\description{
Accessor for mu slot of McmcChains class
}
