% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BatchModel.R,
%   R/methods-MultiBatchModel.R
\name{multiBatchDensities}
\alias{multiBatchDensities}
\alias{multiBatchDensities}
\title{Create a data.frame of the component densities for each batch}
\usage{
multiBatchDensities(model)

multiBatchDensities(model)
}
\arguments{
\item{object}{an object of class \code{BatchModel}}

\item{object}{an object of class \code{BatchModel}}
}
\value{
a \code{{data.frame}}

a \code{{data.frame}}
}
\description{
Create a data.frame of the component densities for each batch

Create a data.frame of the component densities for each batch
}
\examples{
   nbatch <- 3
   k <- 3
   means <- matrix(c(-2.1, -2, -1.95, -0.41, -0.4, -0.395, -0.1,
       0, 0.05), nbatch, k, byrow = FALSE)
   sds <- matrix(0.15, nbatch, k)
   sds[, 1] <- 0.3
   N <- 1000
   truth <- simulateBatchData(N = N, batch = rep(letters[1:3],
                                                 length.out = N),
                              p = c(1/10, 1/5, 1 - 0.1 - 0.2), theta = means,
                              sds = sds)
   mcmcp <- McmcParams(iter = 1000, burnin = 500, thin = 1,
                       nStarts = 10)

   ## this parameter setting for m2.0 allows a lot of varation of the thetas
   ## between batch
   hypp <- CNPBayes:::HyperparametersBatch(m2.0 = 1/60, eta.0 = 1800,
                                           k = 3, a = 1/6, b = 180)
   model <- BatchModel(data = y(truth), batch = batch(truth),
                       k = 3, mcmc.params = mcmcp, hypp = hypp)
   model <- posteriorSimulation(model)
   df <- multiBatchDensities(model)
   df.observed <- data.frame(y=observed(model), batch=batch(model))
   library(ggplot2)
   ggplot(df, aes(x, d)) +
   geom_histogram(data=df.observed,
                  aes(y, ..density..),
                  bins=300, inherit.aes=FALSE) +
   geom_area(stat="identity", aes(color=name, fill=name),
             alpha=0.4) +
   xlab("quantiles") + ylab("density") +
   scale_color_manual(values=colors) +
   scale_fill_manual(values=colors) +
   guides(fill=guide_legend(""), color=guide_legend("")) +
   facet_wrap(~batch, nrow=2)
   nbatch <- 3
   k <- 3
   means <- matrix(c(-2.1, -2, -1.95, -0.41, -0.4, -0.395, -0.1,
       0, 0.05), nbatch, k, byrow = FALSE)
   sds <- matrix(0.15, nbatch, k)
   sds[, 1] <- 0.3
   N <- 1000
   truth <- simulateBatchData(N = N, batch = rep(letters[1:3],
                                                 length.out = N),
                              p = c(1/10, 1/5, 1 - 0.1 - 0.2), theta = means,
                              sds = sds)
   mcmcp <- McmcParams(iter = 1000, burnin = 500, thin = 1,
                       nStarts = 10)

   ## this parameter setting for m2.0 allows a lot of varation of the thetas
   ## between batch
   hypp <- CNPBayes:::HyperparametersBatch(m2.0 = 1/60, eta.0 = 1800,
                                           k = 3, a = 1/6, b = 180)
   model <- BatchModel(data = y(truth), batch = batch(truth),
                       k = 3, mcmc.params = mcmcp, hypp = hypp)
   model <- posteriorSimulation(model)
   df <- multiBatchDensities(model)
   df.observed <- data.frame(y=observed(model), batch=batch(model))
   library(ggplot2)
   ggplot(df, aes(x, d)) +
   geom_histogram(data=df.observed,
                  aes(y, ..density..),
                  bins=300, inherit.aes=FALSE) +
   geom_area(stat="identity", aes(color=name, fill=name),
             alpha=0.4) +
   xlab("quantiles") + ylab("density") +
   scale_color_manual(values=colors) +
   scale_fill_manual(values=colors) +
   guides(fill=guide_legend(""), color=guide_legend("")) +
   facet_wrap(~batch, nrow=2)
}
