% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-McmcChains.R
\docType{methods}
\name{names,McmcChains-method}
\alias{names,McmcChains-method}
\title{Retrieve the names of the parameters estimated in the MCMC chain.}
\usage{
\S4method{names}{McmcChains}(x)
}
\arguments{
\item{x}{an object of class 'McmcChains'}
}
\description{
Retrieve the names of the parameters estimated in the MCMC chain.
}

