% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcParams.R
\name{nStarts}
\alias{nStarts}
\alias{nStarts,McmcParams-method}
\alias{nStarts<-}
\alias{nStarts<-,McmcParams,ANY-method}
\alias{nStarts<-,McmcParams-method}
\title{Number of random starts for MCMC}
\usage{
nStarts(object)

nStarts(object) <- value

\S4method{nStarts}{McmcParams,ANY}(object) <- value
}
\arguments{
\item{object}{a McmcParams instance}

\item{value}{number of chains}
}
\description{
Number of random starts for MCMC
}
