% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcParams.R, R/methods-MixtureModel.R, R/methods-ModelList.R
\docType{methods}
\name{nStarts}
\alias{nStarts}
\alias{nStarts,McmcParams-method}
\alias{nStarts,MixtureModel-method}
\alias{nStarts,ModelList-method}
\title{Number of MCMC chains.}
\usage{
nStarts(object)

\S4method{nStarts}{McmcParams}(object)

\S4method{nStarts}{MixtureModel}(object)

\S4method{nStarts}{ModelList}(object)
}
\arguments{
\item{object}{see showMethods(nStarts)}
}
\description{
This function retrieves the number of chains used for an MCMC simulation.
}

