% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcParams.R,
%   R/methods-MixtureModel.R
\docType{methods}
\name{nStarts}
\alias{nStarts}
\alias{nStarts<-}
\alias{nStarts,McmcParams-method}
\alias{nStarts<-,McmcParams,ANY-method}
\alias{nStarts<-,McmcParams-method}
\alias{nStarts,MixtureModel-method}
\alias{nStarts<-,MixtureModel,ANY-method}
\alias{nStarts<-,MixtureModel-method}
\title{Number of MCMC chains.}
\usage{
nStarts(object)

nStarts(object) <- value

\S4method{nStarts}{McmcParams}(object)

\S4method{nStarts}{McmcParams,ANY}(object) <- value

\S4method{nStarts}{MixtureModel}(object)

\S4method{nStarts}{MixtureModel,ANY}(object) <- value
}
\arguments{
\item{object}{see \code{showMethods(nStarts)}}

\item{value}{new number of chains}
}
\value{
An integer of the number of different starts.
}
\description{
This function retrieves the number of chains used for an MCMC simulation.
}
\details{
Simulating starting values from the priors makes it imperative to
  run a large nubmer of simulations for burnin and to carefully evaluate the
  chains following burning for convergence. The adequacy of the burnin is
  difficult to assess in high-dimensional settings with a large number of
  CNPs. To avoid starting in regions of low posterior probabilitiy, we use
  existing EM-based methods in the package \code{{mclust}} to select starting
  values from \code{N} bootstrap sample of the observed data, where \code{N}
  is specificed as in the example below. For each bootstrap sample, starting
  values for the model are estimated. For each set of simulated starting
  values, the log likelihood of the full data is evaluated. The starting
  values with the largest log likelihood are used as initial values for the
  MCMC simulations.
}
\examples{
number_of_chains <- nStarts(SingleBatchModelExample)
number_of_chains <- 10
nStarts(SingleBatchModelExample) <- number_of_chains
}
