% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcParams.R, R/methods-MixtureModel.R
\docType{methods}
\name{nStarts}
\alias{nStarts}
\alias{nStarts,McmcParams-method}
\alias{nStarts,MixtureModel-method}
\alias{nStarts<-}
\alias{nStarts<-,McmcParams-method}
\alias{nStarts<-,MixtureModel-method}
\title{Number of MCMC chains.}
\usage{
nStarts(object)

nStarts(object) <- value

\S4method{nStarts}{McmcParams}(object)

\S4method{nStarts}{McmcParams}(object) <- value

\S4method{nStarts}{MixtureModel}(object)

\S4method{nStarts}{MixtureModel}(object) <- value
}
\arguments{
\item{object}{see \code{showMethods(nStarts)}}

\item{value}{new number of chains}
}
\value{
An integer of the number of different starts.
}
\description{
This function retrieves the number of chains used for an MCMC simulation.

This function changes the number of chains used for an MCMC simulation.
}
\examples{
number_of_chains <- nStarts(MarginalModelExample)
number_of_chains <- 3
nStarts(MarginalModelExample) <- number_of_chains
}

