% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{nu.0}
\alias{nu.0}
\alias{nu.0,McmcChains-method}
\title{Accessor for nu.0 slot of McmcChains class}
\usage{
nu.0(object)
}
\arguments{
\item{object}{a Hyperparameters instance}
}
\description{
Accessor for nu.0 slot of McmcChains class
}
