% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R, R/methods-McmcChains.R
\docType{methods}
\name{nu.0}
\alias{nu.0}
\alias{nu.0,McmcChains-method}
\alias{nu.0,MixtureModel-method}
\title{Retrieve the shape parameter for the sigma.2 distribution.}
\usage{
nu.0(object)

\S4method{nu.0}{MixtureModel}(object)

\S4method{nu.0}{McmcChains}(object)
}
\arguments{
\item{object}{see \code{showMethods(nu.0)}}
}
\value{
An integer
}
\description{
Retrieve the shape parameter for the sigma.2 distribution.
}
\examples{
     nu.0(MarginalModelExample)
}

