% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{numberObs}
\alias{numberObs}
\title{Number of observations}
\usage{
numberObs(model)
}
\arguments{
\item{model}{a MixtureModel-derived object}
}
\description{
Number of observations
}
\examples{
numberObs(MarginalModelExample)
}

