% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{numberObs,MixtureModel-method}
\alias{numberObs,MixtureModel-method}
\title{Number of observations in mixture model object}
\usage{
\S4method{numberObs}{MixtureModel}(model)
}
\description{
Number of observations in mixture model object
}
