% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{numberObs}
\alias{numberObs}
\alias{numberObs,MixtureModel-method}
\title{Number of observations}
\usage{
numberObs(model)

\S4method{numberObs}{MixtureModel}(model)
}
\arguments{
\item{model}{a MixtureModel-derived object}
}
\description{
Number of observations
}
\examples{
numberObs(MarginalModelExample)
}

