% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{oned}
\alias{oned}
\title{Retrieve data.}
\usage{
oned(object)
}
\arguments{
\item{object}{see \code{showMethods(oned)}}
}
\value{
A vector the length of the data
}
\description{
Retrieve data.
}
