% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{oned}
\alias{oned}
\alias{oned,MixtureModel-method}
\title{Retrieve data.}
\usage{
oned(object)

\S4method{oned}{MixtureModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(oned)}}
}
\value{
A vector the length of the data
}
\description{
Retrieve data.
}

