% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-MarginalModel.R
\name{orderModels}
\alias{orderModels}
\title{Reorder models of varying component sizes.}
\usage{
orderModels(x)
}
\arguments{
\item{x}{the result of a call to \code{computeMarginalLik}.}
}
\description{
Models are ordered according to marginal likelihood. The marginal
likelihood is computed for each chain of each component size model
separately. The mean is taken by model, and ordering by this mean
marginal is performed.
}

