% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{orderModels}
\alias{orderModels}
\title{Order models by Bayes factor}
\usage{
orderModels(models, bf.thr = 10)
}
\arguments{
\item{models}{list of \code{MixtureModel}-derived objects}

\item{bf.thr}{scalar: minimal bayes factor for selecting a model with more parameters over a more parsimonious model}
}
\description{
Order models by Bayes factor
}
