% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{p}
\alias{p}
\title{Retrieve mixture proportions.}
\usage{
p(object)
}
\arguments{
\item{object}{an object of class MarginalModel or BatchModel}
}
\value{
A vector of length the number of components
}
\description{
Retrieve mixture proportions.
}
\examples{
     p(MarginalModelExample)
}

