% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-DensityModel.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,BatchModel,ANY-method}
\alias{plot,DensityBatchModel,ANY-method}
\alias{plot,DensityModel,ANY-method}
\alias{plot,DensityModel,numeric-method}
\alias{plot,MarginalModel,ANY-method}
\title{Plot the densities estimated from a mixture model for a copy number polymorphism}
\usage{
plot(x, y, ...)

\S4method{plot}{DensityModel,ANY}(x, y, ...)

\S4method{plot}{MarginalModel,ANY}(x, y, ...)

\S4method{plot}{BatchModel,ANY}(x, y, show.batch = TRUE, ...)

\S4method{plot}{DensityBatchModel,ANY}(x, show.batch = TRUE, ...)
}
\arguments{
\item{x}{a \code{DensityModel}-derived object, or a
\code{MixtureModel}-derived object.}

\item{y}{If \code{x} is a \code{DensityModel}, \code{y} is a
numeric vector of the one-dimensional summaries for a given copy
number polymorphism. If \code{x} is a \code{MixtureModel}, \code{y}
is ignored.}

\item{...}{Additional arguments passed to \code{hist}.}

\item{show.batch}{a logical. If true, batch specific densities
will be plotted.}
}
\value{
A plot showing the density estimate
}
\description{
Plot estimates of the posterior density for each component and the
overall, marginal density.  For batch models, one can additionally
plot batch-specific density estimates.
}
\examples{
  set.seed(100)
  truth <- simulateData(N=2500,
                        theta=c(-2, -0.4, 0),
                        sds=c(0.3, 0.15, 0.15),
                        p=c(0.05, 0.1, 0.8))

  mcmcp <- McmcParams(iter=500, burnin=500, thin=2)
  model <- MarginalModel(y(truth), k=3, mcmc.params=mcmcp)
  model <- CNPBayes:::startAtTrueValues(model, truth)
  model <- posteriorSimulation(model)
  par(mfrow=c(1,2), las=1)
  plot(truth)
  plot(model)
}

