% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/posteriorSimulation-methods.R
\docType{methods}
\name{posteriorSimulation}
\alias{posteriorSimulation}
\alias{posteriorSimulation,MixtureModel-method}
\alias{posteriorSimulation,TrioBatchModel-method}
\title{Run MCMC simulation.}
\usage{
posteriorSimulation(object, k)

\S4method{posteriorSimulation}{MixtureModel}(object)

\S4method{posteriorSimulation}{TrioBatchModel}(object)
}
\arguments{
\item{object}{see showMethods(posteriorSimulation)}

\item{k}{The number of a priori components. This is optional and if not
specified, the stored k model components are used. This parameters is
useful for running multiple models of varying components.}
}
\value{
An object of class 'MarginalModel' or 'BatchModel'
}
\description{
nStarts chains are run. b burnin iterations are run and then discarded.
Next, s iterations are run in each crain. The user can also specify
an alternative number of components.
The mode of the MCMC simulation is also calculated.
}
\examples{
# Fit model with pre-specified number of components (k=3)
set.seed(123)
## specify small number of iterations so that the example runs quickly
mp <- McmcParams(iter=2, burnin=0, nStarts=3)
sb <- SingleBatchModelExample
mcmcParams(sb) <- mp
posteriorSimulation(sb)

# Run additional iterations, but set nStart = 0 so that the last value of the chain is the first value of the next chain
mcmcParams(sb) <- McmcParams(iter=5, nStarts=0, burnin=0)
posteriorSimulation(sb)
}
\seealso{
See \code{\link{ggMixture}} for plotting the model-based densities.
}
