% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/copynumber-models.R
\docType{methods}
\name{probCopyNumber}
\alias{probCopyNumber}
\alias{probCopyNumber,SingleBatchCopyNumber-method}
\alias{probCopyNumber,MultiBatchCopyNumber-method}
\alias{probCopyNumber,MultiBatchCopyNumberPooled-method}
\title{Posterior probabilities for copy number states}
\usage{
probCopyNumber(model)

\S4method{probCopyNumber}{SingleBatchCopyNumber}(model)

\S4method{probCopyNumber}{MultiBatchCopyNumber}(model)

\S4method{probCopyNumber}{MultiBatchCopyNumberPooled}(model)
}
\description{
In contrast to posterior probabilities for mixture components, this function
returns posterior probabilities for distinct copy number states.
a \code{SingleBatchCopyNumber} or \code{MultiBatchCopyNumber} instance
}
\seealso{
\code{\link{CopyNumber-methods}}
}
