% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{sigma}
\alias{sigma}
\title{Retrieve standard deviations of each component/batch mean.}
\usage{
sigma(object)
}
\arguments{
\item{object}{an object of class MarginalModel or BatchModel}
}
\value{
A vector of length K, or a matrix of size B x K, where
K is the number of components and B is the number of batches
}
\description{
Retrieve standard deviations of each component/batch mean.
}
\examples{
     sigma(MarginalModelExample)
}

