% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-McmcChains.R
\name{sigma2.0,McmcChains-method}
\alias{sigma2.0,McmcChains-method}
\title{Accessor for sigma2.0 slot of McmcChains class}
\usage{
\S4method{sigma2.0}{McmcChains}(object)
}
\description{
Accessor for sigma2.0 slot of McmcChains class
}
