% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulateBatchData}
\alias{simulateBatchData}
\title{Create simulated batch data for testing.}
\usage{
simulateBatchData(N = 2500, p, theta, sds, batch, zz)
}
\arguments{
\item{N}{number of observations}

\item{p}{a vector indicating probability of membership to each component}

\item{theta}{a vector of means, one per component/batch}

\item{sds}{a vector of standard deviations, one per component/batch}

\item{batch}{a vector of labels indication from which batch each simulation should come from}

\item{zz}{a vector indicating latent variable membership. Can be omitted.}
}
\value{
An object of class 'BatchModel'
}
\description{
Create simulated batch data for testing.
}
\examples{
k <- 3
nbatch <- 3
means <- matrix(c(-1.2, -1.0, -0.8,
                  -0.2, 0, 0.2,
                  0.8, 1, 1.2), nbatch, k, byrow=FALSE)
sds <- matrix(0.1, nbatch, k)
N <- 1500
truth <- simulateBatchData(N=N,
                           batch=rep(letters[1:3], length.out=N),
                           theta=means,
                           sds=sds,
                           p=c(1/5, 1/3, 1-1/3-1/5))
}

