% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulateData}
\alias{simulateData}
\title{Create simulated data for testing.}
\usage{
simulateData(N, p, theta, sds)
}
\arguments{
\item{N}{number of observations}

\item{p}{a vector indicating probability of membership to each component}

\item{theta}{a vector of means, one per component}

\item{sds}{a vector of standard deviations, one per component}
}
\value{
An object of class 'MarginalModel'
}
\description{
Create simulated data for testing.
}
\examples{
truth <- simulateData(N=2500, p=rep(1/3, 3),
                      theta=c(-1, 0, 1),
                      sds=rep(0.1, 3))
}

