% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-SingleBatchModel.R
\name{SingleBatchModel2}
\alias{SingleBatchModel2}
\title{Constructor for list of single-batch models}
\usage{
SingleBatchModel2(dat = numeric(), hp = Hyperparameters(),
  mp = McmcParams(iter = 1000, burnin = 1000, thin = 10, nStarts = 4))
}
\arguments{
\item{dat}{numeric vector of average log R ratios}

\item{hp}{an object of class \code{Hyperparameters}}

\item{mp}{an object of class \code{McmcParams}}

\item{k}{numeric vector indicating the number of mixture components for each model}
}
\value{
a list. Each element of the list is a \code{BatchModel}
}
\description{
An object of class MarginalModel is constructed for each k, creating a list of
MarginalModels.
}
\examples{
mlist <- MarginalModelList(data=y(MarginalModelExample), k=1:4)
mcmcParams(mlist) <- McmcParams(iter=1, burnin=1, nStarts=0)
mlist2 <- posteriorSimulation(mlist)
}
\seealso{
\code{\link{MultiBatchModel}} \code{\link{BatchModelList}}
}
