% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-BatchModel.R, R/methods-MarginalModel.R, R/methods-McmcChains.R
\docType{methods}
\name{tau2}
\alias{tau2}
\alias{tau2,BatchModel-method}
\alias{tau2,MarginalModel-method}
\alias{tau2,McmcChains-method}
\title{Retrieve overall variance across components/batches.}
\usage{
tau2(object)

\S4method{tau2}{BatchModel}(object)

\S4method{tau2}{MarginalModel}(object)

\S4method{tau2}{McmcChains}(object)
}
\arguments{
\item{object}{see \code{showMethods(tau2)}}
}
\description{
Retrieve overall variance across components/batches.
}

