% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-BatchModel.R, R/methods-MarginalModel.R, R/methods-McmcChains.R
\docType{methods}
\name{tau2}
\alias{tau2}
\alias{tau2,BatchModel-method}
\alias{tau2,MarginalModel-method}
\alias{tau2,McmcChains-method}
\title{Accessor for the tau2 parameter in the hierarchical mixture model}
\usage{
tau2(object)

\S4method{tau2}{BatchModel}(object)

\S4method{tau2}{MarginalModel}(object)

\S4method{tau2}{McmcChains}(object)
}
\arguments{
\item{object}{see \code{showMethods(tau2)}}
}
\value{
A vector of variances
}
\description{
The interpretation of \code{tau2} depends on whether \code{object}
is a \code{MarginalModel} or a \code{BatchModel}.  For
\code{BatchModel}, \code{tau2} is a vector with length equal to the
number of components.  Each element of the \code{tau2} vector can
be interpreted as the within-component variance of the batch means
(\code{theta}).  For objects of class \code{MarginalModel}
(assumes no batch effect), \code{tau2} is a length-one vector that
describes the variance of the component means between batches.  The
hyperparameters of \code{tau2} are \code{eta.0} and \code{m2.0}. See the
following examples for setting the hyperparameters, accessing the current
value of \code{tau2} from a \code{MixtureModel}-derived object, and
for plotting the chain of \code{tau2} values.
}
\examples{
k(BatchModelExample)
tau2(BatchModelExample)
plot.ts(tau2(chains(BatchModelExample)))
}
\seealso{
\code{Hyperparameters}
}

