% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MixtureModel.R
\name{tauMean}
\alias{tauMean}
\title{Retrieve overall standard deviation averaged across MCMC simulations.}
\usage{
tauMean(object)
}
\arguments{
\item{object}{an object of class MarginalModel or BatchModel}
}
\value{
A vector of size 1 or number of batches
}
\description{
Retrieve overall standard deviation averaged across MCMC simulations.
}
\examples{
     tauMean(MarginalModelExample)
}

