% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-BatchModel.R, R/methods-MarginalModel.R, R/methods-McmcChains.R
\docType{methods}
\name{theta}
\alias{theta}
\alias{theta,BatchModel-method}
\alias{theta,MarginalModel-method}
\alias{theta,McmcChains-method}
\title{Retrieve theta.}
\usage{
theta(object)

\S4method{theta}{BatchModel}(object)

\S4method{theta}{MarginalModel}(object)

\S4method{theta}{McmcChains}(object)
}
\arguments{
\item{object}{see \code{showMethods(theta)}}
}
\description{
This function retrieves theta from an object. Theta is a vector or matrix representing the means of each component/batch.
}

