% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R,
%   R/methods-MultiBatchModel.R
\docType{methods}
\name{theta}
\alias{theta}
\alias{theta,McmcChains-method}
\alias{theta,MixtureModel-method}
\title{Accessor for the theta parameter in the hierarchical mixture model}
\usage{
theta(object)

\S4method{theta}{McmcChains}(object)

\S4method{theta}{MixtureModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(theta)}}
}
\value{
A vector of length number of components or a matrix of size 
number of batches x number of components
}
\description{
The interpretation of \code{theta} depends on whether \code{object}
is a \code{MarginalModel} or a \code{BatchModel}.  For
\code{BatchModel}, \code{theta} is a matrix of size B x K, where B is
the number of batches and K is the number of components.
Each column of the \code{theta} matrix can be interpreted as the
batch means for a particular component. For objects of class
\code{MarginalModel} (assumes no batch effect), \code{theta} is a
vector of length K. Each element of \code{theta} can be interpreted
as the mean for a component. See the following examples for accessing
the current value of \code{theta} from a \code{MixtureModel}-derived
object, and for plotting the chain of \code{theta} values.
}
\examples{
## MarginalModel
\dontrun{
k(SingleBatchModelExample)
theta(SingleBatchModelExample)
plot.ts(theta(chains(SingleBatchModelExample)))
## BatchModel
k(MultiBatchModelExample)
length(unique(batch(MultiBatchModelExample)))
theta(MultiBatchModelExample)
## Plot means for batches in one component
plot.ts(theta(chains(MultiBatchModelExample))[, 1:3])
}
}
