% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcParams.R
\docType{methods}
\name{thin}
\alias{thin}
\alias{thin<-}
\alias{thin,McmcParams-method}
\title{Get or set the number of thinning intervals.}
\usage{
thin(object)

thin(object) <- value

\S4method{thin}{McmcParams}(object)
}
\arguments{
\item{object}{see showMethods(thin)}
}
\value{
An integer of the number of thinning intervals
}
\description{
This function gets or sets the number of thinning intervals used for an MCMC
simulation.
}
\examples{
     thin(SingleBatchModelExample)
     thin(SingleBatchModelExample) <- 10L
}
