% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R,
%   R/methods-McmcParams.R
\docType{methods}
\name{thin}
\alias{thin}
\alias{thin<-}
\alias{thin,MixtureModel-method}
\alias{thin<-,MixtureModel,numeric-method}
\alias{thin,McmcParams-method}
\title{Get or set the number of thinning intervals.}
\usage{
thin(object)

thin(object) <- value

\S4method{thin}{MixtureModel}(object)

\S4method{thin}{MixtureModel,numeric}(object) <- value

\S4method{thin}{McmcParams}(object)
}
\arguments{
\item{object}{see showMethods(thin)}

\item{value}{a length-one numeric vector indicating how often to save MCMC iterations to the chain.  For example, a thin of 10 means that every 10th MCMC simulation is saved to the chain.}
}
\value{
An integer of the number of thinning intervals
}
\description{
This function gets or sets the number of thinning intervals used for an MCMC
simulation.
}
\examples{
     thin(MarginalModelExample)
     thin(MarginalModelExample) <- 10L
}
