% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcParams.R, R/methods-MixtureModel.R
\docType{methods}
\name{thin}
\alias{thin}
\alias{thin,McmcParams-method}
\alias{thin,MixtureModel-method}
\title{Number of thinning intervals.}
\usage{
thin(object)

\S4method{thin}{McmcParams}(object)

\S4method{thin}{MixtureModel}(object)
}
\arguments{
\item{object}{see showMethods(thin)}
}
\value{
An integer of the number of thinning intervals
}
\description{
This function retrieves the number of thinning intervals used for an MCMC simulation.
}
\examples{
thin(MarginalModelExample)
}

