% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-DensityModel.R, R/methods-MixtureModel.R, R/methods-ModelList.R, R/methods-ModelParams.R
\docType{methods}
\name{y}
\alias{y}
\alias{y,DensityModel-method}
\alias{y,MixtureModel-method}
\alias{y,ModelList-method}
\alias{y,ModelParams-method}
\title{Retrieve data.}
\usage{
y(object)

\S4method{y}{DensityModel}(object)

\S4method{y}{MixtureModel}(object)

\S4method{y}{ModelList}(object)

\S4method{y}{ModelParams}(object)
}
\arguments{
\item{object}{see \code{showMethods(y)}}
}
\description{
Retrieve data.
}

