% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-DensityModel.R, R/methods-MixtureModel.R
\docType{methods}
\name{y}
\alias{y}
\alias{y,DensityModel-method}
\alias{y,MixtureModel-method}
\title{Retrieve data.}
\usage{
y(object)

\S4method{y}{DensityModel}(object)

\S4method{y}{MixtureModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(y)}}
}
\value{
A vector containing the data
}
\description{
Retrieve data.
}
\examples{
y(MarginalModelExample)
}

