% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R, R/methods-MixtureModel.R
\docType{methods}
\name{z}
\alias{z}
\alias{z,McmcChains-method}
\alias{z,MixtureModel-method}
\title{Retrieve latent variable assignments.}
\usage{
z(object)

\S4method{z}{McmcChains}(object)

\S4method{z}{MixtureModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(z)}}
}
\value{
A vector the length of the data
}
\description{
Retrieves the simulated latent variable assignments of each observation at each MCMC simulation.
}
\examples{
     z(MarginalModelExample)
}

