% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-McmcChains.R, R/methods-MixtureModel.R
\docType{methods}
\name{zFreq}
\alias{zFreq}
\alias{zFreq,McmcChains-method}
\alias{zFreq,MixtureModel-method}
\alias{zfreq,McmcChains-method}
\alias{zfreq,MixtureModel-method}
\title{Calculates a frequency table of latent variable assigments by observation.}
\usage{
zFreq(object)

\S4method{zFreq}{McmcChains}(object)

\S4method{zFreq}{MixtureModel}(object)
}
\arguments{
\item{object}{see \code{showMethods(zfreq)}}
}
\value{
An integer vector of length the number of components
}
\description{
Calculates a frequency table of latent variable assigments by observation.
}
\examples{
zFreq(MarginalModelExample)
}

